% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_open_payments.R
\name{add_downloadurl}
\alias{add_downloadurl}
\title{Join Vector of Download URLs to Main Dataset}
\usage{
add_downloadurl(x)
}
\arguments{
\item{x}{data.frame}
}
\value{
data.frame
}
\description{
Join Vector of Download URLs to Main Dataset
}
\keyword{internal}
