% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_affiliations.R
\name{affiliations}
\alias{affiliations}
\title{Provider Facility Affiliations}
\usage{
affiliations(
  npi = NULL,
  pac = NULL,
  last_name = NULL,
  first_name = NULL,
  middle_name = NULL,
  suffix = NULL,
  facility_type = NULL,
  ccn_facility = NULL,
  ccn_primary = NULL
)
}
\arguments{
\item{npi}{\verb{<chr>} 10-digit Individual NPI}

\item{pac}{\verb{<chr>} 10-digit PECOS Associate Control (PAC) ID}

\item{first_name, middle_name, last_name, suffix}{\verb{<chr>} Individual provider's name}

\item{facility_type}{\verb{<chr>} Type of facility, one of the following:
\itemize{
\item \code{Hospital} (\code{hp})
\item \verb{Long-term care hospital} (\code{ltch})
\item \verb{Nursing home} (\code{nh})
\item \verb{Inpatient rehabilitation facility} (\code{irf})
\item \verb{Home health agency} (\code{hha})
\item \verb{Skilled nursing facility} (\code{snf})
\item \code{Hospice} (\code{hs})
\item \verb{Dialysis facility} (\code{df})
}}

\item{ccn_facility}{\verb{<chr>} 6-digit CMS Certification Number (CCN) of
facility or unit within hospital where an
individual provider provides service.}

\item{ccn_primary}{\verb{<int>} 6-digit CMS Certification Number (CCN) of a
sub-unit's primary hospital, should the
provider provide services in said unit.}
}
\description{
Facility affiliations data publicly reported in the Provider Data Catalog.
}
\section{Periodicity}{

\itemize{
\item Monthly (R/P1M)
}
}

\examples{
affiliations(last_name = "CURRY", facility_type = "Home health agency")
affiliations(ccn_primary = "670055")
affiliations(ccn_facility = "370781")
affiliations(ccn_facility = "331302")
affiliations(ccn_facility = "33Z302")
affiliations(npi = "1043245657")
affiliations(last_name = "CURRY")
}
