% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_args.R
\name{arg_npi}
\alias{arg_npi}
\alias{arg_state}
\title{\code{npi} argument class}
\usage{
arg_npi(x = NULL)

arg_state(x = NULL)
}
\arguments{
\item{x}{\verb{<chr>} vector of state abbreviations}
}
\value{
An S7 \verb{<arg_npi>} object.

An S7 \verb{<arg_state>} object.
}
\description{
\code{npi} argument class

\code{state} argument class
}
\examples{
arg_npi("1225701881")

arg_npi(1225701881)

try(arg_npi(122570188))

kind <- c(
1225701881,
1174270805,
1235702796,
1962116806,
1013647569,
1306500665,
1982296737,
1083295638,
1841967825,
1891390084,
1275117269,
1992338701,
1891355863,
1548743511,
1023473279,
1861857013,
1689182859,
1982059275)

arg_npi(kind)

arg_state("CA")

try(arg_state("C"))

try(arg_state("AA"))

# Allow duplicates?
arg_state(c("CA", "CA"))

}
