% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrollees.R
\name{as_Dataset}
\alias{as_Dataset}
\title{Public Provider Enrollment Dataset Object}
\usage{
as_Dataset(dataset)
}
\arguments{
\item{dataset}{\verb{<chr>} dataset title}
}
\value{
\verb{<S7_class>} object
}
\description{
Public Provider Enrollment Dataset Object
}
\examples{
as_Dataset(dataset = "Public Provider Enrollment")

}
