% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_misc.R
\name{as_datetime}
\alias{as_datetime}
\title{Parse datetime character vectors}
\usage{
as_datetime(x)
}
\arguments{
\item{x}{\verb{<chr>} vector to parse; format: "YYYY-MM-DDTHH:MM:SS"}
}
\value{
\verb{<chr>} parsed ISOdatetime vector
}
\description{
Parse datetime character vectors
}
\examples{
\dontshow{if (rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
as_datetime("2024-07-29T20:37:53")
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[clock:date-time-parse]{clock::date_time_parse_RFC_3339()}}
}
\keyword{internal}
