% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrollees.R
\name{as_Distribution}
\alias{as_Distribution}
\title{Public Provider Enrollment Distribution Object}
\usage{
as_Distribution(dataset)
}
\arguments{
\item{dataset}{\verb{<chr>} dataset title}
}
\value{
\verb{<S7_class>} object
}
\description{
Public Provider Enrollment Distribution Object
}
\examples{
as_Distribution(dataset = "Public Provider Enrollment")

}
