% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_utils.R
\name{as_fda_date}
\alias{as_fda_date}
\title{Parse \code{openFDA} date character vectors}
\usage{
as_fda_date(i)
}
\arguments{
\item{i}{\verb{<chr>} vector to parse; format: "YYYY-MM-DD"}
}
\value{
\verb{<chr>} parsed ISOdate vector
}
\description{
Parse \code{openFDA} date character vectors
}
\keyword{internal}
