% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_method_request.R
\name{base_request}
\alias{base_request}
\title{Create a new request by class}
\usage{
base_request(x)
}
\arguments{
\item{x}{An object of class \code{care_endpoint}, \code{care_group}, \code{pro_endpoint}, or \code{open_endpoint}}
}
\value{
A new base request
}
\description{
Create a new request by class
}
\examples{
care_endpoint("enrollees") |> base_request()
pro_endpoint("PDC_affiliations") |> base_request()
open_endpoint("PROF_covered") |> base_request()
caid_endpoint("MLR_summary") |> base_request()

care_group("hospital") |> base_request()
pro_group("MIPS") |> base_request()

care_temporal("quality_payment") |> base_request()
caid_temporal("MCP") |> base_request()
open_temporal("GROUP_all") |> base_request()

# care_troup("utilization") |> base_request()
# open_troup("grouped_payment") |> base_request()
}
