% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_getters.R
\name{browse_link}
\alias{browse_link}
\title{Browse Links}
\usage{
browse_link(x, active = is_interactive(), call = caller_env())
}
\arguments{
\item{x}{\verb{<backend>} or \verb{<group>} object}

\item{active}{\verb{<logical>} whether to open links in browser (default: \code{TRUE} if interactive)}

\item{call}{\verb{<env>} environment to use for error reporting}
}
\value{
\verb{<list>} of catalogs
}
\description{
Browse Links
}
\examples{
care_endpoint("care_enrollees") |> browse_link(active = FALSE)
try(caid_endpoint("managed_longterm") |> browse_link(active = FALSE))
}
