% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_caid.R
\name{caidMain}
\alias{caidMain}
\title{Provider Endpoint}
\usage{
caidMain(alias)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint alias}
}
\value{
An S7 \verb{<caidMain>} object.
}
\description{
Provider Endpoint
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
caidMain("affiliations")
caidMain("clinicians")
caidMain("utilization")
\dontshow{\}) # examplesIf}
}
