% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_backend.R
\name{careTempGroup}
\alias{careTempGroup}
\title{Medicare Temporal Group}
\usage{
careTempGroup(alias)
}
\arguments{
\item{alias}{\verb{<chr>} title alias}
}
\value{
An S7 \verb{<careTempGroup>} object.
}
\description{
Medicare Temporal Group
}
\examples{
careTempGroup("utilization")
careTempGroup("prescribers")
careTempGroup("suppliers")
careTempGroup("outpatient")
careTempGroup("inpatient")
}
