% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_care.R
\name{care_dimensions}
\alias{care_dimensions}
\title{Medicare Class Dimensions}
\usage{
care_dimensions(x)
}
\arguments{
\item{x}{\verb{<list>} dimensions list}
}
\value{
An S7 \verb{<care_dimensions>} object.
}
\description{
Medicare Class Dimensions
}
\keyword{internal}
