% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_care.R
\name{care_endpoint}
\alias{care_endpoint}
\title{Medicare Endpoint}
\usage{
care_endpoint(alias)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint alias}
}
\value{
An S7 \verb{<care_endpoint>} object.
}
\description{
Medicare Endpoint
}
\examples{
care_endpoint("contact")
care_endpoint("crosswalk")
care_endpoint("CARE_dialysis")
care_endpoint("enrollees")
care_endpoint("facilities")
care_endpoint("IQIES")
care_endpoint("laboratories")
care_endpoint("long_term")
care_endpoint("opt_out")
care_endpoint("order_refer")
care_endpoint("RBCS")
care_endpoint("transparency")
}
