% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_care.R
\name{care_group}
\alias{care_group}
\title{Medicare Endpoint Group}
\usage{
care_group(alias)
}
\arguments{
\item{alias}{\verb{<chr>} title alias}
}
\value{
An S7 \verb{<care_group>} object.
}
\description{
Medicare Endpoint Group
}
\examples{
care_group("HHA")
care_group("hospice")
care_group("hospital")
care_group("RHC")
care_group("FQHC")
care_group("pending")
care_group("reassignment")
care_group("SNF")
}
