% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_care.R
\name{care_metadata}
\alias{care_metadata}
\title{Medicare Class Metadata}
\usage{
care_metadata(x)
}
\arguments{
\item{x}{\verb{<list>} metadata list}
}
\value{
An S7 \verb{<care_metadata>} object.
}
\description{
Medicare Class Metadata
}
\keyword{internal}
