% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_care.R
\name{care_temporal}
\alias{care_temporal}
\title{Medicare Temporal Endpoint}
\usage{
care_temporal(alias)
}
\arguments{
\item{alias}{\verb{<chr>} title alias}
}
\value{
An S7 \verb{<care_temporal>} object.
}
\description{
Medicare Temporal Endpoint
}
\examples{
care_temporal("quality_payment")
}
