% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_care.R
\name{care_troup}
\alias{care_troup}
\title{Group of Medicare Temporal Endpoints}
\usage{
care_troup(alias)
}
\arguments{
\item{alias}{\verb{<chr>} title alias}
}
\value{
An S7 \verb{<care_troup>} object.
}
\description{
Group of Medicare Temporal Endpoints
}
\examples{
care_troup("inpatient")
care_troup("outpatient")
care_troup("utilization")
care_troup("suppliers")
care_troup("prescribers")
care_troup("staffing")
}
