% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_care.R
\name{careDim}
\alias{careDim}
\title{Medicare Class Dimensions}
\usage{
careDim(x)
}
\arguments{
\item{x}{\verb{<list>} dimensions list}
}
\value{
An S7 \verb{<careDim>} object.
}
\description{
Medicare Class Dimensions
}
\keyword{internal}
