% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_care.R
\name{careGroup}
\alias{careGroup}
\title{Medicare Endpoint Group}
\usage{
careGroup(alias)
}
\arguments{
\item{alias}{\verb{<chr>} title alias}
}
\value{
An S7 \verb{<careGroup>} object.
}
\description{
Medicare Endpoint Group
}
\examples{
careGroup("home_health")
careGroup("hospice")
careGroup("hospitals")
careGroup("rural_health")
careGroup("fqhc")
careGroup("pending")
careGroup("reassignment")
careGroup("skilled_nursing")
}
