% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_care.R
\name{careMeta}
\alias{careMeta}
\title{Medicare Class Metadata}
\usage{
careMeta(x)
}
\arguments{
\item{x}{\verb{<list>} metadata list}
}
\value{
An S7 \verb{<careMeta>} object.
}
\description{
Medicare Class Metadata
}
\keyword{internal}
