% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_care.R
\name{careTempGroup}
\alias{careTempGroup}
\title{Group of Medicare Temporal Endpoints}
\usage{
careTempGroup(alias)
}
\arguments{
\item{alias}{\verb{<chr>} title alias}
}
\value{
An S7 \verb{<careTempGroup>} object.
}
\description{
Group of Medicare Temporal Endpoints
}
\examples{
careTempGroup("inpatient")
careTempGroup("outpatient")
careTempGroup("prescribers")
careTempGroup("suppliers")
careTempGroup("utilization")
}
