% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_open.R
\name{catalog_open}
\alias{catalog_open}
\title{CMS Open Payments Catalog}
\usage{
catalog_open()
}
\value{
\verb{<list>} of Open Payments API catalog information
}
\description{
CMS Open Payments Catalog
}
