% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_old_open.R
\name{catalog_open_}
\alias{catalog_open_}
\title{Load Open Payments Catalog}
\usage{
catalog_open_()
}
\value{
\verb{<list>} of Open Payments API catalog information
}
\description{
Load Open Payments Catalog
}
\keyword{internal}
