% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_open_payments.R
\name{catalog_open_payments}
\alias{catalog_open_payments}
\title{Load Open Payments Catalog}
\usage{
catalog_open_payments()
}
\value{
\verb{<list>} of Open Payments API catalog information
}
\description{
Load Open Payments Catalog
}
\examples{
catalog_open_payments()
}
