% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{Catalog_openpayments}
\alias{Catalog_openpayments}
\title{CMS Open Payments Catalog}
\usage{
Catalog_openpayments()
}
\value{
\verb{<list>} of \verb{<tibbles>}: \code{dataset}, \code{distribution}, \code{downloads}
}
\description{
CMS Open Payments Catalog
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
Catalog_openpayments()
\dontshow{\}) # examplesIf}
}
