% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{Catalog_public}
\alias{Catalog_public}
\title{CMS Public Catalog}
\usage{
Catalog_public()
}
\value{
\verb{<list>} of \verb{<tibbles>}: \code{dataset}, \code{distribution}, \code{downloads}
}
\description{
CMS Public Catalog
}
\examples{
Catalog_public()

}
