% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-catalogs.R
\name{catalogs}
\alias{catalogs}
\title{API Catalogs}
\usage{
catalogs()
}
\value{
\verb{<list>} of catalogs
}
\description{
List of API catalogs:
\enumerate{
\item \code{care} - CMS Medicare API
\item \code{pro} - CMS Provider API
\item \code{open} - CMS Open Payments API
\item \code{caid} - CMS Medicaid API
\item \code{health} - CMS HealthCare.gov API
}
}
\examples{
catalogs()
}
