% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_luhn}
\alias{check_luhn}
\title{Vectorized Luhn Algorithm Check for NPIs}
\usage{
check_luhn(x)
}
\arguments{
\item{x}{\verb{<chr>} NPI}
}
\value{
\verb{<lgl>} \code{TRUE} if valid NPI, \code{FALSE} otherwise
}
\description{
Vectorized Luhn Algorithm Check for NPIs
}
\examples{
check_luhn("1234567890")
check_luhn(c("1417918293", "1234567890"))
}
\keyword{internal}
