% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{class_Identifier}
\alias{class_Identifier}
\title{Identifier Class}
\usage{
class_Identifier(url = NULL)
}
\arguments{
\item{url}{\verb{<chr>} Identifier url}
}
\value{
\verb{<S7_class>} object
}
\description{
\code{class_Identifier} object
}
\examples{
class_Identifier(url =
   paste0(
   "https://data.cms.gov/data-api/v1/dataset/",
   "2457ea29-fc82-48b0-86ec-3b0755de7515/",
   "data-viewer"))

}
