% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{class_publisher}
\alias{class_publisher}
\title{publisher Class}
\usage{
class_publisher(
  type = "org:Organization",
  name = "Centers for Medicare & Medicaid Services"
)
}
\arguments{
\item{type}{\verb{<chr>} type}

\item{name}{\verb{<chr>} publisher name}
}
\value{
\verb{<S7_class>} object
}
\description{
\code{class_publisher} object
}
