% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_misc.R
\name{clean_names}
\alias{clean_names}
\title{Clean column names}
\usage{
clean_names(x)
}
\arguments{
\item{x}{\verb{<chr>} column name}
}
\value{
\verb{<chr>} cleaned column name
}
\description{
Clean column names
}
\keyword{internal}
