% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_request.R
\name{cli_n_results_requests}
\alias{cli_n_results_requests}
\title{Inform number of results and requests}
\usage{
cli_n_results_requests(n, limit)
}
\arguments{
\item{n}{\verb{<int>} Number of results returned in an API request}

\item{limit}{\verb{<int>} API rate limit, i.e. the maximum number of results an
API will return per request.}
}
\value{
cli message
}
\description{
Inform number of results and requests
}
\keyword{internal}
