% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{collection}
\alias{collection}
\title{Endpoint Collection
Load a collection of endpoints by alias.}
\usage{
collection(alias)
}
\arguments{
\item{alias}{\verb{<chr>} collection alias}
}
\value{
An S7 \verb{<class_collection>} object.
}
\description{
Endpoint Collection
Load a collection of endpoints by alias.
}
\examples{
collection("caid_unwind")
collection("caid_services")
}
