% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{Contact}
\alias{Contact}
\title{Contact Class}
\usage{
Contact(name = character(0), email = character(0))
}
\arguments{
\item{name}{\verb{<chr>} Contact name}

\item{email}{\verb{<chr>} Contact email}
}
\value{
\verb{<S7_class>} Contact object
}
\description{
Contact Class
}
