% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{ContactPoint}
\alias{ContactPoint}
\title{ContactPoint Class}
\usage{
ContactPoint(type = "vcard:Contact", fn = NULL, hasEmail = NULL)
}
\arguments{
\item{type}{\verb{<chr>} Schema contact type; default is \code{vcard:Contact}}

\item{fn}{\verb{<chr>} contact's full name}

\item{hasEmail}{\verb{<chr>} contact's email}
}
\value{
\verb{<S7_class>} dataset_contactPoint object
}
\description{
\code{ContactPoint} object
}
