% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_backend.R
\name{Current}
\alias{Current}
\alias{CurrentMain}
\alias{CurrentProvider}
\alias{CurrentOpen}
\title{Current}
\usage{
Current(
  title = character(0),
  description = character(0),
  contact = character(0),
  modified = character(0),
  identifier = character(0),
  rows = integer(0),
  fields = character(0),
  pages = integer(0),
  download = character(0)
)

CurrentMain(
  title = character(0),
  description = character(0),
  contact = character(0),
  modified = character(0),
  identifier = character(0),
  rows = integer(0),
  fields = character(0),
  pages = integer(0),
  download = character(0),
  temporal = character(0),
  periodicity = character(0),
  resources = class_resources(),
  dictionary = character(0),
  site = character(0),
  references = character(0)
)

CurrentProvider(
  title = character(0),
  description = character(0),
  contact = character(0),
  modified = character(0),
  identifier = character(0),
  rows = integer(0),
  fields = character(0),
  pages = integer(0),
  download = character(0),
  issued = character(0),
  released = character(0),
  uuid = character(0),
  site = character(0)
)

CurrentOpen(
  title = character(0),
  description = character(0),
  contact = character(0),
  modified = character(0),
  identifier = character(0),
  rows = integer(0),
  fields = character(0),
  pages = integer(0),
  download = character(0),
  uuid = character(0)
)
}
\arguments{
\item{title}{\verb{<chr>} title}

\item{description}{\verb{<chr>} description}

\item{contact}{\verb{<chr>} contact}

\item{modified}{\verb{<chr>} date last modified}

\item{identifier}{\verb{<chr>} uuid}

\item{rows}{\verb{<int>} number of rows}

\item{fields}{\verb{<chr>} field names}

\item{pages}{\verb{<int>} number of pages}

\item{download}{\verb{<chr>} download URL}

\item{temporal}{\verb{<chr>} temporal}

\item{periodicity}{\verb{<chr>} periodicity}

\item{resources}{\verb{<chr>} resources URL}

\item{dictionary}{\verb{<chr>} dictionary}

\item{site}{\verb{<chr>} landing site}

\item{references}{\verb{<chr>} references link}

\item{issued}{\verb{<chr>} date issued}

\item{released}{\verb{<chr>} date released}

\item{uuid}{\verb{<chr>} uuid}
}
\description{
Current

CurrentMain

CurrentProvider

CurrentOpen
}
\seealso{
Other classes: 
\code{\link{class_resources}()}

Other classes: 
\code{\link{class_resources}()}

Other classes: 
\code{\link{class_resources}()}

Other classes: 
\code{\link{class_resources}()}
}
\concept{classes}
