% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_backend.R
\name{CurrentProvider}
\alias{CurrentProvider}
\title{Provider API Endpoint (Current)}
\usage{
CurrentProvider(alias)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint alias}
}
\value{
An S7 \verb{<CurrentProvider>} object.
}
\description{
Provider API Endpoint (Current)
}
\examples{
CurrentProvider("affiliations")
CurrentProvider("clinicians")
CurrentProvider("utilization")

}
