% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{dataset_contactPoint}
\alias{dataset_contactPoint}
\title{dataset_contactPoint Class}
\usage{
dataset_contactPoint(
  type = "vcard:Contact",
  fn = NULL,
  hasEmail = NULL,
  hasURL = NULL
)
}
\arguments{
\item{type}{\verb{<chr>} type}

\item{fn}{\verb{<chr>} contact name}

\item{hasEmail}{\verb{<chr>} contact email}

\item{hasURL}{\verb{<chr>} contact url}
}
\value{
\verb{<S7_class>} object
}
\description{
\code{dataset_contactPoint} object
}
