% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrollees.R
\name{Dataset_enrollee}
\alias{Dataset_enrollee}
\title{Public Provider Enrollment Dataset Object}
\usage{
Dataset_enrollee()
}
\value{
\verb{<S7_class>} object
}
\description{
Public Provider Enrollment Dataset Object
}
\examples{
Dataset_enrollee()

}
