% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_utils.R
\name{delist_elem}
\alias{delist_elem}
\title{Get List Element}
\usage{
delist_elem(x, el)
}
\arguments{
\item{x}{\verb{<list>} list to get element from}

\item{el}{\verb{<chr>} element to get}
}
\value{
\verb{<list>} list with element
}
\description{
Get List Element
}
\keyword{internal}
