% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect}
\alias{detect}
\title{Detect Regular Expression}
\usage{
detect(x, p, n = FALSE)
}
\arguments{
\item{x}{\verb{<chr>} vector to search}

\item{p}{\verb{<chr>} regular expression pattern}

\item{n}{\verb{<lgl>} negate}
}
\value{
\verb{<lgl>} logical vector
}
\description{
Detect Regular Expression
}
\keyword{internal}
