% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{Distribution}
\alias{Distribution}
\title{Distribution Class}
\usage{
Distribution(
  title = character(0),
  type = "dcat:Distribution",
  accessLevel = "public",
  accrualPeriodicity = character(0),
  bureauCode = "009:38",
  contactPoint = dataset_contactPoint(),
  describedBy = character(0),
  description = character(0),
  keyword = character(0),
  landingPage = character(0),
  programCode = "009:000",
  publisher = dataset_Publisher(),
  references = character(0),
  distributions = list()
)
}
\arguments{
\item{title}{\verb{<chr>} Distribution title}

\item{type}{\verb{<chr>} Schema type; default is \code{dcat:Distribution}}

\item{accessLevel}{\verb{<chr>} Distribution access level; default is \code{public}}

\item{accrualPeriodicity}{\verb{<chr>} Distribution update frequency}

\item{bureauCode}{\verb{<chr>} Distribution bureau code; default is \code{009:38}}

\item{contactPoint}{\verb{<S7_class>} Distribution contact}

\item{describedBy}{\verb{<chr>} Hyperlink to Data dictionary}

\item{description}{\verb{<chr>} Distribution description}

\item{keyword}{\verb{<chr>} Hyperlink to API landing page}

\item{landingPage}{\verb{<chr>} Hyperlink to API landing page}

\item{programCode}{\verb{<chr>} Distribution program code; default is \code{009:000}}

\item{publisher}{\verb{<S7_class>} Distribution publisher}

\item{references}{\verb{<chr>} Distribution references}

\item{distributions}{\verb{<S7_class>} \code{data.frame} of available distributions}
}
\value{
\verb{<S7_class>} Distribution object
}
\description{
\code{Distribution} object
}
