% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{endpoint_current}
\alias{endpoint_current}
\title{\code{endpoint_current} Class}
\usage{
endpoint_current(
  title = character(0),
  description = character(0),
  contact = character(0),
  modified = character(0),
  identifier = character(0),
  download = character(0)
)
}
\arguments{
\item{title}{\verb{<chr>}  Endpoint title}

\item{description}{\verb{<chr>}  Endpoint description}

\item{contact}{\verb{<chr>}  Endpoint contact information}

\item{modified}{\verb{<Date>} Date Endpoint data was last modified}

\item{identifier}{\verb{<chr>}  Endpoint uuid or url}

\item{download}{\verb{<chr>}  Endpoint download url}
}
\value{
\code{endpoint_current} object
}
\description{
\code{endpoint_current} Class
}
\seealso{
Other classes: 
\code{\link{endpoint_temporal}()},
\code{\link{main_current}()},
\code{\link{main_resources}()},
\code{\link{provider_current}()}
}
\concept{classes}
