% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{endpoint_temporal}
\alias{endpoint_temporal}
\title{\code{endpoint_temporal} Class}
\usage{
endpoint_temporal(
  title = character(0),
  description = character(0),
  contact = character(0),
  modified = character(0),
  endpoints = list()
)
}
\arguments{
\item{title}{\verb{<chr>} Endpoint title}

\item{description}{\verb{<chr>} Endpoint description}

\item{contact}{\verb{<chr>} Endpoint contact information}

\item{modified}{\verb{<Date>} Date Endpoint data was last modified}

\item{endpoints}{\verb{<data.frame>} Endpoints data.frame}
}
\value{
\code{endpoint_temporal} object
}
\description{
\code{endpoint_temporal} Class
}
\seealso{
Other classes: 
\code{\link{endpoint_current}()},
\code{\link{main_current}()},
\code{\link{main_resources}()},
\code{\link{provider_current}()}
}
\concept{classes}
