% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrollees.R
\name{enrollapi}
\alias{enrollapi}
\title{Public Provider Enrollment Experiment}
\usage{
enrollapi()
}
\value{
\verb{<tibble>} of search results
}
\description{
Public Provider Enrollment Experiment
}
\examples{
enrollapi()

}
