% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{enrolleeAPI}
\alias{enrolleeAPI}
\title{Enrollee API Class}
\usage{
enrolleeAPI(
  title = character(0),
  description = character(0),
  accrualPeriodicity = character(0),
  modified = numeric(0),
  temporal = character(0),
  identifier = Identifier(),
  accessURL = NULL,
  resourcesAPI = NULL,
  downloadURL = character(0),
  describedBy = character(0),
  landingPage = character(0)
)
}
\arguments{
\item{title}{\verb{<chr>} Source}

\item{description}{\verb{<chr>} Source}

\item{accrualPeriodicity}{\verb{<chr>} Source}

\item{modified}{\verb{<chr>} Source}

\item{temporal}{\verb{<chr>} Source}

\item{identifier}{\verb{<Identifier>} Source}

\item{accessURL}{\verb{<chr>} Source}

\item{resourcesAPI}{\verb{<chr>} Source}

\item{downloadURL}{\verb{<chr>} Source}

\item{describedBy}{\verb{<chr>} Source}

\item{landingPage}{\verb{<chr>} Source}
}
\value{
valid \code{enrolleeAPI} classed object
}
\description{
The \code{enrolleeAPI} object contains validated API metadata
}
\examples{
x <- enrollapi()

enrolleeAPI

enroll <- enrolleeAPI(
  x$title,
  x$description,
  x$accrualPeriodicity,
  x$modified,
  x$temporal,
  Identifier(x$identifier),
  x$accessURL,
  x$resourcesAPI,
  x$downloadURL,
  x$describedBy,
  x$landingPage
)

enroll

}
