% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_enrollees.R
\name{enrollees}
\alias{enrollees}
\title{Medicare Provider Enrollees}
\usage{
enrollees(
  npi = NULL,
  pac = NULL,
  enid = NULL,
  specialty_code = NULL,
  specialty_desc = NULL,
  state = NULL,
  first_name = NULL,
  middle_name = NULL,
  last_name = NULL,
  org_name = NULL
)
}
\arguments{
\item{npi}{\verb{<chr>} 10-digit Individual NPI}

\item{pac}{\verb{<chr>} 10-digit PECOS Associate Control (PAC) ID}

\item{enid}{\verb{<chr>} 15-digit Medicare Enrollment ID}

\item{specialty_code}{\verb{<chr>} Enrollment specialty code}

\item{specialty_desc}{\verb{<chr>} Enrollment specialty description}

\item{state}{\verb{<chr>} Enrollment state abbreviation}

\item{first_name, middle_name, last_name}{\verb{<chr>} Individual provider's name}

\item{org_name}{\verb{<chr>} Organizational provider's name}
}
\value{
\verb{<tibble>} of search results
}
\description{
Individual & Organizational Enrollment-level Data
on Providers Actively Approved to Bill Medicare.
}
\section{Periodicity}{

\itemize{
\item Quarterly (R/P3M)
}
}

\section{Links}{

\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-fee-for-service-public-provider-enrollment}{Enrollment API}
\item \href{https://data.cms.gov/resources/medicare-fee-for-service-public-provider-enrollment-data-dictionary}{Enrollment Data Dictionary}
}
}

\examples{
enrollees(enid = "I20040309000221")
enrollees(npi = "1417918293", specialty_code = "14-41")
enrollees(pac = "2860305554")
enrollees(state = "GA")
}
