% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{eval_params}
\alias{eval_params}
\title{Evaluate Query Parameters}
\usage{
eval_params(args, fields)
}
\arguments{
\item{args}{\verb{<list>} of query parameters}

\item{fields}{\verb{<chr>} vector of API field names}
}
\value{
\verb{<list>} of evaluated query parameters
}
\description{
Evaluate Query Parameters
}
\keyword{internal}
