% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R
\name{field_switch}
\alias{field_switch}
\title{Field Switch
Convert raw field names to a known constant.}
\usage{
field_switch(x)
}
\arguments{
\item{x}{A character vector of raw field names.}
}
\value{
A character vector of standardized field names.
}
\description{
Field Switch
Convert raw field names to a known constant.
}
\examples{
field_switch(c("FIRST NAME - OWNER", "FIRST_NAME"))
}
