% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{fields}
\alias{fields}
\title{Retrieve Fields}
\usage{
fields(x)
}
\arguments{
\item{x}{\verb{<S7_class>} Dataset}
}
\value{
\verb{<chr>} vector of API field names
}
\description{
Retrieve Fields
}
\keyword{internal}
