% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_httr2.R
\name{fields_provider}
\alias{fields_provider}
\title{Request field names}
\usage{
fields_provider(url)
}
\arguments{
\item{url}{\verb{<chr>} API URL}
}
\value{
\verb{<chr>} Field names
}
\description{
Request field names
}
\keyword{internal}
