% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_httr2.R
\name{fields_public}
\alias{fields_public}
\title{Request field names from public catalog}
\usage{
fields_public(url)
}
\arguments{
\item{url}{\verb{<chr>} API URL}
}
\value{
\verb{<chr>} Field names
}
\description{
Request field names from public catalog
}
\keyword{internal}
