% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_utils.R
\name{flatten_column}
\alias{flatten_column}
\title{Flatten Column}
\usage{
flatten_column(i)
}
\arguments{
\item{i}{\verb{<list>} list to flatten}
}
\value{
\verb{<chr>} flattened list
}
\description{
Flatten Column
}
\keyword{internal}
