% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{format_query}
\alias{format_query}
\title{Format API Queries}
\usage{
format_query(args, operator = "=")
}
\arguments{
\item{args}{\verb{<list>} of \verb{<chr>} arguments}

\item{operator}{\verb{<chr>} logical operator to use; options are "=" (default),
">=", "<=", ">", "<", "<>", "STARTS_WITH", "ENDS_WITH", "CONTAINS", "IN",
"NOT IN", "BETWEEN", "NOT BETWEEN", "IS NULL", "IS NOT NULL"}
}
\value{
\verb{<list>} of formatted query \verb{<expr>}
}
\description{
Format API Queries
}
\examples{
format_query(list("NPI" = "1417918293", "PECOS_ASCT_CNTL_ID" = NULL))

}
