% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_misc.R
\name{get_data_elem}
\alias{get_data_elem}
\title{Get List Element Named "data"}
\usage{
get_data_elem(x)
}
\arguments{
\item{x}{\verb{<list>} list to get element from}
}
\value{
\verb{<list>} list with element
}
\description{
Get List Element Named "data"
}
\keyword{internal}
