% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_httr2.R
\name{get_nrows}
\alias{get_nrows}
\title{Request total number of rows from catalog}
\usage{
get_nrows(url)
}
\arguments{
\item{url}{\verb{<chr>} API URL}
}
\value{
\verb{<chr>} Field names
}
\description{
Request total number of rows from catalog
}
\keyword{internal}
