% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nrows_fields.R
\name{get_resources}
\alias{get_resources}
\title{Request resources from catalog}
\usage{
get_resources(url)
}
\arguments{
\item{url}{\verb{<chr>} API URL}
}
\value{
\verb{<tibble>} Resources
}
\description{
Request resources from catalog
}
\keyword{internal}
