% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{glimst}
\alias{glimst}
\title{Wrapper for \code{terse::terse()}}
\usage{
glimst(x, p = "- ", w = 0, m = 20, s = " ")
}
\arguments{
\item{x}{\verb{<list>} or \verb{<data.frame>} to be printed}

\item{p}{\verb{<chr>} prefix to be used for each line}

\item{w}{\verb{<int>} target width; 0 = auto; -1 = no limit}

\item{m}{\verb{<int>} maximum vector length anywhere in original object}

\item{s}{\verb{<chr>} separator to be used for each line}
}
\description{
Wrapper for \code{terse::terse()}
}
\keyword{internal}
