% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_misc.R
\name{\%:=\%}
\alias{\%:=\%}
\title{Assignment operator that allows for a more convenient syntax}
\usage{
left \%:=\% right
}
\arguments{
\item{left}{\verb{<symbol>} name of the object to assign}

\item{right}{\verb{<call>} the value to assign to the object}
}
\value{
\verb{<symbol>} the name of the object assigned
}
\description{
Assignment operator that allows for a more convenient syntax
}
