% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_modifiers.R
\name{greater_than_}
\alias{greater_than_}
\alias{less_than_}
\title{Numeric Modifiers}
\usage{
greater_than_(x, or_equals = FALSE)

less_than_(x, or_equals = FALSE)
}
\arguments{
\item{x}{\verb{<num>}  input}

\item{or_equals}{\verb{<lgl>} append "="; default is \code{FALSE}}
}
\value{
\verb{<list>} of class \code{"modifier"}
}
\description{
Numeric Modifiers
}
\examples{
greater_than_(1000)
greater_than_(1000, or_equals = TRUE)

less_than_(1000)
less_than_(1000, or_equals = TRUE)
}
\seealso{
Other modifiers: 
\code{\link{in_}()},
\code{\link{starts_with_}()}

Other modifiers: 
\code{\link{in_}()},
\code{\link{starts_with_}()}
}
\concept{modifiers}
