% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-alias_care.R, R/01-alias_pro.R,
%   R/new_group.R
\name{care_group}
\alias{care_group}
\alias{pro_group}
\alias{groups}
\alias{new_group}
\title{Medicare API Endpoint Groups}
\usage{
care_group(x, call = caller_env(), ...)

pro_group(x, call = caller_env(), ...)

new_group(
  member_names,
  group_name = NULL,
  quick_call = FALSE,
  offset = 0,
  limit = 10
)
}
\arguments{
\item{x}{\verb{<chr>} endpoint alias}

\item{call}{\verb{<env>} environment to use for error reporting}

\item{...}{Additional arguments passed to the group constructor}

\item{member_names}{\verb{<chr>} Alias representing the CMS data endpoint.}

\item{group_name}{\verb{<chr>} Name of the group.}

\item{quick_call}{\verb{<lgl>} call \code{quick_()} method on group if \code{TRUE}.}

\item{offset}{\verb{<int>} Offset for pagination.}

\item{limit}{\verb{<int>} Limit for pagination.}
}
\value{
S7 \code{class_group} object.
}
\description{
Medicare API Endpoint Groups

Provider API Endpoint Groups

Create Groups of Endpoints
}
\examples{
new_group(c("qhp_bus_rule_variables", "managed_care_share"))
new_group(c("qhp_bus_rule_variables", "managed_care_share"), quick_call = TRUE)
pro_group("pro_dialysis")
care_group("care_hospital")
}
