% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_misc.R
\name{handle_na}
\alias{handle_na}
\title{Handle NAs}
\usage{
handle_na(x)
}
\arguments{
\item{x}{\verb{<list>} list to handle}
}
\value{
\verb{<list>} list with NAs handled
}
\description{
Handle NAs
}
\keyword{internal}
