% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_hgov.R
\name{healthcaregov}
\alias{healthcaregov}
\alias{hgov_endpoint}
\alias{hgov_temporal}
\title{Healthcare.Gov API Endpoint Classes}
\usage{
hgov_endpoint(alias)

hgov_temporal(alias)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint alias}
}
\value{
An S7 \verb{<hgov_endpoint>}, or \verb{<hgov_temporal>} object
}
\description{
Healthcare.Gov API Endpoint Classes
}
\examples{
hgov_endpoint("ab_registration_completion")
hgov_temporal("medical_loss_ratio")
}
