% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-alias_hgov.R
\name{hgov}
\alias{hgov}
\alias{hgov_endpoint}
\alias{hgov_temporal}
\title{Healthcare.Gov API Endpoint Classes}
\usage{
hgov_endpoint(alias, call = caller_env())

hgov_temporal(alias, call = caller_env())
}
\arguments{
\item{alias}{\verb{<chr>} endpoint or group alias}

\item{call}{\verb{<env>} environment to use for error reporting}
}
\value{
An S7 \verb{<class_endpoint>} or \verb{<class_temporal>} object
}
\description{
Healthcare.Gov API Endpoint Classes
}
\examples{
hgov_endpoint("ab_registration_completion")
hgov_temporal("medical_loss_ratio")
}
