% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrsa.R
\name{hrsa_facilities}
\alias{hrsa_facilities}
\title{HRSA Facilities}
\usage{
hrsa_facilities(retrieve = NULL)
}
\arguments{
\item{retrieve}{\verb{<chr>} String indicating what to retrieve.}
}
\value{
A list of endpoints.
}
\description{
HRSA Facilities
}
\examples{
hrsa_facilities()

hrsa_facilities("items")
}
