% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_count.R
\name{identifier}
\alias{identifier}
\title{Return the identifier (URL) for a given object}
\usage{
identifier(obj)
}
\arguments{
\item{obj}{An object of class \verb{<class_endpoint>} or \verb{<class_temporal>}.}
}
\description{
Return the identifier (URL) for a given object
}
\keyword{internal}
