% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_modifiers.R
\name{in_}
\alias{in_}
\alias{equals_}
\alias{is_blank_}
\title{General Modifiers}
\usage{
in_(x, negate = FALSE)

equals_(x, negate = FALSE)

is_blank_(negate = FALSE)
}
\arguments{
\item{x}{\verb{<chr>} input}

\item{negate}{\verb{<lgl>} prepend "NOT"; default is \code{FALSE}}
}
\value{
\verb{<list>} of class \code{"modifier"}
}
\description{
General Modifiers
}
\examples{
in_(state.abb[10:15])
in_(state.abb[10:15], negate = TRUE)

equals_(1000)
equals_(1000, negate = TRUE)

is_blank_()
is_blank_(negate = TRUE)
}
\seealso{
Other modifiers: 
\code{\link{greater_than_}()},
\code{\link{starts_with_}()}
}
\concept{modifiers}
