% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_complete_with_limit}
\alias{is_complete_with_limit}
\title{Helper for \code{iterate_with_offset}}
\usage{
is_complete_with_limit(limit)
}
\arguments{
\item{limit}{\verb{<int>} API rate limit, i.e. the maximum number of results an
API will return per request.}
}
\value{
\verb{<function>} Function to check if API request is complete
}
\description{
Helper for \code{iterate_with_offset}
}
\keyword{internal}
