% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_modifiers.R
\name{is_modifier}
\alias{is_modifier}
\title{Identify Query Modifier}
\usage{
is_modifier(x)
}
\arguments{
\item{x}{input}
}
\value{
\verb{<lgl>} TRUE or FALSE
}
\description{
Identify Query Modifier
}
\examples{
is_modifier(greater_than_(1000))
}
\keyword{internal}
