% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{load_endpoint}
\alias{load_endpoint}
\alias{endpoint}
\alias{collection}
\alias{group}
\title{Load Data Endpoints}
\usage{
endpoint(alias)

collection(alias)

group(..., .description = NULL)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint or collection alias}

\item{...}{Arguments passed to methods.}

\item{.description}{\verb{<chr>} Group description. Defaults to \code{NULL},
which will use the aliases as the description.}
}
\value{
An S7 \verb{<class_care/caid/prov/open/hgov/collection/group>} object.
}
\description{
Load a data endpoint or collection by its alias.
}
\details{
The \code{endpoint()} function loads a single data endpoint, while the
\code{collection()} function loads a collection of endpoints, and the \code{group()}
function loads a group of endpoints. The \code{alias} parameter is used to specify
the endpoint or collection alias. If the alias is not found, an error will be
raised.
}
\examples{
endpoint("dial_facility")
endpoint("man_mltss")
endpoint("ab_reg_comp")
endpoint("asc_facility")
endpoint("dial_listing")
collection("unwind")
collection("managed")
try(collection(c("asc_facility", "enterprise")))
try(collection("asc_facility"))
group("asc_facility", "enterprise")
try(group("asc_facility"))
try(group("util"))
}
