% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{load_openpayments}
\alias{load_openpayments}
\title{CMS Open Payments Catalog}
\usage{
load_openpayments()
}
\value{
\verb{<tibble>} of Open Payments API catalog information
}
\description{
CMS Open Payments Catalog
}
\examples{
load_openpayments()

}
