% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{load_provider}
\alias{load_provider}
\title{CMS Provider Catalog}
\usage{
load_provider()
}
\value{
\verb{<tibble>} of Provider API catalog information
}
\description{
CMS Provider Catalog
}
\examples{
load_provider()

}
