% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{load_public}
\alias{load_public}
\title{CMS Public Catalog}
\usage{
load_public()
}
\value{
\verb{<list>} of \verb{<tibbles>}: \code{dataset}, \code{distribution}
}
\description{
CMS Public Catalog
}
\examples{
load_public()

}
