% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{luhn_check}
\alias{luhn_check}
\title{Luhn Algorithm Check for NPIs}
\usage{
luhn_check(x)
}
\arguments{
\item{x}{\verb{<chr>} NPI}
}
\value{
\verb{<lgl>} \code{TRUE} if valid NPI, \code{FALSE} otherwise
}
\description{
Luhn Algorithm Check for NPIs
}
\examples{
luhn_check("1417918293")

luhn_check("1234567890")

luhn_check("123456789")

}
