% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_backend.R, R/S7_method_resources.R
\name{MainCurrent}
\alias{MainCurrent}
\alias{careTemp}
\alias{list_resources}
\title{Current Main API Endpoint}
\usage{
MainCurrent(alias)

careTemp(alias)

list_resources(x)
}
\arguments{
\item{alias}{\verb{<chr>} title alias}

\item{x}{Object of class \code{MainCurrent} or \code{careTemp}}
}
\value{
An S7 \verb{<MainCurrent>} object.

An S7 \verb{<careTemp>} object.

A list of API resources
}
\description{
Current Main API Endpoint

Medicare Temporal Endpoint

List resources
}
\examples{
MainCurrent("enrollees")
MainCurrent("opt_out")
MainCurrent("order_refer")
MainCurrent("reassignments")
MainCurrent("hospitals")
MainCurrent("laboratories")
MainCurrent("crosswalk")
MainCurrent("rbcs")
MainCurrent("facilities")
MainCurrent("home_health")
MainCurrent("hospice")
MainCurrent("dialysis")
MainCurrent("skilled_nursing")
careTemp("quality_payment")
MainCurrent("enrollees") |> list_resources()
careTemp("quality_payment") |> list_resources()

}
