% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{main_current}
\alias{main_current}
\title{\code{main_current} Class}
\usage{
main_current(
  title = character(0),
  description = character(0),
  contact = character(0),
  modified = character(0),
  identifier = character(0),
  download = character(0),
  temporal = character(0),
  periodicity = character(0),
  resources = main_resources(),
  dictionary = character(0),
  site = character(0),
  references = character(0)
)
}
\arguments{
\item{title}{\verb{<chr>}  API endpoint title}

\item{description}{\verb{<chr>}  API endpoint description}

\item{contact}{\verb{<chr>}  Contact information}

\item{modified}{\verb{<Date>} Date data was last modified}

\item{identifier}{\verb{<chr>}  Query url}

\item{download}{\verb{<chr>}  Download url}

\item{temporal}{\verb{<chr>}  Timespan covered by data}

\item{periodicity}{\verb{<chr>}  Update frequency}

\item{resources}{\verb{<chr>}  Files available for download}

\item{dictionary}{\verb{<chr>}  Link to data dictionary}

\item{site}{\verb{<chr>}  Link to landing page}

\item{references}{\verb{<chr>}  Link to references}
}
\value{
\code{main_current} object
}
\description{
\code{main_current} Class
}
\seealso{
Other classes: 
\code{\link{endpoint_current}()},
\code{\link{endpoint_temporal}()},
\code{\link{main_resources}()},
\code{\link{provider_current}()}
}
\concept{classes}
