% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_main.R
\name{main_dataset}
\alias{main_dataset}
\title{Load Main API \code{Dataset}}
\usage{
main_dataset(dataset)
}
\arguments{
\item{dataset}{\verb{<chr>} dataset title}
}
\value{
\verb{<Dataset>} object
}
\description{
Load Main API \code{Dataset}
}
\examples{
main_dataset("enrollees")

main_dataset("hospitals")

main_dataset("reassignments")

main_dataset("opt_out")

main_dataset("laboratories")

}
