% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{main_resources}
\alias{main_resources}
\title{\code{main_resources} Class}
\usage{
main_resources(url = character(0))
}
\arguments{
\item{url}{\verb{<chr>} default is \code{NA}}
}
\value{
\code{main_resources} object
}
\description{
\code{main_resources} Class
}
\seealso{
Other classes: 
\code{\link{endpoint_current}()},
\code{\link{endpoint_temporal}()},
\code{\link{main_current}()},
\code{\link{provider_current}()}
}
\concept{classes}
