% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_misc.R
\name{main_temp}
\alias{main_temp}
\title{Format temporal string}
\usage{
main_temp(x)
}
\arguments{
\item{x}{\verb{<chr>} string to format}
}
\value{
\verb{<chr>} formatted string
}
\description{
Format temporal string
}
\keyword{internal}
