% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_main.R
\name{main_temp_dims}
\alias{main_temp_dims}
\title{Get Dimensions of Temporal Main Endpoint}
\usage{
main_temp_dims(url)
}
\arguments{
\item{url}{\verb{<chr>} endpoint URL}
}
\value{
\verb{<list>} number of rows and field names
}
\description{
Get Dimensions of Temporal Main Endpoint
}
\keyword{internal}
