% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_main.R
\name{main_temporal}
\alias{main_temporal}
\title{Load \verb{<TemporalMain>} API Endpoint}
\usage{
main_temporal(alias)
}
\arguments{
\item{alias}{\verb{<chr>} dataset title}
}
\value{
\verb{<TemporalMain>} object
}
\description{
Load \verb{<TemporalMain>} API Endpoint
}
\examples{
main_temporal("quality_payment")
}
