% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_main.R
\name{main_temporal_group}
\alias{main_temporal_group}
\title{Load Public API \code{Distribution}}
\usage{
main_temporal_group(alias)
}
\arguments{
\item{alias}{\verb{<chr>} dataset title}
}
\value{
\verb{<Distribution>} object
}
\description{
Load Public API \code{Distribution}
}
\examples{
main_temporal_group("utilization")
main_temporal_group("prescribers")
main_temporal_group("suppliers")
main_temporal_group("outpatient")
main_temporal_group("inpatient")
}
