% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_misc.R
\name{map_na_if}
\alias{map_na_if}
\title{Vectorized \code{na_if}}
\usage{
map_na_if(x)
}
\arguments{
\item{x}{\verb{<list>} list to handle}
}
\value{
\verb{<list>} list with NAs handled
}
\description{
Vectorized \code{na_if}
}
\keyword{internal}
