% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_request.R
\name{map_parse_json_response}
\alias{map_parse_json_response}
\title{Mapped Parse JSON response}
\usage{
map_parse_json_response(response)
}
\arguments{
\item{response}{\verb{<httr_response>} API response}
}
\value{
\verb{<tibble>} Parsed JSON response as tibble
}
\description{
Mapped Parse JSON response
}
\keyword{internal}
