% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-alias_caid.R
\name{medicaid}
\alias{medicaid}
\alias{caid_endpoint}
\alias{caid_temporal}
\alias{caid_group}
\title{Medicaid API Endpoints}
\usage{
caid_endpoint(alias, call = caller_env())

caid_temporal(alias, call = caller_env())

caid_group(alias, call = caller_env(), ...)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint or group alias}

\item{call}{\verb{<env>} environment to use for error reporting}

\item{...}{Additional arguments passed to the group constructor}
}
\value{
An S7 \verb{<class_endpoint>}, \verb{<class_temporal>} or \verb{<class_group>} object
}
\description{
Medicaid API Endpoints
}
\examples{
\dontshow{if (rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
caid_endpoint("managed_longterm")
caid_temporal("state_drug_util")
caid_group("caid_dual_status")
\dontshow{\}) # examplesIf}
}
