% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-alias_care.R, R/02-alias_care.R
\name{medicare}
\alias{medicare}
\alias{care_endpoint}
\alias{care_temporal}
\alias{care_group}
\alias{care_group2}
\title{Medicare API Endpoint Classes}
\usage{
care_endpoint(alias, call = caller_env())

care_temporal(alias, call = caller_env())

care_group(alias, call = caller_env(), ...)

care_group2(alias, call = caller_env(), ...)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint alias}

\item{call}{\verb{<env>} environment to use for error reporting}

\item{...}{Additional arguments passed to the group constructor}
}
\value{
An S7 \verb{<care_endpoint>}, or \verb{<care_temporal>} object
}
\description{
Medicare API Endpoint Classes
}
\examples{
\dontshow{if (rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
care_endpoint("care_dialysis")
care_temporal("quality_payment")
care_group("care_hospital")
\dontshow{\}) # examplesIf}
}
