% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_modifiers.R
\name{modifier_}
\alias{modifier_}
\title{Query Modifier Constructor}
\usage{
modifier_(operator, value, allow)
}
\arguments{
\item{operator}{\verb{<chr>} comparison operator}

\item{value}{\verb{<any>} value to compare against}

\item{allow}{\verb{<chr>} allowed endpoint class(es) for this modifier}
}
\value{
An object of class \code{"modifier"}
}
\description{
Query Modifier Constructor
}
\examples{
modifier_(">", 1000, "all")
}
\keyword{internal}
