% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{modifiers}
\alias{modifiers}
\alias{equal}
\alias{any_of}
\alias{none_of}
\alias{not_equal}
\alias{between}
\alias{not_between}
\alias{greater_than}
\alias{less_than}
\alias{starts_with}
\alias{ends_with}
\alias{contains}
\title{Query Modifiers}
\source{
\href{https://jsonapi.org/format/#query-parameters}{JSON-API: Query Parameters}
}
\usage{
equal(x)

any_of(...)

none_of(...)

not_equal(x)

between(...)

not_between(...)

greater_than(x, or_equal = FALSE)

less_than(x, or_equal = FALSE)

starts_with(x)

ends_with(x)

contains(x)
}
\arguments{
\item{x}{input}

\item{...}{input}

\item{or_equal}{\verb{<lgl>} append \code{=} to \code{>} or \code{<}}
}
\value{
An S7 \verb{<class_modifier>} object.
}
\description{
Helpers for use in constructing conditions in queries.
}
\details{
Query modifiers are part of a small DSL for use in query construction, in the \href{https://www.drupal.org/docs/core-modules-and-themes/core-modules/jsonapi-module/filtering}{JSON-API} format.
}
\examples{
equal(1000)
any_of(state.abb[10:15])
none_of(state.abb[1:3])
not_equal(10000.23)
between(1000, 1100, 125)
not_between(0.95, 0.67, 0.75)
greater_than(1000)
greater_than(0.125, or_equal = TRUE)
less_than(1000)
less_than(0.125, or_equal = TRUE)
starts_with("foo")
ends_with("bar")
contains("baz")
}
