% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_count.R
\name{ncount}
\alias{ncount}
\title{Return the number of results for a given query}
\usage{
ncount(obj, url, qstr = NULL, ...)
}
\arguments{
\item{obj}{An object of class \verb{<class_endpoint>} or \verb{<class_temporal>}.}

\item{url}{An object of class \verb{<class_query>}. If \code{NULL}, the default query for the endpoint is used.}

\item{qstr}{A \verb{<chr>} string representing the query to be passed to the API. If \code{NULL}, the default query for the endpoint is used.}

\item{...}{Additional arguments passed to the query.}
}
\value{
An \verb{<int>} representing the number of results found.
}
\description{
Return the number of results for a given query
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
end_dial <- endpoint("care_dial_end")
end_qpp  <- endpoint("quality_payment")
end_rhc  <- collection("care_rhc")
qry      <- query(
              STATE = in_(c("CA", "GA", "NY"), care = TRUE),
              `STATE - OWNER` = in_(c("CA", "GA", "NY"), care = TRUE),
              .type = "medicare")

ncount(end_dial, identifier(end_dial), qry)

#ncount(end_rhc, identifier(end_rhc), qry)
\dontshow{\}) # examplesIf}
}
