% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_group.R
\name{new_collection}
\alias{new_collection}
\title{Create Collection of Endpoints}
\usage{
new_collection(alias, call = caller_env())
}
\arguments{
\item{alias}{\verb{<chr>} Alias representing the CMS data endpoint.}

\item{call}{\verb{<env>} Environment to use for error reporting.}
}
\value{
S7 \code{class_collection} object.
}
\description{
Create Collection of Endpoints
}
\examples{
\dontshow{if (rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
new_collection("caid_demographics")
new_collection("caid_unwind")
new_collection("caid_services") |> query_nresults()
\dontshow{\}) # examplesIf}
}
