% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-new_end.R
\name{new_endpoint}
\alias{new_endpoint}
\title{API Endpoints}
\usage{
new_endpoint(alias)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint alias}
}
\value{
An S7 object
}
\description{
API Endpoints
}
\examples{
\dontshow{if (rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
new_endpoint("care_dial_end")
new_endpoint("care_dial_tmp")
new_endpoint("quality_payment")
new_endpoint("managed_longterm")
new_endpoint("state_drug_util")
new_endpoint("hgov_ab_reg_comp")
new_endpoint("hgov_puf_rate")
new_endpoint("profile_covered")
new_endpoint("grouped_state_nature")
new_endpoint("asc_facility")
new_endpoint("prov_dialysis")
\dontshow{\}) # examplesIf}
}
