% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-new_end.R
\name{new_group}
\alias{new_group}
\title{Create Groups of Endpoints}
\usage{
new_group(aliases, description = NULL)
}
\arguments{
\item{aliases}{\verb{<chr>} Alias representing the CMS data endpoint.}

\item{description}{\verb{<chr>} Name of the group.}
}
\value{
S7 \code{class_group} object.
}
\description{
Create Groups of Endpoints
}
\examples{
\dontshow{if (rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
new_group(c("cahps_hhc_patient", "hgov_qhp_business"))
new_group(c("hgov_local_help", "hgov_qhp_business"))
\dontshow{\}) # examplesIf}
}
