% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_request.R
\name{new_request}
\alias{new_request}
\title{Create a new request by class}
\arguments{
\item{x}{An object of class \code{CurrentMain}.}
}
\value{
A new request.
}
\description{
Create a new request by class
}
\examples{
x <- CurrentMain("enrollees")
new_request(x)

}
