% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_nppes.R
\name{npi_nlm}
\alias{npi_nlm}
\title{Search the NLM NPI Registry}
\source{
\href{https://clinicaltables.nlm.nih.gov/apidoc/npi_idv/v3/doc.html}{API Documentation: Individuals}

\href{https://clinicaltables.nlm.nih.gov/apidoc/npi_org/v3/doc.html}{API Documentation: Organizations}
}
\usage{
npi_nlm(terms, npi = NULL)
}
\arguments{
\item{terms}{\verb{<chr>} Search terms, separated by spaces}

\item{npi}{\verb{<chr>} Search terms, separated by spaces}
}
\value{
\verb{<tibble>} of search results
}
\description{
Search the NLM NPI Registry
}
\examples{
npi_nlm("john bethesda")
npi_nlm("Wiregrass Georgia")
npi_nlm("Dentist Valdosta")
npi_nlm("Valdosta")
npi_nlm("Atlanta")
}
