% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_nppes.R
\name{npi_nppes}
\alias{npi_nppes}
\title{Search the NPPES NPI Registry}
\source{
\href{https://npiregistry.cms.hhs.gov/api-page}{API Documentation}
}
\usage{
npi_nppes(
  npi = NULL,
  entity = NULL,
  first = NULL,
  last = NULL,
  organization = NULL,
  name_type = NULL,
  taxonomy_desc = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  country = NULL
)
}
\arguments{
\item{npi}{\verb{<chr>} Search terms, separated by spaces}

\item{entity}{\verb{<chr>} Entity type: \code{individual} or \code{organization}}

\item{first}{\verb{<chr>} Individual provider's first name}

\item{last}{\verb{<chr>} Individual provider's last name}

\item{organization}{\verb{<chr>} Organizational provider's name}

\item{name_type}{\verb{<chr>} Name purpose: \code{legal} or \code{de}}

\item{taxonomy_desc}{\verb{<chr>} Taxonomy description}

\item{city}{\verb{<chr>} City}

\item{state}{\verb{<chr>} State}

\item{zip}{\verb{<chr>} Zip code}

\item{country}{\verb{<chr>} Country code}
}
\value{
\verb{<tibble>} of search results
}
\description{
Search the NPPES NPI Registry
}
\examples{
#npi_nppes(npi = npi_ex$v)
}
