% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_nppes.R
\name{npi_nlm}
\alias{npi_nlm}
\alias{npi_nppes}
\title{Search the NLM NPI Registry}
\source{
\href{https://clinicaltables.nlm.nih.gov/apidoc/npi_idv/v3/doc.html}{API Documentation: Individuals}

\href{https://clinicaltables.nlm.nih.gov/apidoc/npi_org/v3/doc.html}{API Documentation: Organizations}

\href{https://npiregistry.cms.hhs.gov/api-page}{API Documentation}
}
\usage{
npi_nlm(terms, npi = NULL)

npi_nppes(
  npi = NULL,
  entity = NULL,
  first = NULL,
  last = NULL,
  organization = NULL,
  name_type = NULL,
  taxonomy_desc = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  country = NULL
)
}
\arguments{
\item{terms}{\verb{<chr>} Search terms, separated by spaces}

\item{npi}{\verb{<chr>} Search terms, separated by spaces}

\item{entity}{\verb{<chr>} Entity type: \code{individual} or \code{organization}}

\item{first}{\verb{<chr>} Individual provider's first name}

\item{last}{\verb{<chr>} Individual provider's last name}

\item{organization}{\verb{<chr>} Organizational provider's name}

\item{name_type}{\verb{<chr>} Name purpose: \code{legal} or \code{de}}

\item{taxonomy_desc}{\verb{<chr>} Taxonomy description}

\item{city}{\verb{<chr>} City}

\item{state}{\verb{<chr>} State}

\item{zip}{\verb{<chr>} Zip code}

\item{country}{\verb{<chr>} Country code}
}
\value{
\verb{<tibble>} of search results

\verb{<tibble>} of search results
}
\description{
Search the NLM NPI Registry

Search the NPPES NPI Registry
}
\examples{
npi_nlm("john bethesda")
npi_nlm("Wiregrass Georgia")
npi_nlm("Dentist Valdosta")
npi_nlm("Valdosta")
npi_nlm("Atlanta")
npi_nppes(npi = npi_ex$k[1:2]) |> str()
npi_nppes(npi = npi_ex$k[1]) |> str()
}
