% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_nppes.R
\name{nppes_nlm}
\alias{nppes_nlm}
\title{Search the NLM NPI Registry}
\source{
API Documentation: \url{https://clinicaltables.nlm.nih.gov/api/npi_idv/v3}
}
\usage{
nppes_nlm(terms)
}
\arguments{
\item{terms}{\verb{<chr>} Search terms, separated by spaces}
}
\value{
\verb{<tibble>} of search results
}
\description{
Search the NLM NPI Registry
}
\examples{
nppes_nlm("john bethesda")

nppes_nlm("Dentist Valdosta")

# Not returning all results
nppes_nlm("Valdosta")
}
