% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_httr2.R
\name{nrows_provider}
\alias{nrows_provider}
\title{Request number of results from provider catalog}
\usage{
nrows_provider(url)
}
\arguments{
\item{url}{\verb{<chr>} API URL}
}
\value{
\verb{<int>} Number of results
}
\description{
Request number of results from provider catalog
}
\keyword{internal}
