% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2_get_nrows_fields.R
\name{nrows_public}
\alias{nrows_public}
\title{Request number of results from public catalog}
\usage{
nrows_public(url)
}
\arguments{
\item{url}{\verb{<chr>} API URL}
}
\value{
\verb{<int>} Number of results
}
\description{
Request number of results from public catalog
}
\keyword{internal}
