% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cli.R
\name{num}
\alias{num}
\title{Format Number with Commas}
\usage{
num(x)
}
\arguments{
\item{x}{\verb{<int>} number to format}
}
\value{
\verb{<chr>} formatted number
}
\description{
Format Number with Commas
}
\keyword{internal}
