% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_misc.R
\name{offset}
\alias{offset}
\alias{offset_seq}
\alias{offset_size}
\title{Generate API Request "Offset" Sequence}
\usage{
offset_seq(n, limit)

offset_size(n, limit)
}
\arguments{
\item{n}{\verb{<int>} Number of results returned in an API request}

\item{limit}{\verb{<int>} API rate limit, i.e. the maximum number of results an
API will return per request.}
}
\value{
\verb{<int>} If \code{n <= limit}, simply returns \code{n}. If \code{n > limit}, an
integer sequence is returned, beginning at \code{0} and of
length equal to \code{celing(n / limit)}.
}
\description{
Generate API Request "Offset" Sequence
}
\examples{
offset_seq(100, 10)
offset_size(100, 10)

offset_seq(10, 100)
offset_size(10, 100)

offset_seq(47984, 5000)
offset_size(47984, 5000)

offset_seq(147984, 2000)
offset_size(147984, 2000)
}
