% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{offset_sequence}
\alias{offset_sequence}
\title{Generate API Request "Offset" Sequence}
\usage{
offset_sequence(nobs, limit)
}
\arguments{
\item{nobs}{\verb{<int>} Number of results returned in the API request}

\item{limit}{\verb{<int>} Maximum number of results an API will return per request}
}
\value{
\verb{<int>} If \code{nobs > limit}, an integer sequence beginning at \code{0} of
length equal to \code{nobs / limit + 1}. If \code{nobs <= limit}, the function
simply returns \code{nobs}.
}
\description{
Generate API Request "Offset" Sequence
}
\examples{
offset_sequence(100, 10)

offset_sequence(10, 100)

offset_sequence(47984, 5000)

offset_sequence(147984, 2000)

}
