% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_backend.R
\name{OpenCurrent}
\alias{OpenCurrent}
\alias{OpenTemporal}
\alias{OpenTemporalGroup}
\title{Current Open Payments API Endpoint}
\usage{
OpenCurrent(alias)

OpenTemporal(alias)

OpenTemporalGroup(
  title = character(0),
  description = character(0),
  contact = character(0),
  rows = integer(0),
  pages = integer(0),
  fields = character(0),
  years = integer(0),
  endpoints = list()
)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint alias}

\item{title}{\verb{<chr>} title}

\item{description}{\verb{<chr>} description}

\item{contact}{\verb{<chr>} contact name and email address}

\item{rows}{\verb{<int>} number of rows}

\item{pages}{\verb{<int>} number of pages}

\item{fields}{\verb{<chr>} field names}

\item{years}{\verb{<int>} years available}

\item{endpoints}{\verb{<data.frame>} endpoints}
}
\value{
An S7 \verb{<OpenCurrent>} object.

An S7 \verb{<OpenTemporal>} object.

An S7 \verb{<TemporalOpenGroup>} object.
}
\description{
Current Open Payments API Endpoint

Temporal Open Payments API Endpoint

Open Payments API Endpoint Temporal Group
}
\examples{
OpenCurrent("prof_cov")
OpenCurrent("prof_phys")
OpenCurrent("prof_info")
OpenCurrent("prof_map")
OpenCurrent("prof_entity")
OpenCurrent("prof_teach")
OpenCurrent("dashboard")
OpenCurrent("pay_state_total")
OpenCurrent("pay_state_group")
OpenCurrent("pay_nat_group")
OpenCurrent("pay_nat_total")

OpenTemporal("general")
OpenTemporal("ownership")
OpenTemporal("research")
OpenTemporal("recipient_nature")
OpenTemporal("recipient_entity")
OpenTemporal("entity_nature")
OpenTemporal("entity_recipient_nature")
OpenTemporal("state_nature")
}
