% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_open_payments.R
\name{open_add_dlurl}
\alias{open_add_dlurl}
\title{Join Vector of Download URLs to Main Dataset}
\usage{
open_add_dlurl(x)
}
\arguments{
\item{x}{data.frame}
}
\value{
data.frame
}
\description{
Join Vector of Download URLs to Main Dataset
}
\keyword{internal}
