% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_open_payments.R
\name{open_catalog}
\alias{open_catalog}
\title{Load Open Payments Catalog}
\usage{
open_catalog()
}
\value{
\verb{<list>} of Open Payments API catalog information
}
\description{
Load Open Payments Catalog
}
\examples{
open_catalog()
}
