% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_open_payments.R
\name{open_clean_grouped}
\alias{open_clean_grouped}
\title{Clean Open Payments Grouped Data}
\usage{
open_clean_grouped(x)
}
\arguments{
\item{x}{data.frame}
}
\value{
data.frame
}
\description{
Clean Open Payments Grouped Data
}
\keyword{internal}
