% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_open_payments.R
\name{open_clean_temp}
\alias{open_clean_temp}
\title{Clean Open Payments Temporal Data}
\usage{
open_clean_temp(x)
}
\arguments{
\item{x}{data.frame}
}
\value{
data.frame
}
\description{
Clean Open Payments Temporal Data
}
\keyword{internal}
