% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_open.R
\name{open_current}
\alias{open_current}
\title{Load \verb{<CurrentOpen>} API Endpoint}
\usage{
open_current(alias)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint alias}
}
\value{
\verb{<CurrentOpen>} object
}
\description{
Load \verb{<CurrentOpen>} API Endpoint
}
\examples{
open_current("prof_cov")
open_current("prof_phys")
open_current("prof_info")
open_current("prof_map")
open_current("prof_entity")
open_current("prof_teach")
open_current("dashboard")
open_current("pay_state_total")
open_current("pay_state_group")
open_current("pay_nat_group")
open_current("pay_nat_total")
}
