% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_dashboard.R
\name{open_dashboard}
\alias{open_dashboard}
\title{Open Payments Summary Dashboard}
\usage{
open_dashboard()
}
\value{
A \verb{<tibble>}
}
\description{
Open Payments Summary Dashboard
}
\examples{
open_dashboard()
}
