% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_Dataset_Distribution.R
\name{open_Dataset}
\alias{open_Dataset}
\title{Load Open Payments API \code{Dataset}}
\usage{
open_Dataset(period = NULL, group = NULL, dataset = NULL)
}
\arguments{
\item{period}{\verb{<chr>} dataset year}

\item{group}{\verb{<chr>} dataset group keyword}

\item{dataset}{\verb{<chr>} dataset title}
}
\value{
\verb{<Dataset>} object
}
\description{
Load Open Payments API \code{Dataset}
}
\examples{
open_Dataset(period = "All", group = "Summary", dataset = "Covered Recipient Profile Supplement")

}
