% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ref.R
\name{open_dictionary}
\alias{open_dictionary}
\title{Open Payments Dictionary}
\usage{
open_dictionary()
}
\value{
\verb{<tibble>} of search results
}
\description{
Open Payments Dictionary
}
\examples{
open_dictionary()
}
\keyword{internal}
