% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_open_payments.R
\name{open_general}
\alias{open_general}
\title{Load Open Payments General Data}
\usage{
open_general()
}
\value{
\verb{<list>} of Open Payments API general data
}
\description{
Load Open Payments General Data
}
\examples{
open_general()
}
