% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_old_open.R
\name{open_grouped_summary}
\alias{open_grouped_summary}
\title{Load Open Payments Grouped Summaries}
\usage{
open_grouped_summary()
}
\value{
\verb{<list>} of Open Payments API grouped information
}
\description{
Load Open Payments Grouped Summaries
}
\examples{
open_grouped_summary()
}
