% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_open.R
\name{open_grouped_yearly}
\alias{open_grouped_yearly}
\title{Load Open Payments Grouped by Year}
\usage{
open_grouped_yearly()
}
\value{
\verb{<list>} of Open Payments API grouped information
}
\description{
Load Open Payments Grouped by Year
}
\examples{
open_grouped_yearly()
}
