% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_open_payments.R
\name{open_nrows_fields}
\alias{open_nrows_fields}
\title{Get Field Names and Number of Rows from Open Payments Endpoint}
\usage{
open_nrows_fields(uuid)
}
\arguments{
\item{uuid}{\verb{<chr>} Endpoint UUID}
}
\value{
\verb{<list>} of number of rows and field names
}
\description{
Get Field Names and Number of Rows from Open Payments Endpoint
}
\keyword{internal}
