% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_open_payments.R
\name{open_profiles}
\alias{open_profiles}
\title{Load Open Payments Profile}
\usage{
open_profiles(x)
}
\arguments{
\item{x}{\verb{<list>} of Open Payments API catalog information}
}
\value{
\verb{<list>} of Open Payments API profile information
}
\description{
Load Open Payments Profile
}
\examples{
open_profiles(open_catalog())
}
