% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_open.R
\name{open_research}
\alias{open_research}
\title{Load Open Payments Research Data}
\usage{
open_research()
}
\value{
\verb{<list>} of Open Payments API research data
}
\description{
Load Open Payments Research Data
}
\examples{
open_research()
}
