% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_summary.R
\name{open_summary}
\alias{open_summary}
\alias{open_dashboard}
\alias{open_national}
\alias{open_dictionary}
\title{Open Payments Summaries}
\usage{
open_dashboard()

open_national()

open_dictionary()
}
\value{
A \verb{<tibble>}
}
\description{
Open Payments Summaries
}
\examples{
open_dashboard()
open_national()
open_dictionary()
}
