% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_open.R
\name{open_temporal}
\alias{open_temporal}
\title{Load \verb{<TemporalOpen>} API Endpoint}
\usage{
open_temporal(alias)
}
\arguments{
\item{alias}{\verb{<chr>} dataset title}
}
\value{
\verb{<TemporalOpen>} object
}
\description{
Load \verb{<TemporalOpen>} API Endpoint
}
\examples{
open_temporal("general")
open_temporal("ownership")
open_temporal("research")
open_temporal("recipient_nature")
open_temporal("recipient_entity")
open_temporal("entity_nature")
open_temporal("entity_recipient_nature")
open_temporal("state_nature")
}
