% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_open.R
\name{open_url}
\alias{open_url}
\title{Open Payments UUID to URL}
\usage{
open_url(uuid)
}
\arguments{
\item{uuid}{\verb{<chr>} endpoint UUID}
}
\value{
\verb{<chr>} endpoint URL
}
\description{
Open Payments UUID to URL
}
\keyword{internal}
