% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_open.R
\name{open_uuid_url}
\alias{open_uuid_url}
\title{Convert Open Payments UUID to URL}
\usage{
open_uuid_url(x)
}
\arguments{
\item{x}{\verb{<chr>} UUID}
}
\value{
\verb{<chr>} URL
}
\description{
Convert Open Payments UUID to URL
}
\keyword{internal}
