% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_open.R
\name{openGroup}
\alias{openGroup}
\title{Open Payments Endpoint Group}
\usage{
openGroup(alias)
}
\arguments{
\item{alias}{\verb{<chr>} title alias}
}
\value{
An S7 \verb{<openGroup>} object.
}
\description{
Open Payments Endpoint Group
}
\examples{
openGroup("profile")
}
