% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_open.R
\name{openpayments}
\alias{openpayments}
\alias{open_endpoint}
\alias{open_temporal}
\title{Open Payments API Endpoint Classes}
\usage{
open_endpoint(alias)

open_temporal(alias)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint alias}
}
\value{
An S7 \verb{<open_endpoint>} or \verb{<open_temporal>} object
}
\description{
Open Payments API Endpoint Classes
}
\examples{
open_endpoint("profile_covered")
open_temporal("grouped_state_nature")
}
