% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_dashboard.R
\name{OpenResp}
\alias{OpenResp}
\alias{openRespDash}
\alias{open_dashboard}
\title{Open Dashboard Response Class}
\usage{
openRespDash(response = list())

open_dashboard()
}
\arguments{
\item{response}{\verb{<openRespDash>} object returned from the API call to the dashboard.}
}
\value{
An S7 \verb{<openRespDash>} object.
}
\description{
Open Dashboard Response Class
}
\examples{
open_dashboard()
}
