% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_open.R
\name{openTemp}
\alias{openTemp}
\title{Open Payments Temporal Endpoint}
\usage{
openTemp(alias)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint alias}
}
\value{
An S7 \verb{<openTemp>} object.
}
\description{
Open Payments Temporal Endpoint
}
\examples{
openTemp("general") |> str()
openTemp("ownership") |> str()
openTemp("research") |> str()
}
