% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_open.R
\name{openTempGroup}
\alias{openTempGroup}
\title{Group of Open Payments Temporal Endpoints}
\usage{
openTempGroup(alias)
}
\arguments{
\item{alias}{\verb{<chr>} title alias}
}
\value{
An S7 \verb{<openTempGroup>} object.
}
\description{
Group of Open Payments Temporal Endpoints
}
\examples{
openTempGroup("grouped_payment")
openTempGroup("detailed_payment")
}
