% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_httr2.R
\name{parse_json_response_public}
\alias{parse_json_response_public}
\title{Parse JSON response}
\usage{
parse_json_response_public(resp)
}
\arguments{
\item{resp}{\verb{<httr_response>} object}
}
\value{
\verb{<tibble>} Parsed JSON response
}
\description{
Parse JSON response
}
\keyword{internal}
