% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2_perform_request.R
\name{perform_request_provider}
\alias{perform_request_provider}
\title{Perform Provider API request}
\usage{
perform_request_provider(url, query)
}
\arguments{
\item{url}{\verb{<url>} API identifier url}

\item{query}{\verb{<chr>} vector of query parameters}
}
\value{
\verb{<int>} Number of results
}
\description{
Perform Provider API request
}
\keyword{internal}
