% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2_perform_request.R
\name{perform_request_public}
\alias{perform_request_public}
\title{Perform Public API request}
\usage{
perform_request_public(url, query, limit = 5000L)
}
\arguments{
\item{url}{\verb{<url>} API identifier url}

\item{query}{\verb{<chr>} vector of query parameters}

\item{limit}{\verb{<int>} API rate limit; default is 5000}
}
\value{
\verb{<int>} Number of results
}
\description{
Perform Public API request
}
\keyword{internal}
