% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_httr2.R
\name{perform_simple}
\alias{perform_simple}
\title{Perform Request and Parse Simple JSON Response}
\usage{
perform_simple(req, ...)
}
\arguments{
\item{req}{\verb{<httr2_request>} object}

\item{...}{Additional arguments}
}
\value{
\verb{<list>} Parsed JSON response
}
\description{
Perform Request and Parse Simple JSON Response
}
\keyword{internal}
