% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_caid.R, R/S7_pro.R
\name{caidGroup}
\alias{caidGroup}
\alias{proMain}
\alias{proGroup}
\title{Medicaid Endpoint Group}
\usage{
caidGroup(alias)

proMain(alias)

proGroup(alias)
}
\arguments{
\item{alias}{\verb{<chr>} title alias}
}
\value{
An S7 \verb{<caidGroup>} object.

An S7 \verb{<proMain>} object.

An S7 \verb{<proGroup>} object.
}
\description{
Medicaid Endpoint Group

Provider Endpoint

Provider Endpoint Group
}
\examples{
caidGroup("demographics")
proMain("affiliations")
proMain("clinicians")
proMain("utilization")
proGroup("PDC")
proGroup("MIPS")
}
