% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_provider.R
\name{pro_dict}
\alias{pro_dict}
\title{Convert Provider UUID to Data Dictionary Hyperlink}
\usage{
pro_dict(uuid)
}
\arguments{
\item{uuid}{\verb{<chr>} endpoint UUID}
}
\value{
\verb{<chr>} dictionary URL
}
\description{
Convert Provider UUID to Data Dictionary Hyperlink
}
\keyword{internal}
