% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_backend.R
\name{pro_group}
\alias{pro_group}
\title{Provider API Group Endpoint}
\usage{
pro_group(alias)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint alias}
}
\value{
\verb{<data.frame>}
}
\description{
Provider API Group Endpoint
}
\examples{
pro_group("mips")
}
