% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{process_params}
\alias{process_params}
\title{Process Query Parameters}
\usage{
process_params(arg_names, field_names)
}
\arguments{
\item{arg_names}{\verb{<chr>} vector of argument names}

\item{field_names}{\verb{<chr>} vector of API field names}
}
\value{
\verb{<expr>} of processed query parameters
}
\description{
Process Query Parameters
}
\keyword{internal}
