% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_misc.R
\name{prop_empty}
\alias{prop_empty}
\title{Check if a property is empty}
\usage{
prop_empty(obj, nm)
}
\arguments{
\item{obj}{\verb{<S7_object>} to check}

\item{nm}{\verb{<chr>} property name}
}
\value{
\verb{<lgl>} \code{TRUE} if empty, \code{FALSE} otherwise
}
\description{
Check if a property is empty
}
\keyword{internal}
