% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pro_utilization.R
\name{proUtilization}
\alias{proUtilization}
\title{Utilization Class}
\usage{
proUtilization(
  npi = NULL,
  pac = NULL,
  last_name = NULL,
  first_name = NULL,
  middle_name = NULL,
  suffix = NULL,
  procedure = NULL,
  count = NULL,
  percentile = NULL
)
}
\arguments{
\item{npi}{\verb{<chr>} 10-digit Individual NPI}

\item{pac}{\verb{<chr>} 10-digit PECOS Associate Control (PAC) ID}

\item{first_name, middle_name, last_name, suffix}{\verb{<chr>} Individual provider's name}

\item{procedure}{\verb{<chr>} Procedure category}

\item{count}{\verb{<int>} Number of procedures performed}

\item{percentile}{\verb{<int>} Percentile of procedures performed}
}
\value{
An S7 \verb{<proUtilization>} object.
}
\description{
Utilization Class
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
proUtilization(last_name = "CURRY", middle_name = "")
\dontshow{\}) # examplesIf}
}
