% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_provider.R
\name{prov_nrows_fields}
\alias{prov_nrows_fields}
\title{Get Field Names and Number of Rows from Provider Endpoint}
\usage{
prov_nrows_fields(uuid)
}
\arguments{
\item{uuid}{\verb{<chr>} endpoint UUID}
}
\value{
\verb{<list>} number of rows and field names
}
\description{
Get Field Names and Number of Rows from Provider Endpoint
}
\keyword{internal}
