% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_provider.R
\name{prov_uuid_dict}
\alias{prov_uuid_dict}
\title{Convert Provider UUID to Data Dictionary Hyperlink}
\usage{
prov_uuid_dict(x)
}
\arguments{
\item{x}{\verb{<chr>} UUID}
}
\value{
\verb{<chr>} URL
}
\description{
Convert Provider UUID to Data Dictionary Hyperlink
}
\keyword{internal}
