% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_provider.R
\name{prov_uuid_url}
\alias{prov_uuid_url}
\title{Convert Provider UUID to URL}
\usage{
prov_uuid_url(uuid)
}
\arguments{
\item{uuid}{\verb{<chr>} endpoint UUID}
}
\value{
\verb{<chr>} endpoint URL
}
\description{
Convert Provider UUID to URL
}
\keyword{internal}
