% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_pro.R
\name{pro_endpoint}
\alias{pro_endpoint}
\alias{pro_group}
\title{Provider Endpoint}
\usage{
pro_endpoint(alias)

pro_group(alias)
}
\arguments{
\item{alias}{\verb{<chr>} title alias}
}
\value{
An S7 \verb{<pro_endpoint>} object.

An S7 \verb{<pro_group>} object.
}
\description{
Provider Endpoint

Provider Endpoint Group
}
\examples{
pro_endpoint("PDC_affiliations")
pro_endpoint("PDC_clinicians")
pro_endpoint("PDC_utilization")
pro_group("PDC")
pro_group("MIPS")
}
