% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{provider_current}
\alias{provider_current}
\title{\code{provider_current} Class}
\usage{
provider_current(
  title = character(0),
  description = character(0),
  contact = character(0),
  modified = character(0),
  identifier = character(0),
  download = character(0),
  issued = character(0),
  released = character(0),
  site = character(0)
)
}
\arguments{
\item{title}{\verb{<chr>} Endpoint title}

\item{description}{\verb{<chr>} Endpoint description}

\item{contact}{\verb{<chr>} Endpoint contact information}

\item{modified}{\verb{<Date>} Date Endpoint data was last modified}

\item{identifier}{\verb{<chr>} Endpoint url}

\item{download}{\verb{<chr>} Endpoint download url}

\item{issued}{\verb{<Date>} Date Endpoint data was issued}

\item{released}{\verb{<Date>} Date Endpoint data was released}

\item{site}{\verb{<chr>} Hyperlink to API landing page}
}
\value{
\code{provider_current} object
}
\description{
\code{provider_current} Class
}
\seealso{
Other classes: 
\code{\link{endpoint_current}()},
\code{\link{endpoint_temporal}()},
\code{\link{main_current}()},
\code{\link{main_resources}()}
}
\concept{classes}
