% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_provider.R
\name{provider_current_group}
\alias{provider_current_group}
\title{Load \verb{<CurrentProviderGroup>} API Endpoint}
\usage{
provider_current_group(alias)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint alias}
}
\value{
\verb{<CurrentProviderGroup>} object
}
\description{
Load \verb{<CurrentProviderGroup>} API Endpoint
}
\examples{
provider_current_group("mips")
}
