% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_provider.R
\name{provider_dataset}
\alias{provider_dataset}
\title{Load \code{CurrentProvider} API Endpoint}
\usage{
provider_dataset(alias)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint alias}
}
\value{
\verb{<CurrentProvider>} object
}
\description{
Load \code{CurrentProvider} API Endpoint
}
\examples{
provider_dataset("affiliations")
provider_dataset("clinicians")
provider_dataset("utilization")
provider_dataset("utilization")
provider_dataset("group_mips")
provider_dataset("group_patient")
provider_dataset("clin_mips")
provider_dataset("clin_overall")
provider_dataset("vgroup_mips")
}
