% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrollees.R
\name{provider_enrollment}
\alias{provider_enrollment}
\title{Public Provider Enrollment Experiment}
\usage{
provider_enrollment()
}
\value{
\verb{<tibble>} of search results
}
\description{
Public Provider Enrollment Experiment
}
\examples{
provider_enrollment()

}
