% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{provider_Identifier}
\alias{provider_Identifier}
\title{provider_Identifier Class}
\usage{
provider_Identifier(id = NULL)
}
\arguments{
\item{id}{\verb{<chr>} identifier}
}
\value{
\verb{<S7_class>} provider_Identifier object
}
\description{
\code{provider_Identifier} object
}
\examples{
provider_Identifier(id = "mj5m-pzi6")

}
