% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/providers.R
\name{providers2}
\alias{providers2}
\title{Provider Enrollment in Medicare}
\usage{
providers2(
  npi = NULL,
  pac = NULL,
  enid = NULL,
  spec_code = NULL,
  spec_desc = NULL,
  first = NULL,
  middle = NULL,
  last = NULL,
  org = NULL,
  state = NULL,
  gender = NULL
)
}
\arguments{
\item{npi}{\verb{<chr>} 10-digit Individual NPI}

\item{pac}{\verb{<chr>} 10-digit PECOS Associate Control (PAC) ID}

\item{enid}{\verb{<chr>} 15-digit Medicare Enrollment ID}

\item{spec_code}{\verb{<chr>} Enrollment specialty code}

\item{spec_desc}{\verb{<chr>} Enrollment specialty description}

\item{first, middle, last}{\verb{<chr>} Individual provider's name}

\item{org}{\verb{<chr>} Organizational provider's name}

\item{state}{\verb{<chr>} Enrollment state abbreviation}

\item{gender}{\verb{<chr>} Individual provider's gender:
\itemize{
\item \code{"F"} (Female)
\item \code{"M"} (Male)
\item \code{"9"} (Unknown/Organization)
}}
}
\value{
\verb{<tibble>} of search results
}
\description{
Provides access to individual and organizational enrollment level data on
providers actively approved to bill Medicare.
}
\section{accrualPeriodicity}{

\itemize{
\item \code{R/P1Y} (Annually)
}
}

\section{Links}{

\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-fee-for-service-public-provider-enrollment}{Provider Enrollment API}
\item \href{https://data.cms.gov/resources/medicare-fee-for-service-public-provider-enrollment-data-dictionary}{Provider Enrollment Data Dictionary}
}
}

\examples{
providers2(enid = "I20040309000221")

providers2(npi = "1417918293", spec_code = "14-41")

providers2(pac = "2860305554", gender = "9")

}
