% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/providertwo-package.R
\docType{package}
\name{providertwo-package}
\alias{providertwo}
\alias{providertwo-package}
\title{providertwo: Tidy Healthcare Provider API Interface}
\description{
Provides easy access to healthcare provider data through publicly available APIs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andrewallenbruce/providertwo}
  \item \url{https://andrewallenbruce.github.io/providertwo/}
  \item Report bugs at \url{https://github.com/andrewallenbruce/providertwo/issues}
}

}
\author{
\strong{Maintainer}: Andrew Allen Bruce \email{andrewallenbruce@gmail.com} [copyright holder]

}
\keyword{internal}
