% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{public_dataset}
\alias{public_dataset}
\title{Public CMS Dataset Object}
\usage{
public_dataset()
}
\value{
\verb{<list>} of \verb{<tibbles>}: \code{dataset}, \code{distribution}, \code{downloads}
}
\description{
Public CMS Dataset Object
}
\examples{
public_dataset()

}
