% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_class.R
\name{public_Distribution}
\alias{public_Distribution}
\title{Load Public API \code{Distribution}}
\usage{
public_Distribution(dataset)
}
\arguments{
\item{dataset}{\verb{<chr>} dataset title}
}
\value{
\verb{<Distribution>} object
}
\description{
Load Public API \code{Distribution}
}
\examples{
public_Distribution("Public Provider Enrollment")

public_Distribution("Quality Payment Program Experience")

}
