% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{public_filter}
\alias{public_filter}
\title{Filter Public Catalog}
\usage{
public_filter(endpoint = c("dataset", "latest", "api", "csv"), title = NULL)
}
\arguments{
\item{endpoint}{\verb{<chr>} API endpoint; options are \code{"dataset"} (default),
\code{"latest"}, \code{"api"}, \code{"csv"}}

\item{title}{\verb{<chr>} dataset title to search for; use \code{NULL} to return all}
}
\value{
\verb{<tibble>} of search results
}
\description{
Filter Public Catalog
}
\examples{
public_filter(
  endpoint = "latest",
  title = "Medicare Fee-For-Service Public Provider Enrollment : 2024-09-01")

public_filter(
  endpoint = "dataset",
  title = "Medicare Fee-For-Service  Public Provider Enrollment")

}
