% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{public_Identifier}
\alias{public_Identifier}
\title{public_Identifier Class}
\usage{
public_Identifier(url = NULL)
}
\arguments{
\item{url}{\verb{<chr>} identifier url}
}
\value{
\verb{<S7_class>} public_Identifier object
}
\description{
\code{public_Identifier} object
}
\examples{
public_Identifier(url =
   paste0(
   "https://data.cms.gov/data-api/v1/dataset/",
   "2457ea29-fc82-48b0-86ec-3b0755de7515/",
   "data-viewer"))

}
