% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_main.R
\name{public_temporal}
\alias{public_temporal}
\title{Load Public API \code{Distribution}}
\usage{
public_temporal(dataset, fname = TRUE)
}
\arguments{
\item{dataset}{\verb{<chr>} dataset title}

\item{fname}{\verb{<lgl>} Is \code{dataset} a function name?; default is \code{TRUE}}
}
\value{
\verb{<Distribution>} object
}
\description{
Load Public API \code{Distribution}
}
\examples{
public_temporal("utilization_provider")

public_temporal("quality_payment")

}
