% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{Publisher}
\alias{Publisher}
\title{Publisher Class}
\usage{
Publisher(
  type = "org:Organization",
  name = "Centers for Medicare & Medicaid Services"
)
}
\arguments{
\item{type}{\verb{<chr>} type}

\item{name}{\verb{<chr>} publisher name}
}
\value{
\verb{<S7_class>} object
}
\description{
\code{Publisher} object
}
