% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query-helpers}
\alias{query-helpers}
\alias{greater_than}
\alias{greater_equal}
\alias{less_than}
\alias{less_equal}
\alias{starts_with}
\alias{ends_with}
\alias{includes}
\alias{inside}
\alias{outside}
\alias{is_in}
\alias{not_in}
\title{Query Formatting Helpers}
\usage{
greater_than(x)

greater_equal(x)

less_than(x)

less_equal(x)

starts_with(x)

ends_with(x)

includes(x)

inside(x)

outside(x)

is_in(x)

not_in(x)
}
\arguments{
\item{x}{\verb{<chr>} function argument input}
}
\value{
\verb{<list>} of function argument and operator
}
\description{
Query Formatting Helpers
}
\examples{
greater_than(1)

starts_with("foo")

is_in(state.abb[10:15])
}
