% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_modifiers.R
\name{query-modifiers}
\alias{query-modifiers}
\alias{modifier_}
\alias{greater_than_}
\alias{less_than_}
\alias{starts_with_}
\alias{ends_with_}
\alias{contains_}
\alias{in_}
\alias{equals_}
\alias{is_blank_}
\title{Query Modifiers}
\usage{
modifier_(operator, value)

greater_than_(x, or_equals = FALSE)

less_than_(x, or_equals = FALSE)

starts_with_(x)

ends_with_(x)

contains_(x)

in_(x, negate = FALSE)

equals_(x, negate = FALSE)

is_blank_(negate = FALSE)
}
\arguments{
\item{x}{\verb{<chr>}  input}

\item{or_equals}{\verb{<lgl>}  append "="; default is \code{FALSE}}

\item{negate}{\verb{<lgl>}  prepend "NOT"; default is \code{FALSE}}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    `<list>` of class `"modifier"`
}\if{html}{\out{</div>}}
}
\description{
Query Modifiers
}
\examples{
starts_with_("foo")
ends_with_("bar")
contains_("baz")

in_(state.abb[10:15])
in_(state.abb[10:15], negate = TRUE)

equals_(1000)
equals_(1000, negate = TRUE)

greater_than_(1000)
greater_than_(1000, or_equals = TRUE)

less_than_(1000)
less_than_(1000, or_equals = TRUE)

is_blank_()
is_blank_(negate = TRUE)
}
\keyword{internal}
