% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_count.R
\name{query_count}
\alias{query_count}
\title{Return the number of results for a given query}
\usage{
query_count(obj, ...)
}
\arguments{
\item{obj}{An object of class \verb{<class_care>}, \verb{<class_group>}, or \verb{<class_collection>}.}

\item{...}{Additional arguments passed to the query.}
}
\value{
An \verb{<int>} representing the number of results found.
}
\description{
Return the number of results for a given query
}
\examples{
endpoint("care_dial_end") |>
query_count(query(
state = in_(c("CA", "GA", "NY"),
care = TRUE),
.type = "medicare"))

collection("care_rhc") |>
query_count(
query(STATE = in_(c("CA", "GA", "NY"), care = TRUE),
      `STATE - OWNER` = in_(c("CA", "GA", "NY"), care = TRUE),
      .type = "medicare"))

}
