% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_groupers.R
\name{query_grouper}
\alias{query_grouper}
\alias{and}
\alias{or}
\title{Query Groups}
\source{
\href{https://jsonapi.org/format/#query-parameters}{JSON-API: Query Parameters}
}
\usage{
and(...)

or(...)
}
\arguments{
\item{...}{input}
}
\value{
An S7 \verb{<class_junction>} object.
}
\description{
Helpers for use in constructing groups in queries.
}
\details{
Query groupers are part of a small DSL for use in query construction, in the \href{https://www.drupal.org/docs/core-modules-and-themes/core-modules/jsonapi-module/filtering}{JSON-API} format.
}
\examples{
and("foo", "bar")
or("foo", "bar")
}
