% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_modifiers.R
\name{query_modifier}
\alias{query_modifier}
\alias{equals}
\alias{greater_than}
\alias{less_than}
\alias{between}
\title{Query Modifiers}
\usage{
equals(x, negate = FALSE)

greater_than(x, or_equal = FALSE)

less_than(x, or_equal = FALSE)

between(x, y, negate = FALSE)
}
\arguments{
\item{x, y}{input}

\item{negate}{\verb{<lgl>} prepend \code{NOT}}

\item{or_equal}{\verb{<lgl>} append \code{=}}
}
\value{
An object of class \verb{<modifier>}
}
\description{
Helpers for use in constructing conditions in queries.
}
\details{
Query modifiers are a small DSL for use in constructing query conditions,
in the \href{https://www.drupal.org/docs/core-modules-and-themes/core-modules/jsonapi-module/filtering}{JSON-API} format.
}
\examples{
equals(1000)
equals(1000, negate = TRUE)
greater_than(1000)
greater_than(0.125, or_equal = TRUE)
less_than(1000)
less_than(0.125, or_equal = TRUE)
between(1000, 1100)
between(0.125, 2, negate = TRUE)
}
