% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_base_request.R
\name{query_nresults}
\alias{query_nresults}
\title{Request the number of results for a query}
\usage{
query_nresults(x, ...)
}
\arguments{
\item{x}{An \verb{<int>} vector}

\item{...}{Additional arguments}
}
\value{
An \verb{<int>} vector
}
\description{
Request the number of results for a query
}
