% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2_perform_request.R
\name{query_nrows_public}
\alias{query_nrows_public}
\title{Request number of results from Public Query}
\usage{
query_nrows_public(req)
}
\arguments{
\item{req}{\verb{<httr_request>} API request}
}
\value{
\verb{<int>} Number of results
}
\description{
Request number of results from Public Query
}
\keyword{internal}
