% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_results.R
\name{query_results}
\alias{query_results}
\title{Return the number of results for a given query}
\usage{
query_results(obj, query = NULL, ...)
}
\arguments{
\item{obj}{An object of class \verb{<class_endpoint>} or \verb{<class_temporal>}.}

\item{query}{A \verb{<class_query>} object. If \code{NULL} (the default), no query is used.}

\item{...}{Additional arguments.}
}
\value{
A \verb{<list>} with the number of results found, total number of rows, and the URL used for the query.
}
\description{
Return the number of results for a given query
}
\examples{
query_results(endpoint("care_dial_end"), query(state = any_of(c("GA", "TX"))))

query_results(
   endpoint("quality_payment"),
   query(year = 2017:2023,
         `practice state or us territory` = any_of(c("GA", "TX"))))

}
