% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick.R
\name{quick}
\alias{quick}
\title{Quickly access CMS data}
\usage{
quick(x, offset = 0L, limit = 10000L, call = caller_env())
}
\arguments{
\item{x}{A string representing the CMS data endpoint or category.}

\item{offset}{An integer representing the offset for pagination. Default is 0.}

\item{limit}{An integer representing the maximum number of records to retrieve. Default is 10,000.}

\item{call}{The calling environment. Default is the current environment.}
}
\value{
A data frame or object containing the requested CMS data.
}
\description{
\code{quick()} is a convenience function to quickly access various CMS data endpoints.
It provides a simple way to retrieve data from different CMS programs and categories.
}
\examples{
quick("MLR", limit = 1) |> str()
quick("enterprise", limit = 1) |> str()
quick("demographics", limit = 1) |> str()

quick("suppliers", limit = 1) |> str()
quick("PDC", limit = 1) |> str()
quick("MIPS", limit = 1) |> str()
quick("LTCH", limit = 1) |> str()
quick("IRF", limit = 1) |> str()
quick("SPICE", limit = 1) |> str()
quick("CAHPS_hospice", limit = 1) |> str()
quick("HHVBP", limit = 1) |> str()
quick("HHCAHPS", limit = 1) |> str()
quick("HHCAHPS", limit = 1) |> str()
quick("SNF_VBP", limit = 1) |> str()
quick("SNF_quality", limit = 1) |> str()
quick("NH_pro", limit = 1) |> str()

quick("summary", limit = 1) |> str()
quick("profile", limit = 1) |> str()

quick("HHA", limit = 1) |> str()
quick("hospice", limit = 1) |> str()
quick("hospital", limit = 1) |> str()
quick("RHC", limit = 1) |> str()
quick("FQHC", limit = 1) |> str()
quick("pending", limit = 1) |> str()
quick("reassignment", limit = 1) |> str()
quick("SNF", limit = 1) |> str()

quick("contact", limit = 1) |> str()
quick("crosswalk", limit = 1) |> str()
quick("CARE_dialysis", limit = 1) |> str()
quick("enrollees", limit = 1) |> str()
quick("facilities", limit = 1) |> str()
quick("IQIES", limit = 1) |> str()
quick("laboratories", limit = 1) |> str()
quick("long_term", limit = 1) |> str()
quick("opt_out", limit = 1) |> str()
quick("order_refer", limit = 1) |> str()
quick("RBCS", limit = 1) |> str()
quick("transparency", limit = 1) |> str()
}
