% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_quick_query.R
\name{quick_query_}
\alias{quick_query_}
\title{Quick Query}
\usage{
quick_query_(x, ...)
}
\arguments{
\item{x}{An S7 \verb{<endpoint>} object}

\item{...}{Additional arguments}
}
\value{
A list of results from the API query, or \code{NULL} if no results are found.
}
\description{
Quick Query
}
\examples{
\dontshow{if (rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
asc <- new_endpoint("asc_facility")

q0 <- list(
`conditions[0][property]` = "state",
`conditions[0][operator]` = "=",
`conditions[0][value]` = "ZZ")

q1 <- list(
`conditions[0][property]` = "state",
`conditions[0][operator]` = "=",
`conditions[0][value]` = "NY")

q2 <- list(
  `conditions[0][property]` = "state",
  `conditions[0][operator]` = "IN",
  `conditions[0][value][1]` = "CA",
  `conditions[0][value][2]` = "GA",
  `conditions[0][value][3]` = "NY")

q3 <- list(
  `conditions[0][property]` = "asc2_rate",
  `conditions[0][operator]` = "<",
  `conditions[0][value]` = "0.02")

quick_query_(asc, q0)
quick_query_(asc, q1)
#quick_query_(asc, q2)
#quick_query_(asc, q3)
\dontshow{\}) # examplesIf}
}
