% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{recode_iso8601}
\alias{recode_iso8601}
\title{Parse ISO8601 Recurrence Rules}
\usage{
recode_iso8601(x)
}
\arguments{
\item{x}{\verb{<chr>} vector of ISO8601 recurrence rules}
}
\value{
\verb{<chr>} vector of human-readable recurrence rule descriptions
}
\description{
Parse ISO8601 Recurrence Rules
}
\examples{
data.frame(
  accrualPeriodicity = c(
  "R/PT1S",   "R/PT1H",  "R/P1D", "R/P3.5D",
  "R/P0.33W", "R/P0.5W", "R/P1W", "R/P2W",
  "R/P0.33M", "R/P0.5M", "R/P1M", "R/P2M",
  "R/P3M",    "R/P4M",   "R/P6M", "R/P1Y",
  "R/P2Y",    "R/P3Y",   "R/P4Y", "R/P10Y"),
  accrualDescription = recode_iso8601(accrualPeriodicity)
)

}
