% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_utils.R
\name{reduce_contact}
\alias{reduce_contact}
\title{Concatenate Contact Point}
\usage{
reduce_contact(x)
}
\arguments{
\item{x}{\verb{<list>} list to get element from}
}
\value{
\verb{<chr>} element
}
\description{
Concatenate Contact Point
}
\keyword{internal}
