% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2_utils.R
\name{req_perform_iterative_offset}
\alias{req_perform_iterative_offset}
\title{Perform Iterative API request with Offset}
\usage{
req_perform_iterative_offset(req, limit)
}
\arguments{
\item{req}{\verb{<httr2_request>} object}

\item{limit}{\verb{<int>} API rate limit, i.e. the maximum number of results an
API will return per request.}
}
\value{
list of \verb{<httr2_response>}s
}
\description{
Perform Iterative API request with Offset
}
