% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{Resources}
\alias{Resources}
\title{Resources Class}
\usage{
Resources(url = NA_character_)
}
\arguments{
\item{url}{\verb{<chr>} \code{resourcesAPI} url; default is \code{NA}}
}
\value{
\verb{<S7_class>} Resources object
}
\description{
Resources Class
}
\seealso{
Other classes: 
\code{\link{Contact}()},
\code{\link{Dataset}()},
\code{\link{Identifier}()},
\code{\link{Publisher}()}
}
\concept{classes}
