% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_backend.R
\name{ResourcesMain}
\alias{ResourcesMain}
\title{Main API Endpoint Resources}
\usage{
ResourcesMain(url = character(0))
}
\arguments{
\item{url}{\verb{<chr>} resources URL}
}
\value{
An S7 \verb{<ResourcesMain>} object.
}
\description{
Main API Endpoint Resources
}
\keyword{internal}
