% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sbt_detect}
\alias{sbt_detect}
\title{Subset by Regular Expression}
\usage{
sbt_detect(i, x, p, n = FALSE)
}
\arguments{
\item{i}{\verb{<data.frame>} to search}

\item{x}{\verb{<chr>} vector to detect}

\item{p}{\verb{<chr>} regular expression pattern}

\item{n}{\verb{<lgl>} negate}
}
\value{
\verb{<data.frame>} subsetted data.frame
}
\description{
Subset by Regular Expression
}
\keyword{internal}
