% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_misc.R
\name{set_clean}
\alias{set_clean}
\title{Set clean column names}
\usage{
set_clean(i, x)
}
\arguments{
\item{i}{\verb{<chr>} vector to be named}

\item{x}{\verb{<chr>} vector of column names}
}
\value{
\verb{<chr>} vector with clean column names
}
\description{
Set clean column names
}
\keyword{internal}
