% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_misc.R
\name{smush_elem}
\alias{smush_elem}
\title{Get List Element}
\usage{
smush_elem(i, el)
}
\arguments{
\item{i}{\verb{<list>} list to get element from}

\item{el}{\verb{<chr>} element to get}
}
\value{
\verb{<list>} list with element
}
\description{
Get List Element
}
\keyword{internal}
