% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_standardize.R
\name{standardise}
\alias{standardise}
\title{Standardize a Query to an Endpoint}
\usage{
standardise(obj, qry)
}
\arguments{
\item{obj}{An \verb{<endpoint>}, \verb{<collection>} or \verb{<group>} object.}

\item{qry}{A \verb{<query>} object.}
}
\value{
A named list of query parameters that match the fields of the endpoint.
}
\description{
Standardize a Query to an Endpoint
}
\examples{
qry <- new_query(
   first_name            = starts_with("Andr"),
   middle_name           = ends_with("e"),
   last_name             = contains("J"),
   state_cd              = any_of(c("CA", "GA", "NY")),
   state                 = none_of(c("GA", "NY")),
   city                  = all_but(c("Atlanta", "Los Angeles")),
   npi                   = npi_ex$k,
   covered_recipient_npi = npi_ex$k[1:5],
   ccn                   = "01256",
   rate                  = between(0.45, 0.67),
   year                  = 2021:2025)

standardise(endpoint("enroll_prov"), qry)

standardise(endpoint("dial_facility"), qry)

standardise(endpoint("quality_payment"), qry)

standardise(endpoint("pdc_clinician"), qry)

standardise(endpoint("grp_cover_nature"), qry)

standardise(collection("hha"), qry)

standardise(collection("hhc"), qry)

standardise(collection("hhc"))

}
