% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_standardize.R
\name{standardize}
\alias{standardize}
\title{Standardize a Query for an Endpoint}
\usage{
standardize(obj, qry)
}
\arguments{
\item{obj}{An \verb{<endpoint>}, \verb{<collection>} or \verb{<group>} object.}

\item{qry}{A \verb{<query>} object.}
}
\value{
A list of query parameters matched to an endpoint's fields.
}
\description{
Standardize a Query for an Endpoint
}
\examples{
standardize(
  endpoint("drug_state"),
  query(year = 2022:2024,
  state = any_of("GA", "NY")))

standardize(
  endpoint("enroll_prov"),
  query(enrlmt_id = "I20040309000221"))

standardize(
  collection("dialysis"),
  query(state = "AL"))

standardize(
  endpoint("dial_facility"),
  query(
    state = "GA",
    city = "Atlanta",
    provcity = "Atlanta",
    provider_name = starts_with("C"),
    provname = starts_with("C")
  )
)

standardize(
  endpoint("qppe"),
  query(
    year = 2021:2023,
    practice_state_or_us_territory = any_of("GA", "FL"),
    practice_size = less_than(10, or_equal = TRUE)
  )
)

standardize(
  endpoint("hc_quality"),
  query(state = any_of("Georgia", "Alabama"), year = 2020:2023))

standardize(
  endpoint("pdc_clinician"),
  query(
    provider_first_name = starts_with("An"),
    provider_last_name = contains("JE"),
    state = any_of("CA", "GA", "NY")))

}
