% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_modifiers.R
\name{starts_with_}
\alias{starts_with_}
\alias{ends_with_}
\alias{contains_}
\title{Character Modifiers}
\usage{
starts_with_(x)

ends_with_(x)

contains_(x)
}
\arguments{
\item{x}{\verb{<chr>} input}
}
\value{
\verb{<list>} of class \code{"modifier"}
}
\description{
Character Modifiers
}
\examples{
starts_with_("foo")
ends_with_("bar")
contains_("baz")
}
\seealso{
Other modifiers: 
\code{\link{greater_than_}()},
\code{\link{in_}()}

Other modifiers: 
\code{\link{greater_than_}()},
\code{\link{in_}()}

Other modifiers: 
\code{\link{greater_than_}()},
\code{\link{in_}()}
}
\concept{modifiers}
