% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_misc.R
\name{subset_detect}
\alias{subset_detect}
\title{Subset by Regular Expression}
\usage{
subset_detect(i, j, p, n = FALSE)
}
\arguments{
\item{i}{\verb{<data.frame>} to search}

\item{j}{\verb{<chr>} vector to detect}

\item{p}{\verb{<chr>} regular expression pattern}

\item{n}{\verb{<lgl>} negate}
}
\value{
\verb{<data.frame>} subsetted data.frame
}
\description{
Subset by Regular Expression
}
\keyword{internal}
