% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_backend.R
\name{Temporal}
\alias{Temporal}
\alias{TemporalMain}
\alias{TemporalMainGroup}
\alias{TemporalOpen}
\alias{TemporalOpenGroup}
\title{Temporal}
\usage{
Temporal(
  title = character(0),
  description = character(0),
  contact = character(0),
  rows = integer(0),
  fields = character(0),
  pages = integer(0),
  years = integer(0),
  endpoints = list()
)

TemporalMain(
  title = character(0),
  description = character(0),
  contact = character(0),
  rows = integer(0),
  fields = character(0),
  pages = integer(0),
  years = integer(0),
  endpoints = list()
)

TemporalMainGroup(
  title = character(0),
  description = character(0),
  contact = character(0),
  rows = integer(0),
  fields = character(0),
  pages = integer(0),
  years = integer(0),
  endpoints = list()
)

TemporalOpen(
  title = character(0),
  description = character(0),
  contact = character(0),
  rows = integer(0),
  fields = character(0),
  pages = integer(0),
  years = integer(0),
  endpoints = list()
)

TemporalOpenGroup(
  title = character(0),
  description = character(0),
  contact = character(0),
  rows = integer(0),
  fields = character(0),
  pages = integer(0),
  years = integer(0),
  endpoints = list()
)
}
\arguments{
\item{title}{\verb{<chr>} dataset title}

\item{description}{\verb{<chr>} dataset description}

\item{contact}{\verb{<chr>} dataset contact}

\item{rows}{\verb{<int>} number of rows}

\item{fields}{\verb{<chr>} field names}

\item{pages}{\verb{<int>} number of pages}

\item{years}{\verb{<int>} years available}

\item{endpoints}{\verb{<data.frame>} endpoints}
}
\description{
Temporal

TemporalMain

TemporalMainGroup

TemporalOpen

TemporalOpenGroup
}
\concept{base-classes}
\concept{main-classes}
\concept{open-classes}
