% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_backend.R
\name{TemporalMain}
\alias{TemporalMain}
\title{Main API Endpoint (Temporal)}
\usage{
TemporalMain(alias)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint alias}
}
\value{
An S7 \verb{<TemporalMain>} object.
}
\description{
Main API Endpoint (Temporal)
}
\examples{
TemporalMain("quality_payment")

}
