% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_backend.R
\name{TemporalMainGroup}
\alias{TemporalMainGroup}
\title{Main API Endpoint Temporal Group}
\usage{
TemporalMainGroup(alias)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint alias}
}
\value{
An S7 \verb{<TemporalMainGroup>} object.
}
\description{
Main API Endpoint Temporal Group
}
\examples{
TemporalMainGroup("utilization")
TemporalMainGroup("prescribers")
TemporalMainGroup("suppliers")
TemporalMainGroup("outpatient")
TemporalMainGroup("inpatient")

}
