% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_backend.R
\name{TemporalOpenGroup}
\alias{TemporalOpenGroup}
\title{Open Payments API Endpoint Temporal Group}
\usage{
TemporalOpenGroup(
  title = character(0),
  description = character(0),
  contact = character(0),
  rows = integer(0),
  pages = integer(0),
  fields = character(0),
  years = integer(0),
  endpoints = list()
)
}
\arguments{
\item{title}{\verb{<chr>} title}

\item{description}{\verb{<chr>} description}

\item{contact}{\verb{<chr>} contact name and email address}

\item{rows}{\verb{<int>} number of rows}

\item{pages}{\verb{<int>} number of pages}

\item{fields}{\verb{<chr>} field names}

\item{years}{\verb{<int>} years available}

\item{endpoints}{\verb{<data.frame>} endpoints}
}
\value{
An S7 \verb{<TemporalOpenGroup>} object.
}
\description{
Open Payments API Endpoint Temporal Group
}
