% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_httr2.R
\name{tidyup}
\alias{tidyup}
\title{Clean JSON response}
\usage{
tidyup(x, names)
}
\arguments{
\item{x}{\verb{<data.frame>} tibble of API response}

\item{names}{\verb{<chr>} vector of column names}
}
\value{
\verb{<tibble>} Parsed JSON response as tibble
}
\description{
Clean JSON response
}
\keyword{internal}
