% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_affiliations.R
\name{utilization}
\alias{utilization}
\title{Clinicians Utilization}
\usage{
utilization(
  npi = NULL,
  pac = NULL,
  last_name = NULL,
  first_name = NULL,
  middle_name = NULL,
  suffix = NULL,
  procedure = NULL,
  count = NULL,
  percentile = NULL
)
}
\arguments{
\item{npi}{\verb{<chr>} 10-digit Individual NPI}

\item{pac}{\verb{<chr>} 10-digit PECOS Associate Control (PAC) ID}

\item{first_name, middle_name, last_name, suffix}{\verb{<chr>} Individual provider's name}

\item{procedure}{\verb{<chr>} Procedure category}

\item{count}{\verb{<int>} Number of procedures performed}

\item{percentile}{\verb{<int>} Percentile of procedures performed}
}
\value{
\verb{<tibble>} of search results
}
\description{
The Doctors and Clinicians utilization data file reports volume
information for procedures of interest on clinician profile pages and
in the provider data catalog (PDC) to inform patients and caregivers
about clinicians' experience.
}
\examples{
utilization(last_name = "CURRY")
utilization(npi = "1043245657")
utilization(npi = "1003001785")
}
