% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pro_affiliations.R
\name{affiliations}
\alias{affiliations}
\title{Provider Facility Affiliations}
\usage{
affiliations(
  npi = NULL,
  pac = NULL,
  last_name = NULL,
  first_name = NULL,
  middle_name = NULL,
  suffix = NULL,
  facility_type = NULL,
  ccn_facility = NULL,
  ccn_primary = NULL
)
}
\arguments{
\item{npi}{\verb{<chr>} 10-digit Individual NPI}

\item{pac}{\verb{<chr>} 10-digit PECOS Associate Control (PAC) ID}

\item{first_name, middle_name, last_name, suffix}{\verb{<chr>} Individual provider's name}

\item{facility_type}{\verb{<chr>} Type of facility, one of the following:
\itemize{
\item \code{"Hospital"}
\item \code{"Long-term care hospital"}
\item \code{"Nursing home"}
\item \code{"Inpatient rehabilitation facility"}
\item \code{"Home health agency"}
\item \code{"Skilled nursing facility"}
\item \code{"Hospice"}
\item \code{"Dialysis facility"}
}}

\item{ccn_facility}{\verb{<chr>} 6-digit CMS Certification Number (CCN)
of a facility or unit within hospital where an individual provider
provides service.}

\item{ccn_primary}{\verb{<int>} 6-digit CMS Certification Number (CCN)
of a sub-unit's primary hospital, should the provider provide
services in said unit.}
}
\value{
\verb{<tibble>} of search results
}
\description{
Facility affiliations data publicly reported in the Provider Data Catalog.
}
\section{Periodicity}{

\itemize{
\item Monthly (R/P1M)
}
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
affiliations(facility_type = "Home health agency")
affiliations(last_name = "CURRY")
affiliations(ccn_primary = "670055")
affiliations(ccn_facility = "370781")
affiliations(ccn_facility = "331302")
affiliations(ccn_facility = "33Z302")
affiliations(npi = "1043245657")
\dontshow{\}) # examplesIf}
}
