% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_affiliations.R
\name{affiliations}
\alias{affiliations}
\title{Provider Facility Affiliations}
\usage{
affiliations(
  npi = NULL,
  pac = NULL,
  first = NULL,
  middle = NULL,
  last = NULL,
  suffix = NULL,
  facility_type = NULL,
  ccn_facility = NULL,
  ccn_parent = NULL
)
}
\arguments{
\item{npi}{\verb{<chr>} 10-digit Individual NPI}

\item{pac}{\verb{<chr>} 10-digit PECOS Associate Control (PAC) ID}

\item{first, middle, last, suffix}{\verb{<chr>} Individual provider's name(s)}

\item{facility_type}{\verb{<chr>} Type of facility, one of the following:
\itemize{
\item \code{Hospital} (\code{hp})
\item \verb{Long-term care hospital} (\code{ltch})
\item \verb{Nursing home} (\code{nh})
\item \verb{Inpatient rehabilitation facility} (\code{irf})
\item \verb{Home health agency} (\code{hha})
\item \verb{Skilled nursing facility} (\code{snf})
\item \code{Hospice} (\code{hs})
\item \verb{Dialysis facility} (\code{df})
}}

\item{ccn_facility}{\verb{<chr>} 6-digit CMS Certification Number (CCN) of
facility or unit within hospital where an
individual provider provides service.}

\item{ccn_parent}{\verb{<int>} 6-digit CMS Certification Number (CCN) of a
sub-unit's primary hospital, should the
provider provide services in said unit.}
}
\description{
Facility affiliations data publicly reported in the Provider Data Catalog.
}
\section{Accrual Periodicity}{

\itemize{
\item Monthly (R/P1M)
}
}

\examples{
\dontshow{if (rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
affiliations(last = "CURRY", facility_type = "Home health agency")
affiliations(ccn_parent = "670055")
affiliations(ccn_facility = "370781")
affiliations(ccn_facility = "331302")
affiliations(ccn_facility = "33Z302")
affiliations(npi = "1043245657")
affiliations(last = "CURRY")
\dontshow{\}) # examplesIf}
}
