% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_props.R
\name{arg_npi}
\alias{arg_npi}
\title{\code{npi} argument class}
\usage{
arg_npi(input = NULL)
}
\arguments{
\item{input}{\verb{<chr>} npi}
}
\value{
An S7 \verb{<arg_npi>} object.
}
\description{
\code{npi} argument class
}
\examples{
arg_npi("1225701881")

kind <- c(
1225701881,
1174270805,
1235702796,
1962116806,
1013647569,
1306500665,
1982296737,
1083295638,
1841967825,
1891390084,
1275117269,
1992338701,
1891355863,
1548743511,
1023473279,
1861857013,
1689182859,
1982059275)

arg_npi(kind)

}
