% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_base_request.R
\name{base_request}
\alias{base_request}
\title{Create a new \verb{<httr2_request>} by class}
\usage{
base_request(x, ...)
}
\arguments{
\item{x}{A \code{class_endpoint}, \code{class_temporal} or \code{class_group} object}

\item{...}{Additional arguments}
}
\value{
An \verb{<httr2_request>} object or list of \verb{<httr2_request>} objects
}
\description{
Create a new \verb{<httr2_request>} by class
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
endpoint("quality_payment")
collection("care_hospital")
endpoint("pdc_affiliations")
endpoint("mlr_summary")
endpoint("hgov_catastrophic")
endpoint("hgov_mlr")
\dontshow{\}) # examplesIf}
}
