% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_base_request.R
\name{base_request}
\alias{base_request}
\title{Create a new request by class}
\usage{
base_request(x)
}
\arguments{
\item{x}{An object of class \code{care_endpoint}, \code{care_group}, \code{pro_endpoint}, or \code{open_endpoint}}
}
\value{
A new base request
}
\description{
Create a new request by class
}
\examples{
care_endpoint("care_enrollees") |> base_request()
pro_endpoint("pdc_affiliations") |> base_request()
open_endpoint("profile_covered") |> base_request()
caid_endpoint("mlr_summary") |> base_request()

care_group("care_hospital") |> base_request()
pro_group("pro_mips") |> base_request()

# care_temporal("quality_payment") |> base_request()
# caid_temporal("healthcare_quality") |> base_request()
# open_temporal("payment_general") |> base_request()

# care_troup("care_utilization") |> base_request()
# open_troup("payment_grouped") |> base_request()
}
