% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_build.R
\name{build}
\alias{build}
\title{Build a Query for an Endpoint}
\usage{
build(obj, qry)
}
\arguments{
\item{obj}{An \verb{<endpoint>}, \verb{<collection>} or \verb{<group>} object.}

\item{qry}{A \verb{<query>} object.}
}
\value{
A list of query parameters matched to an endpoint's fields.
}
\description{
Build a Query for an Endpoint
}
\examples{
qry <- new_query(
   first_name = starts_with("An"),
   middle_name = ends_with("e"),
   last_name = contains("J"),
   state = any_of(c("CA", "GA", "NY")),
   practice_state_or_us_territory = any_of(c("CA", "GA", "NY")),
   year = 2021:2025)

build(endpoint("enroll_prov"), qry)

build(endpoint("dial_facility"), qry)

build(endpoint("quality_payment"), qry)

build(endpoint("pdc_clinician"), qry)

build(endpoint("grp_cover_nature"), qry)

build(collection("hha"), qry)

build(collection("hhc"), qry)

}
