% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_build.R
\name{build}
\alias{build}
\title{Build a Query for an Endpoint}
\usage{
build(obj, qry)
}
\arguments{
\item{obj}{An \verb{<endpoint>}, \verb{<collection>} or \verb{<group>} object.}

\item{qry}{A \verb{<query>} object.}
}
\value{
A list of query parameters matched to an endpoint's fields.
}
\description{
Build a Query for an Endpoint
}
\examples{
build(
  endpoint("drug_state"),
  query(
    year = 2022:2024,
    state = any_of(c("GA", "NY"))))

build(endpoint("enroll_prov"),
      query(first_name = starts_with("And"),
      last_name = ends_with("ce")))

build(endpoint("dial_facility"), query(city = "BIRMINGHAM"))

build(collection("dialysis"), query(state = "AL"))

build(
  endpoint("qppe"),
  query(
    year = 2021:2023,
    practice_state_or_us_territory = any_of(c("GA", "FL")),
    allowed_charges = less_than(10000)))

build(
  endpoint("pdc_clinician"),
  query(
    provider_first_name = starts_with("An"),
    provider_last_name = contains("J"),
    state = any_of(c("CA", "GA", "NY"))))

}
