% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_caid.R
\name{caidMain}
\alias{caidMain}
\title{Provider Endpoint}
\usage{
caidMain(alias)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint alias}
}
\value{
An S7 \verb{<caidMain>} object.
}
\description{
Provider Endpoint
}
\examples{
caidMain("mlr")
caidMain("mesd")
caidMain("wcv")
caidMain("mhsud")
caidMain("disability")
caidMain("livebirth")
caidMain("lang")
caidMain("race")
caidMain("rural")
caidMain("waive")
caidMain("newdrug_01")
caidMain("newdrug_16")
}
