% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_caid.R
\name{caidMain}
\alias{caidMain}
\title{Medicaid Endpoint}
\usage{
caidMain(alias)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint alias}
}
\value{
An S7 \verb{<caidMain>} object.
}
\description{
Medicaid Endpoint
}
\examples{
caidMain("MLR")
caidMain("enterprise")
}
