% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_caid.R
\name{caid_endpoint}
\alias{caid_endpoint}
\alias{caid_group}
\title{Medicaid Endpoint}
\usage{
caid_endpoint(alias)

caid_group(alias)
}
\arguments{
\item{alias}{\verb{<chr>} title alias}
}
\value{
An S7 \verb{<caid_endpoint>} object.

An S7 \verb{<caid_group>} object.
}
\description{
Medicaid Endpoint

Medicaid Endpoint Group
}
\examples{
caid_endpoint("MLR")
caid_endpoint("enterprise")
caid_group("demographics")
}
