% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_care.R
\name{careMain}
\alias{careMain}
\title{Medicare Endpoint}
\usage{
careMain(alias)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint alias}
}
\value{
An S7 \verb{<careMain>} object.
}
\description{
Medicare Endpoint
}
\examples{
careMain("enrollees")
careMain("opt_out")
careMain("order_refer")
careMain("reassignments")
# careMain("hospitals")
careMain("laboratories")
careMain("crosswalk")
careMain("rbcs")
careMain("facilities")
careMain("home_health")
careMain("hospice")
careMain("dialysis")
careMain("skilled_nursing")
}
