% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_care.R
\name{careTemp}
\alias{careTemp}
\title{Medicare Temporal Endpoint}
\usage{
careTemp(alias)
}
\arguments{
\item{alias}{\verb{<chr>} title alias}
}
\value{
An S7 \verb{<careTemp>} object.
}
\description{
Medicare Temporal Endpoint
}
\examples{
careTemp("quality_payment")
}
