% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_main.R
\name{catalog_main}
\alias{catalog_main}
\title{CMS Main Catalog}
\usage{
catalog_main()
}
\value{
\verb{<list>} of CMS Main API catalog information
}
\description{
CMS Main Catalog
}
\examples{
catalog_main()

}
