% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_provider.R
\name{catalog_provider}
\alias{catalog_provider}
\title{CMS Provider Catalog}
\usage{
catalog_provider()
}
\value{
\verb{<list>} of Provider API catalog information
}
\description{
CMS Provider Catalog
}
\examples{
\dontshow{if (rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
catalog_provider()
\dontshow{\}) # examplesIf}
}
\keyword{internal}
