% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{Catalog_provider}
\alias{Catalog_provider}
\title{CMS Provider Catalog}
\usage{
Catalog_provider()
}
\value{
\verb{<tibble>} of Provider API catalog information
}
\description{
CMS Provider Catalog
}
\examples{
Catalog_provider()

}
