% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-catalogs.R
\name{catalogs}
\alias{catalogs}
\title{API Catalogs}
\usage{
catalogs()
}
\value{
\verb{<list>} of catalogs
}
\description{
List of API catalogs:
\itemize{
\item \code{catalog_care}: CMS Medicare API
\item \code{catalog_prov}: CMS Provider API
\item \code{catalog_open}: CMS Open Payments API
\item \code{catalog_caid}: CMS Medicaid API
\item \code{catalog_hgov}: CMS HealthCare.gov API
}
}
\examples{
catalogs()
}
