% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{catalogs}
\alias{catalogs}
\title{API Catalogs}
\usage{
catalogs(backup = TRUE)
}
\arguments{
\item{backup}{\verb{<lgl>} default \code{TRUE}; use local backup of catalog data}
}
\value{
\verb{<list>} of catalogs
}
\description{
List of API catalogs:
\itemize{
\item \code{care}: CMS Medicare API
\item \code{prov}: CMS Provider API
\item \code{open}: CMS Open Payments API
\item \code{caid}: CMS Medicaid API
\item \code{hgov}: CMS HealthCare.gov API
}
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
catalogs()
\dontshow{\}) # examplesIf}
}
\keyword{internal}
