% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{class_API}
\alias{class_API}
\title{API Class}
\usage{
class_API(
  title = character(0),
  description = character(0),
  accrualPeriodicity = character(0),
  modified = numeric(0),
  temporal = character(0),
  identifier = class_Identifier(),
  accessURL = character(0),
  resourcesAPI = class_Resources(),
  downloadURL = character(0),
  describedBy = character(0),
  landingPage = character(0)
)
}
\arguments{
\item{title}{\verb{<chr>} Source}

\item{description}{\verb{<chr>} Source}

\item{accrualPeriodicity}{\verb{<chr>} Source}

\item{modified}{\verb{<chr>} Source}

\item{temporal}{\verb{<chr>} Source}

\item{identifier}{\verb{<class_Identifier>} Source}

\item{accessURL}{\verb{<chr>} Source}

\item{resourcesAPI}{\verb{<class_Resources>} Source}

\item{downloadURL}{\verb{<chr>} Source}

\item{describedBy}{\verb{<chr>} Source}

\item{landingPage}{\verb{<chr>} Source}
}
\value{
valid \code{class_API} object
}
\description{
The \code{class_API} object contains validated API metadata
}
\examples{
x <- enrollapi()

enroll <- class_API(
  x$title,
  x$description,
  x$accrualPeriodicity,
  x$modified,
  x$temporal,
  class_Identifier(url = x$identifier),
  x$accessURL,
  class_Resources(url = x$resourcesAPI),
  x$downloadURL,
  x$describedBy,
  x$landingPage
)

enroll

}
