% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{class_API}
\alias{class_API}
\title{API Class}
\usage{
class_API(
  type = "dcat:Dataset",
  accessLevel = "public",
  bureauCode = "009:38",
  programCode = "009:000",
  contactPoint = class_contactPoint(),
  dataQuality = TRUE,
  publisher = class_publisher(),
  references = character(0),
  title = character(0),
  description = character(0),
  accrualPeriodicity = character(0),
  issued = NULL,
  modified = NULL,
  released = NULL,
  temporal = character(0),
  identifier = class_Identifier(),
  accessURL = character(0),
  resourcesAPI = class_Resources(),
  downloadURL = character(0),
  describedBy = character(0),
  landingPage = character(0)
)
}
\arguments{
\item{type}{\verb{<chr>} type}

\item{accessLevel}{\verb{<chr>} Dataset access level}

\item{bureauCode}{\verb{<chr>} Dataset bureau code}

\item{programCode}{\verb{<chr>} Dataset program code}

\item{contactPoint}{\verb{<S7_class>}}

\item{dataQuality}{\verb{<lgl>} Dataset data quality}

\item{publisher}{\verb{<S7_class>}}

\item{references}{\verb{<chr>} Dataset references}

\item{title}{\verb{<chr>} Dataset title}

\item{description}{\verb{<chr>} Dataset description}

\item{accrualPeriodicity}{\verb{<chr>} Dataset update frequency}

\item{issued}{\verb{<chr>} Date Dataset was issued}

\item{modified}{\verb{<chr>} Date Dataset was last modified}

\item{released}{\verb{<chr>} Date Dataset was released}

\item{temporal}{\verb{<chr>} Date range the Current dataset covers}

\item{identifier}{\verb{<S7_class>} dcat:Dataset url and nrows in dataset}

\item{accessURL}{\verb{<chr>} dcat:Distribution url}

\item{resourcesAPI}{\verb{<S7_class>} \code{data.frame} of available supplemental resources}

\item{downloadURL}{\verb{<chr>} dcat:Distribution url to csv versions}

\item{describedBy}{\verb{<chr>} Hyperlink to Data dictionary}

\item{landingPage}{\verb{<chr>} Hyperlink to API landing page}
}
\value{
\verb{<S7_class>} object
}
\description{
\code{class_API} object
}
