% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{class_Identifier}
\alias{class_Identifier}
\title{Identifier Class}
\usage{
class_Identifier(url = NULL)
}
\arguments{
\item{url}{\verb{<chr>} Identifier url}
}
\value{
\verb{<S7_class>} object
}
\description{
\code{class_Identifier} object
}
