% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{class_Identifier}
\alias{class_Identifier}
\title{Identifier Class}
\usage{
class_Identifier(.data = character(0), url = NULL)
}
\arguments{
\item{.data}{\verb{<tibble>} Identifier data}

\item{url}{\verb{<chr>} Identifier url}
}
\value{
\verb{<tibble>} of resourcesAPI data
}
\description{
The \code{class_Identifier} object
}
\examples{
x <- class_Identifier(url = "https://data.cms.gov/data-api/v1/dataset/2457ea29-fc82-48b0-86ec-3b0755de7515/data-viewer")

x

}
