% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_backend.R
\name{class_resources}
\alias{class_resources}
\title{class_resources}
\usage{
class_resources(url = character(0))
}
\arguments{
\item{url}{\verb{<chr>} dataset URL}
}
\description{
class_resources
}
\seealso{
Other classes: 
\code{\link{Current}()}
}
\concept{classes}
