% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_main.R
\name{class_resources}
\alias{class_resources}
\title{Resources Class}
\usage{
class_resources(url = character(0))
}
\arguments{
\item{url}{\verb{<chr>} resourcesAPI URL}
}
\value{
\verb{<S7_class>} class_resources object
}
\description{
Resources Class
}
\seealso{
Other classes: 
\code{\link{Contact}()},
\code{\link{Dataset}()},
\code{\link{Identifier}()},
\code{\link{Publisher}()},
\code{\link{Resources}()}
}
\concept{classes}
