% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{class_Resources}
\alias{class_Resources}
\title{Resources Class}
\usage{
class_Resources(url = NULL)
}
\arguments{
\item{url}{\verb{<chr>} ResourcesAPI url}
}
\description{
The \code{class_Resources} object
}
