% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_backend.R
\name{class_resources}
\alias{class_resources}
\title{API Endpoint Resources Object}
\usage{
class_resources(url = character(0))
}
\arguments{
\item{url}{\verb{<chr>} resources URL}
}
\value{
An S7 \verb{<class_resources>} object.
}
\description{
API Endpoint Resources Object
}
\keyword{internal}
