% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{class_Resources}
\alias{class_Resources}
\title{Resources Class}
\usage{
class_Resources(url = NULL)
}
\arguments{
\item{url}{\verb{<chr>} resourcesAPI url}
}
\value{
\verb{<S7_class>} object
}
\description{
\code{class_Resources} object
}
