% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_affiliations.R
\name{clinicians}
\alias{clinicians}
\title{Clinicians}
\usage{
clinicians()
}
\description{
The Doctors and Clinicians utilization
data file reports volume information for
procedures of interest on clinician
profile pages and in the provider data
catalog (PDC) to inform patients and
caregivers about clinicians' experience.
}
\examples{
clinicians()
}
