% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{collection}
\alias{collection}
\title{Create Collection of Endpoints}
\usage{
collection(alias)
}
\arguments{
\item{alias}{\verb{<chr>} Alias representing the CMS data endpoint.}
}
\value{
S7 \code{class_collection} object.
}
\description{
Create Collection of Endpoints
}
\examples{
\dontshow{if (rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
collection("caid_demographics")
collection("caid_unwind")
collection("caid_services")
\dontshow{\}) # examplesIf}
}
