% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{collection}
\alias{collection}
\title{Load a collection of endpoints by alias}
\usage{
collection(alias)
}
\arguments{
\item{alias}{\verb{<chr>} collection alias}
}
\value{
An S7 \verb{<class_collection>} object.
}
\description{
Load a collection of endpoints by alias
}
\examples{
collection("unwind")
}
