% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_backend.R
\name{Current}
\alias{Current}
\title{API Current Endpoint Class}
\usage{
Current(
  title = character(0),
  description = character(0),
  contact = character(0),
  modified = character(0),
  identifier = character(0),
  rows = integer(0),
  pages = integer(0),
  fields = character(0),
  download = character(0)
)
}
\arguments{
\item{title}{\verb{<chr>} title}

\item{description}{\verb{<chr>} description}

\item{contact}{\verb{<chr>} contact name and email address}

\item{modified}{\verb{<chr>} date last modified}

\item{identifier}{\verb{<chr>} uuid}

\item{rows}{\verb{<int>} number of rows}

\item{pages}{\verb{<int>} number of pages}

\item{fields}{\verb{<chr>} field names}

\item{download}{\verb{<chr>} download URL}
}
\value{
An S7 \verb{<Current>} object.
}
\description{
API Current Endpoint Class
}
\keyword{internal}
