% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_backend.R
\name{Current}
\alias{Current}
\alias{ResourcesMain}
\alias{CurrentMain}
\alias{CurrentProvider}
\alias{CurrentOpen}
\title{Current}
\usage{
Current(
  title = character(0),
  description = character(0),
  contact = character(0),
  modified = character(0),
  identifier = character(0),
  download = character(0)
)

ResourcesMain(url = character(0))

CurrentMain(
  title = character(0),
  description = character(0),
  contact = character(0),
  modified = character(0),
  identifier = character(0),
  download = character(0),
  temporal = character(0),
  periodicity = character(0),
  resources = ResourcesMain(),
  dictionary = character(0),
  site = character(0),
  references = character(0)
)

CurrentProvider(
  title = character(0),
  description = character(0),
  contact = character(0),
  modified = character(0),
  identifier = character(0),
  download = character(0),
  issued = character(0),
  released = character(0),
  site = character(0)
)

CurrentOpen(
  title = character(0),
  description = character(0),
  contact = character(0),
  modified = character(0),
  identifier = character(0),
  download = character(0)
)
}
\arguments{
\item{title}{\verb{<chr>} dataset title}

\item{description}{\verb{<chr>} dataset description}

\item{contact}{\verb{<chr>} dataset contact}

\item{modified}{\verb{<chr>} dataset modified date}

\item{identifier}{\verb{<chr>} dataset identifier}

\item{download}{\verb{<chr>} dataset download URL}

\item{url}{\verb{<chr>} dataset URL}

\item{temporal}{\verb{<chr>} dataset temporal}

\item{periodicity}{\verb{<chr>} dataset periodicity}

\item{resources}{\verb{<ResourcesMain>} dataset resources}

\item{dictionary}{\verb{<chr>} dataset dictionary}

\item{site}{\verb{<chr>} dataset site}

\item{references}{\verb{<chr>} dataset references}

\item{issued}{\verb{<chr>} dataset issued date}

\item{released}{\verb{<chr>} dataset released date}
}
\description{
Current

ResourcesMain

CurrentMain

CurrentProvider

CurrentOpen
}
\seealso{
Other base-classes: 
\code{\link{Temporal}()}

Other main-classes: 
\code{\link{Temporal}()}

Other main-classes: 
\code{\link{Temporal}()}

Other open-classes: 
\code{\link{Temporal}()}
}
\concept{base-classes}
\concept{main-classes}
\concept{open-classes}
\concept{provider-classes}
