% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_backend.R
\name{CurrentMain}
\alias{CurrentMain}
\title{CurrentMain API Endpoint}
\usage{
CurrentMain(alias)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint alias}
}
\value{
An S7 \verb{<CurrentMain>} object.
}
\description{
CurrentMain API Endpoint
}
\examples{
CurrentMain("enrollees")
CurrentMain("opt_out")
CurrentMain("order_refer")
CurrentMain("reassignments")
CurrentMain("hospitals")
CurrentMain("laboratories")
CurrentMain("crosswalk")
CurrentMain("rbcs")
CurrentMain("facilities")
CurrentMain("home_health")
CurrentMain("hospice")
CurrentMain("dialysis")
CurrentMain("snf")

}
