% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_backend.R
\name{CurrentOpen}
\alias{CurrentOpen}
\title{CurrentOpen Payments API Endpoint}
\usage{
CurrentOpen(alias)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint alias}
}
\value{
An S7 \verb{<CurrentOpen>} object.
}
\description{
CurrentOpen Payments API Endpoint
}
\examples{
CurrentOpen("prof_cov")
CurrentOpen("prof_phys")
CurrentOpen("prof_info")
CurrentOpen("prof_map")
CurrentOpen("prof_entity")
CurrentOpen("prof_teach")
CurrentOpen("dashboard")
CurrentOpen("pay_state_total")
CurrentOpen("pay_state_group")
CurrentOpen("pay_nat_group")
CurrentOpen("pay_nat_total")

}
