% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{Dataset}
\alias{Dataset}
\title{Dataset Class}
\usage{
Dataset(
  type = "dcat:Dataset",
  accessLevel = "public",
  accrualPeriodicity = character(0),
  bureauCode = "009:38",
  contactPoint = dataset_contactPoint(),
  describedBy = character(0),
  description = character(0),
  identifier = dataset_Identifier(),
  keyword = character(0),
  landingPage = character(0),
  modified = NULL,
  programCode = "009:000",
  publisher = dataset_Publisher(),
  references = character(0),
  temporal = character(0),
  title = character(0),
  resourcesAPI = dataset_Resources()
)
}
\arguments{
\item{type}{\verb{<chr>} Schema type; default is \code{dcat:Dataset}}

\item{accessLevel}{\verb{<chr>} Dataset access level; default is \code{public}}

\item{accrualPeriodicity}{\verb{<chr>} Dataset update frequency}

\item{bureauCode}{\verb{<chr>} Dataset bureau code; default is \code{009:38}}

\item{contactPoint}{\verb{<S7_class>} Dataset contact}

\item{describedBy}{\verb{<chr>} Hyperlink to Data dictionary}

\item{description}{\verb{<chr>} Dataset description}

\item{identifier}{\verb{<S7_class>} dcat:Dataset url and nrows in dataset}

\item{keyword}{\verb{<chr>} Hyperlink to API landing page}

\item{landingPage}{\verb{<chr>} Hyperlink to API landing page}

\item{modified}{\verb{<chr>} Date Dataset was last modified}

\item{programCode}{\verb{<chr>} Dataset program code; default is \code{009:000}}

\item{publisher}{\verb{<S7_class>} Dataset publisher}

\item{references}{\verb{<chr>} Dataset references}

\item{temporal}{\verb{<chr>} Date range the Current dataset covers}

\item{title}{\verb{<chr>} Dataset title}

\item{resourcesAPI}{\verb{<S7_class>} \code{data.frame} of available supplemental resources}
}
\value{
\verb{<S7_class>} Dataset object
}
\description{
\code{Dataset} object
}
