% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{Dataset}
\alias{Dataset}
\title{Dataset Class}
\usage{
Dataset(
  type = "dcat:Dataset",
  accessLevel = "public",
  accrualPeriodicity = character(0),
  bureauCode = "009:38",
  contactPoint = class_contactPoint(),
  describedBy = character(0),
  description = character(0),
  identifier = class_Identifier(),
  keyword = list(),
  landingPage = character(0),
  modified = NULL,
  programCode = "009:000",
  publisher = class_publisher(),
  references = character(0),
  temporal = character(0),
  title = character(0),
  accessURL = character(0),
  resourcesAPI = class_Resources(),
  downloadURL = character(0)
)
}
\arguments{
\item{type}{\verb{<chr>} type}

\item{accessLevel}{\verb{<chr>} Dataset access level}

\item{accrualPeriodicity}{\verb{<chr>} Dataset update frequency}

\item{bureauCode}{\verb{<chr>} Dataset bureau code}

\item{contactPoint}{\verb{<S7_class>}}

\item{describedBy}{\verb{<chr>} Hyperlink to Data dictionary}

\item{description}{\verb{<chr>} Dataset description}

\item{identifier}{\verb{<S7_class>} dcat:Dataset url and nrows in dataset}

\item{keyword}{\verb{<list>} Hyperlink to API landing page}

\item{landingPage}{\verb{<chr>} Hyperlink to API landing page}

\item{modified}{\verb{<chr>} Date Dataset was last modified}

\item{programCode}{\verb{<chr>} Dataset program code}

\item{publisher}{\verb{<S7_class>}}

\item{references}{\verb{<chr>} Dataset references}

\item{temporal}{\verb{<chr>} Date range the Current dataset covers}

\item{title}{\verb{<chr>} Dataset title}

\item{accessURL}{\verb{<chr>} dcat:Distribution url}

\item{resourcesAPI}{\verb{<S7_class>} \code{data.frame} of available supplemental resources}

\item{downloadURL}{\verb{<chr>} dcat:Distribution url to csv versions}
}
\value{
\verb{<S7_class>} object
}
\description{
\code{Dataset} object
}
