% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{dataset_contactPoint}
\alias{dataset_contactPoint}
\title{dataset_contactPoint Class}
\usage{
dataset_contactPoint(
  type = "vcard:Contact",
  fn = NULL,
  hasEmail = NULL,
  hasURL = NULL
)
}
\arguments{
\item{type}{\verb{<chr>} Schema contact type; default is \code{vcard:Contact}}

\item{fn}{\verb{<chr>} contact's full name}

\item{hasEmail}{\verb{<chr>} contact's email}

\item{hasURL}{\verb{<chr>} contact's url}
}
\value{
\verb{<S7_class>} dataset_contactPoint object
}
\description{
\code{dataset_contactPoint} object
}
