% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{endpoint}
\alias{endpoint}
\title{Load API Endpoint}
\usage{
endpoint(alias)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint alias}
}
\value{
S7 \verb{<class_[Catalog Name]>} object.
}
\description{
Load API Endpoint
}
\examples{
endpoint("care_dial_end")
endpoint("care_dial_tmp")
endpoint("managed_mltss")
endpoint("hgov_ab_reg_comp")
endpoint("asc_facility")
endpoint("dialysis_by_facility")
}
