% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{endpoint}
\alias{endpoint}
\title{Endpoint from Catalog
Load an endpoint by alias.}
\usage{
endpoint(alias)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint alias}
}
\value{
An S7 \verb{<class_care/caid/prov/open/hgov>} object.
}
\description{
Endpoint from Catalog
Load an endpoint by alias.
}
\examples{
endpoint("care_dial_end")
endpoint("care_dial_tmp")
endpoint("managed_mltss")
endpoint("hgov_ab_reg_comp")
endpoint("asc_facility")
endpoint("dialysis_by_facility")
}
