% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{endpoint}
\alias{endpoint}
\title{Retrieve Identifier URL}
\usage{
endpoint(x)
}
\arguments{
\item{x}{\verb{<S7_class>} Dataset}
}
\value{
\verb{<chr>} Identifier URL
}
\description{
Retrieve Identifier URL
}
\keyword{internal}
