% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{endpoint}
\alias{endpoint}
\alias{collection}
\alias{group}
\title{Load Endpoints}
\usage{
endpoint(alias)

collection(alias)

group(..., .title = NULL)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint or collection alias}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> endpoint aliases to be grouped}

\item{.title}{\verb{<chr>} Group title. Defaults to \code{NULL},
which will concatenate the aliases for the title.}
}
\value{
An S7 \verb{<class_care/caid/prov/open/hgov/collection/group>} object.
}
\description{
Load a data endpoint or collection by its alias.
}
\details{
The \code{endpoint()} function loads a single data endpoint, while the
\code{collection()} function loads a collection of endpoints, and the \code{group()}
function loads a group of endpoints. The \code{alias} parameter is used to specify
the endpoint or collection alias. If the alias is not found, an error will be
raised.
}
\examples{
endpoint("dial_facility")
endpoint("man_state")
try(endpoint(c("asc_facility", "enterprise")))
try(endpoint("unwind"))
try(endpoint("ex"))
collection("unwind")
try(collection(c("asc_facility", "enterprise")))
try(collection("asc_facility"))
group("asc_facility", "enterprise")
try(group("asc_facility"))
try(group("util"))
}
