% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrollees.R
\name{enrollee_API}
\alias{enrollee_API}
\title{Public Provider Enrollment Experiment}
\usage{
enrollee_API()
}
\value{
\verb{<S7_class>} object
}
\description{
Public Provider Enrollment Experiment
}
\examples{
enrollee_API()

}
