% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{enrolleeAPI}
\alias{enrolleeAPI}
\title{Enrollee API Class}
\usage{
enrolleeAPI(
  title = character(0),
  description = character(0),
  periodicity = character(0),
  modified = (function (.data = double()) 
 {
     .Date(.data)
 })(),
  temporal = (function (.data = double()) 
 {
     .Date(.data)
 })(),
  identifier = list(),
  accessurl = list(),
  resourceapi = list(),
  downloadurl = character(0),
  dictionary = character(0),
  landingpage = character(0)
)
}
\arguments{
\item{title}{\verb{<chr>} Source}

\item{description}{\verb{<chr>} Source}

\item{periodicity}{\verb{<chr>} Source}

\item{modified}{\verb{<chr>} Source}

\item{temporal}{\verb{<chr>} Source}

\item{identifier}{\verb{<chr>} Source}

\item{accessurl}{\verb{<chr>} Source}

\item{resourceapi}{\verb{<chr>} Source}

\item{downloadurl}{\verb{<chr>} Source}

\item{dictionary}{\verb{<chr>} Source}

\item{landingpage}{\verb{<chr>} Source}
}
\value{
valid \code{enrolleeAPI} classed object
}
\description{
The \code{enrolleeAPI} object contains validated API metadata
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

if (!exists(".__public")) .__public <<- public_dataset()

x <- as.list(
     collapse::sbt(.__public[["dataset"]],
     sf_detect(title, "Public Provider Enrollment")))

y <- as.list(
     collapse::sbt(.__public[["api"]],
     sf_detect(title, "Public Provider Enrollment"))[1, 4:5])

z <- collapse::sbt(.__public[["csv"]],
     sf_detect(title, "Public Provider Enrollment"),
     downloadURL)[1,][[1]]

enrolleeAPI(
  title       = x$title,
  description = x$description,
  periodicity = x$accrualPeriodicity,
  modified    = x$modified,
  temporal    = x$temporal,
  identifier  = x$identifier,
  accessurl   = y$accessURL,
  resourceapi = y$resourcesAPI,
  downloadurl = z,
  dictionary  = x$describedBy,
  landingpage = x$landingPage)
\dontshow{\}) # examplesIf}
}
