% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{enrolleeAPI}
\alias{enrolleeAPI}
\title{Enrollee API Class}
\usage{
enrolleeAPI(
  title = character(0),
  description = character(0),
  periodicity = character(0),
  modified = numeric(0),
  temporal = character(0),
  identifier = NULL,
  accessurl = NULL,
  resourceapi = NULL,
  downloadurl = character(0),
  dictionary = character(0),
  landingpage = character(0)
)
}
\arguments{
\item{title}{\verb{<chr>} Source}

\item{description}{\verb{<chr>} Source}

\item{periodicity}{\verb{<chr>} Source}

\item{modified}{\verb{<chr>} Source}

\item{temporal}{\verb{<chr>} Source}

\item{identifier}{\verb{<chr>} Source}

\item{accessurl}{\verb{<chr>} Source}

\item{resourceapi}{\verb{<chr>} Source}

\item{downloadurl}{\verb{<chr>} Source}

\item{dictionary}{\verb{<chr>} Source}

\item{landingpage}{\verb{<chr>} Source}
}
\value{
valid \code{enrolleeAPI} classed object
}
\description{
The \code{enrolleeAPI} object contains validated API metadata
}
\examples{
x <- enrollapi()

enroll <- enrolleeAPI(
          x$title,
          x$description,
          x$accrualPeriodicity,
          x$modified,
          x$temporal,
          x$identifier,
          x$accessURL,
          x$resourcesAPI,
          x$downloadURL,
          x$describedBy,
          x$landingPage)

enroll

}
