% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_enrollees.R
\name{enrollees}
\alias{enrollees}
\title{Medicare Provider Enrollees}
\usage{
enrollees(
  npi = NULL,
  pac = NULL,
  enid = NULL,
  spec_code = NULL,
  spec_desc = NULL,
  state = NULL,
  first = NULL,
  middle = NULL,
  last = NULL,
  org = NULL,
  gender = NULL
)
}
\arguments{
\item{npi}{\verb{<chr>} 10-digit Individual NPI}

\item{pac}{\verb{<chr>} 10-digit PECOS Associate Control (PAC) ID}

\item{enid}{\verb{<chr>} 15-digit Medicare Enrollment ID}

\item{spec_code}{\verb{<chr>} Enrollment specialty code}

\item{spec_desc}{\verb{<chr>} Enrollment specialty description}

\item{state}{\verb{<chr>} Enrollment state abbreviation}

\item{first, middle, last}{\verb{<chr>} Individual provider's name(s)}

\item{org}{\verb{<chr>} Organizational provider's name}

\item{gender}{\verb{<chr>} Individual provider's gender:
\itemize{
\item \code{"F"} (Female)
\item \code{"M"} (Male)
\item \code{"9"} (Unknown/Organization)
}}
}
\value{
\verb{<tibble>} of search results
}
\description{
Individual & Organizational Enrollment-level Data
on Providers Actively Approved to Bill Medicare.
}
\section{Accrual Periodicity}{

\itemize{
\item Quarterly (R/P3M)
}
}

\section{Links}{

\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-fee-for-service-public-provider-enrollment}{Enrollment API}
\item \href{https://data.cms.gov/resources/medicare-fee-for-service-public-provider-enrollment-data-dictionary}{Enrollment Data Dictionary}
}
}

\examples{
\dontshow{if (rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
enrollees(enid = "I20040309000221")

enrollees(npi = "1417918293", spec_code = "14-41")

enrollees(pac = "2860305554", gender = "9")

enrollees(state = "GA", gender = "9")

enrollees(state = "GA", gender = "F")
\dontshow{\}) # examplesIf}
}
