% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2_get_nrows_fields.R
\name{fields_provider}
\alias{fields_provider}
\title{Request field names from provider catalog}
\usage{
fields_provider(url)
}
\arguments{
\item{url}{\verb{<chr>} API URL}
}
\value{
\verb{<chr>} Field names
}
\description{
Request field names from provider catalog
}
\keyword{internal}
