% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{format_query}
\alias{format_query}
\title{Format API Queries}
\usage{
format_query(args)
}
\arguments{
\item{args}{\verb{<list>} of \verb{<chr>} arguments}
}
\value{
\verb{<list>} of formatted query \verb{<expr>}
}
\description{
Format API Queries
}
\examples{
format_query(list("NPI" = "1417918293", "PECOS_ASCT_CNTL_ID" = NULL))

}
