% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nrows_fields.R
\name{get_fields}
\alias{get_fields}
\title{Request field names from catalog}
\usage{
get_fields(url)
}
\arguments{
\item{url}{\verb{<chr>} API URL}
}
\value{
\verb{<chr>} Field names
}
\description{
Request field names from catalog
}
\keyword{internal}
