% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nrows_fields.R
\name{get_resources}
\alias{get_resources}
\title{Get Resources}
\usage{
get_resources(url)
}
\arguments{
\item{url}{resourcesAPI url}
}
\value{
\verb{<tibble>} of Resource Files Data
}
\description{
Get Resources
}
\keyword{internal}
