% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{group}
\alias{group}
\title{Create Group of Endpoints}
\usage{
group(alias, description = NULL)
}
\arguments{
\item{alias}{\verb{<chr>} Alias representing the CMS data endpoint.}

\item{description}{\verb{<chr>} Group description. Defaults to \code{NULL},
which will use the aliases as the description.}
}
\value{
S7 \verb{<class_group>} object.
}
\description{
Create Group of Endpoints
}
\examples{
\dontshow{if (rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
group(c("cahps_hhc_patient", "hgov_qhp_business"))
group(c("hgov_local_help", "hgov_qhp_business"))
group(c("asc_facility", "care_dial_end"))
\dontshow{\}) # examplesIf}
}
