% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{group}
\alias{group}
\title{Create Groups of Endpoints}
\usage{
group(alias, desc = NULL)
}
\arguments{
\item{alias}{\verb{<chr>} Alias representing the CMS data endpoint.}

\item{desc}{\verb{<chr>} Name of the group.}
}
\value{
S7 \code{class_group} object.
}
\description{
Create Groups of Endpoints
}
\examples{
\dontshow{if (rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
group(c("cahps_hhc_patient", "hgov_qhp_business"))
group(c("hgov_local_help", "hgov_qhp_business"))
\dontshow{\}) # examplesIf}
}
