% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_group.R
\name{groups}
\alias{groups}
\alias{new_group}
\alias{pro_group}
\title{Create Groups of Endpoints}
\usage{
new_group(
  member_names,
  group_name = NULL,
  quick_call = FALSE,
  offset = 0,
  limit = 10
)

pro_group(alias, ...)
}
\arguments{
\item{member_names}{\verb{<chr>} Alias representing the CMS data endpoint.}

\item{group_name}{\verb{<chr>} Name of the group.}

\item{quick_call}{\verb{<lgl>} call \code{quick_()} method on group if \code{TRUE}.}

\item{offset}{\verb{<int>} Offset for pagination.}

\item{limit}{\verb{<int>} Limit for pagination.}

\item{alias}{\verb{<chr>} endpoint alias}

\item{...}{Additional arguments passed to the group constructor}
}
\value{
S7 \code{class_group} object.
}
\description{
Create Groups of Endpoints
}
\examples{
new_group(c("qhp_bus_rule_variables", "managed_care_share"))
new_group(c("qhp_bus_rule_variables", "managed_care_share"), quick_call = TRUE)
pro_group("pro_dialysis")
}
