% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{Identifier}
\alias{Identifier}
\title{Identifier Class}
\usage{
Identifier(url = character(0))
}
\arguments{
\item{url}{\verb{<chr>} identifier url}
}
\value{
\verb{<S7_class>} Identifier object
}
\description{
\code{Identifier} object
}
\examples{
Identifier(url =
   paste0(
   "https://data.cms.gov/data-api/v1/dataset/",
   "2457ea29-fc82-48b0-86ec-3b0755de7515/",
   "data-viewer"))

}
