% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{Identifier}
\alias{Identifier}
\title{Identifier Class}
\usage{
Identifier(url = NULL)
}
\arguments{
\item{url}{\code{NULL} | \code{class_list} API identifier url}
}
\description{
The \code{Identifier} class object contains validated API metadata
}
