% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_method_resources.R
\name{list_resources}
\alias{list_resources}
\title{List resources}
\usage{
list_resources(x)
}
\arguments{
\item{x}{\code{careMain}, \code{careGroup}, \code{careTemp}, or \code{careTempGroup} object}
}
\value{
A list of API resources
}
\description{
List resources
}
\examples{
careMain("enrollees") |> list_resources()
careGroup("HHA") |> list_resources()
careTemp("quality_payment") |> list_resources()
#careTempGroup("inpatient") |> list_resources()
}
