% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_method_resources.R
\name{list_resources}
\alias{list_resources}
\title{List resources}
\usage{
list_resources(x)
}
\arguments{
\item{x}{\code{care_endpoint}, \code{care_group}, \code{care_temporal}, or \code{care_troup} object}
}
\value{
A list of API resources
}
\description{
List resources
}
\examples{
care_endpoint("enrollees") |> list_resources()
care_group("HHA") |> list_resources()
care_temporal("quality_payment") |> list_resources()
care_troup("inpatient") |> list_resources()
}
