% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_resources.R
\name{list_resources}
\alias{list_resources}
\title{List resources}
\usage{
list_resources(x)
}
\arguments{
\item{x}{\code{class_endpoint} or \code{class_temporal} object}
}
\value{
A list of available API resources, either for download or to browse online.
}
\description{
List resources
}
\examples{
\dontshow{if (rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
new_endpoint("care_enroll_prov") |> list_resources()
new_endpoint("quality_payment") |> list_resources()
new_collection("care_hha") |> list_resources()
new_collection("care_in") |> list_resources()
\dontshow{\}) # examplesIf}
}
