% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_resources.R
\name{list_resources}
\alias{list_resources}
\title{List resources}
\usage{
list_resources(obj)
}
\arguments{
\item{obj}{\verb{<class_care>} object}
}
\value{
A list of available API resources.
}
\description{
List resources
}
\examples{
roll <- endpoint("enroll_prov")
qppe <- endpoint("qppe")
hosp <- collection("in_hosp")
rand <- group("asc_facility", "enterprise", "lab_fee")
list_resources(roll)
list(qppe, hosp, rand) |> purrr::map(list_resources)
}
