% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_list_resources.R
\name{list_resources}
\alias{list_resources}
\title{List resources from a \code{MainCurrent} or \code{class_character} object}
\usage{
list_resources(x, ...)
}
\arguments{
\item{x}{An object of class \code{MainCurrent} or \code{class_character}}

\item{...}{Additional arguments?}
}
\value{
A list of API resources
}
\description{
List resources from a \code{MainCurrent} or \code{class_character} object
}
\examples{
MainCurrent("enrollees") |> list_resources()

}
