% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_list_resources.R
\name{list_resources}
\alias{list_resources}
\title{List resources}
\usage{
list_resources(x)
}
\arguments{
\item{x}{\code{care_endpoint}, \code{care_group}, \code{care_temporal}}
}
\value{
A list of API resources
}
\description{
List resources
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
care_endpoint("care_enrollees") |> list_resources()
care_temporal("quality_payment") |> list_resources()
care_group("care_hha") |> list_resources()
care_troup("care_inpatient") |> list_resources()
\dontshow{\}) # examplesIf}
}
