% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{load_openpayments}
\alias{load_openpayments}
\title{CMS Open Payments Catalog}
\usage{
load_openpayments()
}
\value{
\verb{<list>} of \verb{<tibbles>}: \code{dataset}, \code{distribution}, \code{downloads}
}
\description{
CMS Open Payments Catalog
}
\examples{
load_openpayments()

}
