% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_backend.R, R/S7_method_resources.R
\name{careMain}
\alias{careMain}
\alias{careTemp}
\alias{list_resources}
\title{Current Main API Endpoint}
\usage{
careMain(alias)

careTemp(alias)

list_resources(x)
}
\arguments{
\item{alias}{\verb{<chr>} title alias}

\item{x}{Object of class \code{careMain} or \code{careTemp}}
}
\value{
An S7 \verb{<MainCurrent>} object.

An S7 \verb{<careTemp>} object.

A list of API resources
}
\description{
Current Main API Endpoint

Medicare Temporal Endpoint

List resources
}
\examples{
careMain("enrollees")
careMain("opt_out")
careMain("order_refer")
careMain("reassignments")
careMain("hospitals")
careMain("laboratories")
careMain("crosswalk")
careMain("rbcs")
careMain("facilities")
careMain("home_health")
careMain("hospice")
careMain("dialysis")
careMain("skilled_nursing")
careTemp("quality_payment")
careMain("enrollees") |> list_resources()
careTemp("quality_payment") |> list_resources()

}
