% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_backend.R, R/S7_method_resources.R
\name{MainCurrent}
\alias{MainCurrent}
\alias{MainTemporal}
\alias{MainTemporalGroup}
\alias{list_resources}
\title{Current Main API Endpoint}
\usage{
MainCurrent(alias)

MainTemporal(alias)

MainTemporalGroup(alias)

list_resources(x, ...)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint alias}

\item{x}{An object of class \code{MainCurrent} or \code{class_character}}

\item{...}{Additional arguments?}
}
\value{
An S7 \verb{<MainCurrent>} object.

An S7 \verb{<MainTemporal>} object.

An S7 \verb{<MainTemporalGroup>} object.

A list of API resources
}
\description{
Current Main API Endpoint

TemporalMain API Endpoint

Main API Endpoint Temporal Group

List resources
}
\examples{
MainCurrent("enrollees")
MainCurrent("opt_out")
MainCurrent("order_refer")
MainCurrent("reassignments")
MainCurrent("hospitals")
MainCurrent("laboratories")
MainCurrent("crosswalk")
MainCurrent("rbcs")
MainCurrent("facilities")
MainCurrent("home_health")
MainCurrent("hospice")
MainCurrent("dialysis")
MainCurrent("snf")

MainTemporal("quality_payment")

MainTemporalGroup("utilization")
MainTemporalGroup("prescribers")
MainTemporalGroup("suppliers")
MainTemporalGroup("outpatient")
MainTemporalGroup("inpatient")

MainCurrent("enrollees") |> list_resources()
MainTemporal("quality_payment") |> list_resources()

}
