% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_backend.R
\name{careMain}
\alias{careMain}
\alias{careTemp}
\title{Current Main API Endpoint}
\usage{
careMain(alias)

careTemp(alias)
}
\arguments{
\item{alias}{\verb{<chr>} title alias}
}
\value{
An S7 \verb{<MainCurrent>} object.

An S7 \verb{<careTemp>} object.
}
\description{
Current Main API Endpoint

Medicare Temporal Endpoint
}
\examples{
careMain("enrollees")
careMain("opt_out")
careMain("order_refer")
careMain("reassignments")
careMain("hospitals")
careMain("laboratories")
careMain("crosswalk")
careMain("rbcs")
careMain("facilities")
careMain("home_health")
careMain("hospice")
careMain("dialysis")
careMain("skilled_nursing")
careTemp("quality_payment")
}
