% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_backend.R, R/S7_method_resources.R
\name{MainCurrent}
\alias{MainCurrent}
\alias{mainTemp}
\alias{mainTempGroup}
\alias{list_resources}
\title{Current Main API Endpoint}
\usage{
MainCurrent(alias)

mainTemp(alias)

mainTempGroup(alias)

list_resources(x)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint alias}

\item{x}{Object of class \code{MainCurrent} or \code{mainTemp}}
}
\value{
An S7 \verb{<MainCurrent>} object.

An S7 \verb{<mainTemp>} object.

An S7 \verb{<mainTempGroup>} object.

A list of API resources
}
\description{
Current Main API Endpoint

Temporal Main API Endpoint

Main API Endpoint Temporal Group

List resources
}
\examples{
MainCurrent("enrollees")
MainCurrent("opt_out")
MainCurrent("order_refer")
MainCurrent("reassignments")
MainCurrent("hospitals")
MainCurrent("laboratories")
MainCurrent("crosswalk")
MainCurrent("rbcs")
MainCurrent("facilities")
MainCurrent("home_health")
MainCurrent("hospice")
MainCurrent("dialysis")
MainCurrent("skilled_nursing")
mainTemp("quality_payment")
mainTempGroup("utilization")
mainTempGroup("prescribers")
mainTempGroup("suppliers")
mainTempGroup("outpatient")
mainTempGroup("inpatient")
MainCurrent("enrollees") |> list_resources()
mainTemp("quality_payment") |> list_resources()

}
