% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_backend.R
\name{MainCurrent}
\alias{MainCurrent}
\alias{MainTemporal}
\alias{MainTemporalGroup}
\title{Current Main API Endpoint}
\usage{
MainCurrent(alias)

MainTemporal(alias)

MainTemporalGroup(alias)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint alias}
}
\value{
An S7 \verb{<MainCurrent>} object.

An S7 \verb{<MainTemporal>} object.

An S7 \verb{<MainTemporalGroup>} object.
}
\description{
Current Main API Endpoint

TemporalMain API Endpoint

Main API Endpoint Temporal Group
}
\examples{
MainCurrent("enrollees")
MainCurrent("opt_out")
MainCurrent("order_refer")
MainCurrent("reassignments")
MainCurrent("hospitals")
MainCurrent("laboratories")
MainCurrent("crosswalk")
MainCurrent("rbcs")
MainCurrent("facilities")
MainCurrent("home_health")
MainCurrent("hospice")
MainCurrent("dialysis")
MainCurrent("snf")

MainTemporal("quality_payment")

MainTemporalGroup("utilization")
MainTemporalGroup("prescribers")
MainTemporalGroup("suppliers")
MainTemporalGroup("outpatient")
MainTemporalGroup("inpatient")

}
