% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_main.R
\name{main_current}
\alias{main_current}
\title{Load Main API \code{CurrentMain}}
\usage{
main_current(alias)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint alias}
}
\value{
\verb{<CurrentMain>} object
}
\description{
Load Main API \code{CurrentMain}
}
\examples{
main_current("enrollees")
main_current("opt_out")
main_current("order_refer")
main_current("reassignments")
main_current("hospitals")
main_current("laboratories")
main_current("crosswalk")
main_current("rbcs")
main_current("facilities")
main_current("home_health")
main_current("hospice")
main_current("dialysis")
main_current("snf")
}
