% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_main.R
\name{main_nrows_fields}
\alias{main_nrows_fields}
\title{Get Field Names and Number of Rows from Main Endpoint}
\usage{
main_nrows_fields(url)
}
\arguments{
\item{url}{\verb{<chr>} endpoint URL}
}
\value{
\verb{<list>} number of rows and field names
}
\description{
Get Field Names and Number of Rows from Main Endpoint
}
\keyword{internal}
