% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_main.R
\name{main_temporal}
\alias{main_temporal}
\title{Load Public API \code{Distribution}}
\usage{
main_temporal(dataset)
}
\arguments{
\item{dataset}{\verb{<chr>} dataset title}
}
\value{
\verb{<Distribution>} object
}
\description{
Load Public API \code{Distribution}
}
\examples{
main_temporal("utilization_provider")

main_temporal("quality_payment")

}
