% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_main.R
\name{main_temporal_group}
\alias{main_temporal_group}
\title{Load \verb{<TemporalMainGroup>} API Endpoints}
\usage{
main_temporal_group(alias = "utilization")
}
\arguments{
\item{alias}{\verb{<chr>} dataset title}
}
\value{
\verb{<TemporalMainGroup>} object
}
\description{
Load \verb{<TemporalMainGroup>} API Endpoints
}
\examples{
main_temporal_group("utilization")
main_temporal_group("prescribers")
main_temporal_group("suppliers")
main_temporal_group("outpatient")
main_temporal_group("inpatient")
}
