% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_care.R
\name{medicare}
\alias{medicare}
\alias{care_endpoint}
\alias{care_temporal}
\alias{pro_endpoint}
\title{Medicare API Endpoint Classes}
\usage{
care_endpoint(alias)

care_temporal(alias)

pro_endpoint(alias)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint alias}
}
\value{
An S7 \verb{<care_endpoint>}, or \verb{<care_temporal>} object
}
\description{
Medicare API Endpoint Classes
}
\examples{
care_endpoint("care_dialysis")
care_temporal("quality_payment")
pro_endpoint("asc_facility")
}
