% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-alias_care.R
\name{medicare}
\alias{medicare}
\alias{new_endpoint}
\title{Medicare API Endpoints}
\usage{
new_endpoint(alias)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint alias}
}
\value{
An S7 \verb{<class_care>} object
}
\description{
Medicare API Endpoints
}
\examples{
\dontshow{if (rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
new_endpoint("care_dialysis")
new_endpoint("quality_payment")
\dontshow{\}) # examplesIf}
}
