% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_care.R
\name{medicare}
\alias{medicare}
\alias{care_endpoint}
\alias{care_group}
\alias{care_temporal}
\alias{care_troup}
\title{Medicare API Endpoint Classes}
\usage{
care_endpoint(alias)

care_group(alias)

care_temporal(alias)

care_troup(alias)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint alias}
}
\value{
An S7 \verb{<care_endpoint>}, \verb{<care_group>}, \verb{<care_temporal>}, or \verb{<care_troup>} object
}
\description{
Medicare API Endpoint Classes
}
\examples{
care_endpoint("CARE_dialysis")
care_endpoint("long_term")
care_endpoint("transparency")
care_temporal("quality_payment")
care_group("HHA")
care_group("hospice")
care_group("hospital")
care_group("RHC")
care_group("FQHC")
care_group("pending")
care_group("reassignment")
care_group("SNF")
care_troup("inpatient")
care_troup("outpatient")
care_troup("utilization")
care_troup("suppliers")
care_troup("prescribers")
care_troup("staffing")
}
