% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-alias_care.R
\name{medicare}
\alias{medicare}
\alias{new_endpoint}
\alias{care_collection}
\title{Medicare API Endpoints}
\usage{
new_endpoint(alias)

care_collection(alias, call = caller_env(), ...)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint alias}

\item{call}{\verb{<env>} environment to use for error reporting}

\item{...}{Additional arguments passed to the group constructor}
}
\value{
An S7 \verb{<class_care>} object
}
\description{
Medicare API Endpoints
}
\examples{
\dontshow{if (rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
care_endpoint("care_dialysis")
care_endpoint("quality_payment")
care_collection("care_hospital")
\dontshow{\}) # examplesIf}
}
