% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_modifiers.R
\name{modifier_}
\alias{modifier_}
\title{Query Modifier Constructor}
\usage{
modifier_(operator, value)
}
\arguments{
\item{operator}{\verb{<chr>} comparison operator}

\item{value}{\verb{<any>} value to compare against}
}
\value{
object of class \code{"modifier"}
}
\description{
Query Modifier Constructor
}
\examples{
modifier_(
  operator = ">",
  value    = 1000)
}
\keyword{internal}
