% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_count.R
\name{ncount}
\alias{ncount}
\title{Return the number of results for a given query}
\usage{
ncount(obj, ...)
}
\arguments{
\item{obj}{An object of class \verb{<class_care>}, \verb{<class_group>}, or \verb{<class_collection>}.}

\item{...}{Additional arguments passed to the query.}
}
\value{
An \verb{<int>} representing the number of results found.
}
\description{
Return the number of results for a given query
}
\examples{
end_dial <- endpoint("care_dial_end")
qr_state <- query(state = in_(c("CA", "GA", "NY"), care = TRUE),
                  .type = "medicare")

ncount(end_dial, qr_state)

end_rhc <- collection("care_rhc")
qr_own  <- query(STATE = in_(c("CA", "GA", "NY"), care = TRUE),
                `STATE - OWNER` = in_(c("CA", "GA", "NY"), care = TRUE),
                .type = "medicare")

ncount(end_rhc, qr_own)

}
