% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_group.R
\name{new_group}
\alias{new_group}
\title{Create Groups of Endpoints}
\usage{
new_group(
  member_names,
  group_name = NULL,
  quick = FALSE,
  offset = 0,
  limit = 10
)
}
\arguments{
\item{member_names}{\verb{<chr>} Alias representing the CMS data endpoint.}

\item{group_name}{\verb{<chr>} Name of the group.}

\item{quick}{\verb{<lgl>} call \code{quick_()} method on group if \code{TRUE}.}

\item{offset}{\verb{<int>} Offset for pagination.}

\item{limit}{\verb{<int>} Limit for pagination.}
}
\value{
S7 \code{class_group} object.
}
\description{
Create Groups of Endpoints
}
\examples{
new_group(c("local_help", "managed_care_share"))
new_group(c("local_help", "managed_care_share"), quick = TRUE)
}
