% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-new_end.R
\name{new_group}
\alias{new_group}
\title{Create Groups of Endpoints}
\usage{
new_group(members, name = NULL, quick = FALSE, offset = 0, limit = 10)
}
\arguments{
\item{members}{\verb{<chr>} Alias representing the CMS data endpoint.}

\item{name}{\verb{<chr>} Name of the group.}

\item{quick}{\verb{<lgl>} call \code{quick_()} method on group if \code{TRUE}.}

\item{offset}{\verb{<int>} Offset for pagination.}

\item{limit}{\verb{<int>} Limit for pagination.}
}
\value{
S7 \code{class_group} object.
}
\description{
Create Groups of Endpoints
}
\examples{
\dontshow{if (rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
new_group(c("hgov_local_help", "hgov_qhp_business"))
new_group(c("hgov_local_help", "hgov_qhp_business"), quick = TRUE)
\dontshow{\}) # examplesIf}
}
