% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_group.R
\name{new_group}
\alias{new_group}
\title{Create a New Group of Endpoints}
\usage{
new_group(x, q = FALSE, offset = 0, limit = 10)
}
\arguments{
\item{x}{\verb{<chr>} Alias representing the CMS data endpoint.}

\item{q}{\verb{<lgl>} return a call to quick_() if \code{TRUE}.}

\item{offset}{\verb{<int>} Offset for pagination.}

\item{limit}{\verb{<int>} Limit for pagination.}
}
\value{
S7 \code{class_group} object.
}
\description{
Create a New Group of Endpoints
}
\examples{
new_group(c("qhp_bus_rule_variables", "managed_care_share"))
new_group(x = c("qhp_bus_rule_variables", "managed_care_share"), q = TRUE)
}
