% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_method_request.R
\name{new_request}
\alias{new_request}
\title{Create a new request by class}
\usage{
new_request(x)
}
\arguments{
\item{x}{An object of class \code{care_endpoint}, \code{care_group}, \code{pro_endpoint}, or \code{open_endpoint}}
}
\value{
A new request
}
\description{
Create a new request by class
}
\examples{
care_endpoint("enrollees") |> new_request()
care_group("hospital") |> new_request()

pro_endpoint("PDC_affiliations") |> new_request()

open_endpoint("PROF_covered") |> new_request()

}
