% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_method_request.R
\name{new_request}
\alias{new_request}
\title{Create a new request by class}
\usage{
new_request(x)
}
\arguments{
\item{x}{An object of class \code{pro_endpoint}, \code{careMain}, or \code{open_endpoint}}
}
\value{
A new request
}
\description{
Create a new request by class
}
\examples{
careMain("enrollees") |> new_request()
pro_endpoint("PDC_affiliations") |> new_request()
open_endpoint("PROF_covered") |> new_request()
careGroup("hospital") |> new_request()
}
