% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_method_request.R
\name{new_request}
\alias{new_request}
\title{Create a new request by class}
\usage{
new_request(x)
}
\arguments{
\item{x}{An object of class \code{proMain}, \code{careMain}, or \code{openMain}}
}
\value{
A new request
}
\description{
Create a new request by class
}
\examples{
careMain("enrollees") |> new_request()
proMain("affiliations") |> new_request()
openMain("dashboard") |> new_request()

}
