% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_method_request.R
\name{new_request}
\alias{new_request}
\title{Create a new request by class}
\usage{
new_request(x, ...)
}
\arguments{
\item{x}{An object of class \code{proCurr}, \code{MainCurrent}, or \code{openCurr}}

\item{...}{Additional arguments?}
}
\value{
A new request
}
\description{
Create a new request by class
}
\examples{
MainCurrent("enrollees") |> new_request()
proCurr("affiliations") |> new_request()
openCurr("dashboard") |> new_request()

}
