% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_nppes.R
\name{npi_nlm}
\alias{npi_nlm}
\alias{npi_nppes}
\title{Search the NLM NPI Registry}
\source{
\href{https://clinicaltables.nlm.nih.gov/apidoc/npi_idv/v3/doc.html}{API Documentation: Individuals}

\href{https://clinicaltables.nlm.nih.gov/apidoc/npi_org/v3/doc.html}{API Documentation: Organizations}

\href{https://npiregistry.cms.hhs.gov/api-page}{API Documentation}
}
\usage{
npi_nlm(terms, npi = NULL)

npi_nppes(
  npi = NULL,
  entity = NULL,
  first = NULL,
  last = NULL,
  organization = NULL,
  name_type = NULL,
  taxonomy_desc = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  country = NULL
)
}
\arguments{
\item{terms}{\verb{<chr>} Search terms, separated by spaces}

\item{npi}{\verb{<chr>} Unique 10-digit National Provider Identifier number issued by CMS to US healthcare providers through NPPES.}

\item{entity}{\verb{<chr>} Entity type; one of either I for Individual (NPI-1) or O for Organizational (NPI-2)}

\item{first, last}{\verb{<chr>} Individual provider's name}

\item{organization}{\verb{<chr>} Organizational provider's name}

\item{name_type}{\verb{<chr>} Type of individual the first and last name parameters refer to; one of either AO for Authorized Officials or Provider for Individual Providers.}

\item{taxonomy_desc}{\verb{<chr>} Provider's taxonomy description, e.g. Pharmacist, Pediatrics}

\item{city}{\verb{<chr>} City name. For military addresses, search for either APO or FPO.}

\item{state}{\verb{<chr>} 2-character state abbreviation. If it is the only input, one other parameter besides entype and country is required.}

\item{zip}{\verb{<chr>} WC 5- to 9-digit zip code, without a hyphen.}

\item{country}{\verb{<chr>} 2-character country abbreviation. Can be the only input, as long as it is not US.}
}
\value{
\verb{<tibble>} of search results

\verb{<tibble>} of search results
}
\description{
Search the NLM NPI Registry

Search the NPPES NPI Registry
}
\examples{
npi_nlm("john bethesda")
npi_nlm("Wiregrass Georgia")
npi_nlm("Dentist Valdosta")
npi_nlm("Valdosta")
npi_nlm("Atlanta")
npi_nppes(npi = npi_ex$k[1:2]) |> str()
npi_nppes(npi = npi_ex$k[1]) |> str()
}
