% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{offset}
\alias{offset}
\alias{offset_sequence}
\alias{offset_length}
\alias{seq_along0}
\title{Generate API Request "Offset" Sequence}
\usage{
offset_sequence(n, limit, start = 0)

offset_length(n, limit, start = 0)

seq_along0(x)
}
\arguments{
\item{n}{\verb{<int>} Number of results returned in an API request}

\item{limit}{\verb{<int>} API rate limit, i.e. the maximum number of results an
API will return per request.}

\item{start}{\verb{<int>} Offset start; either \code{0} (default) or \code{limit}}
}
\value{
\verb{<int>} If \code{n <= limit}, simply returns \code{n}. If \code{n > limit}, an
integer sequence is returned, beginning at \code{start} and of
length equal to \code{n / limit}.
}
\description{
Generate API Request "Offset" Sequence
}
\examples{
offset_sequence(100, 10)
offset_length(100, 10)

offset_sequence(10, 100)
offset_length(10, 100)

offset_sequence(47984, 5000)
offset_length(47984, 5000)

offset_sequence(147984, 2000)
offset_length(147984, 2000)

}
\keyword{internal}
