% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{offset_sequence}
\alias{offset_sequence}
\title{Create Offset Sequence}
\usage{
offset_sequence(rows, size = 5000)
}
\arguments{
\item{rows}{\verb{<int>} The number of rows found in the dataset}

\item{size}{\verb{<int>} The size or limit of the offset sequence}
}
\value{
\verb{<int>} vector, length equal to rows / size, as sequence begins
at zero. If \code{rows} is less than or equal to \code{size}, the
function simply returns \code{rows}.
}
\description{
Create Offset Sequence
}
\examples{
offset_sequence(rows = 100, size = 10)

offset_sequence(rows = 47984, size = 5000)

}
