% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_backend.R
\name{openMain}
\alias{openMain}
\alias{openTemp}
\alias{openTempGroup}
\title{Current Open Payments API Endpoint}
\usage{
openMain(alias)

openTemp(alias)

openTempGroup(
  title = character(0),
  description = character(0),
  contact = character(0),
  rows = integer(0),
  pages = integer(0),
  fields = character(0),
  years = integer(0),
  endpoints = list()
)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint alias}

\item{title}{\verb{<chr>} title}

\item{description}{\verb{<chr>} description}

\item{contact}{\verb{<chr>} contact name and email address}

\item{rows}{\verb{<int>} number of rows}

\item{pages}{\verb{<int>} number of pages}

\item{fields}{\verb{<chr>} field names}

\item{years}{\verb{<int>} years available}

\item{endpoints}{\verb{<data.frame>} endpoints}
}
\value{
An S7 \verb{<openCurr>} object.

An S7 \verb{<openTemp>} object.

An S7 \verb{<openTempGroup>} object.
}
\description{
Current Open Payments API Endpoint

Temporal Open Payments API Endpoint

Open Payments API Endpoint Temporal Group
}
\examples{
openMain("prof_cov")
openMain("prof_phys")
openMain("prof_info")
openMain("prof_map")
openMain("prof_entity")
openMain("prof_teach")
openMain("dashboard")
openMain("pay_state_total")
openMain("pay_state_group")
openMain("pay_nat_group")
openMain("pay_nat_total")
openTemp("general")
openTemp("ownership")
openTemp("research")
# openTemp("recipient_nature")
# openTemp("recipient_entity")
# openTemp("entity_nature")
# openTemp("entity_recipient_nature")
# openTemp("state_nature")
}
