% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_open_payments.R
\name{open_dataset}
\alias{open_dataset}
\title{Load Open Payments API \code{Dataset}}
\usage{
open_dataset(period = NULL, group = NULL, dataset = NULL)
}
\arguments{
\item{period}{\verb{<chr>} dataset year}

\item{group}{\verb{<chr>} dataset group keyword}

\item{dataset}{\verb{<chr>} dataset title}
}
\value{
\verb{<Dataset>} object
}
\description{
Load Open Payments API \code{Dataset}
}
\examples{
open_dataset(period  = "All",
             group   = "Summary",
             dataset = "Covered Recipient Profile Supplement")
}
