% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_open_payments.R
\name{open_grouped}
\alias{open_grouped}
\title{Load Open Payments Grouped}
\usage{
open_grouped(x)
}
\arguments{
\item{x}{\verb{<list>} of Open Payments API catalog information}
}
\value{
\verb{<list>} of Open Payments API grouped information
}
\description{
Load Open Payments Grouped
}
\examples{
open_grouped(open_catalog())
}
