% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_open.R
\name{open_nrows_fields}
\alias{open_nrows_fields}
\title{Open Payments Endpoint Fields, Rows, Pages}
\usage{
open_nrows_fields(uuid)
}
\arguments{
\item{uuid}{\verb{<chr>} endpoint UUID}
}
\value{
\verb{<list>} number of rows and field names
}
\description{
Open Payments Endpoint Fields, Rows, Pages
}
\keyword{internal}
