% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_open_payments.R
\name{open_ownership}
\alias{open_ownership}
\title{Load Open Payments Ownership Data}
\usage{
open_ownership()
}
\value{
\verb{<list>} of Open Payments API ownership data
}
\description{
Load Open Payments Ownership Data
}
\examples{
open_ownership()
}
