% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_open.R
\name{open_profiles}
\alias{open_profiles}
\title{Load Open Payments Profile Sets}
\usage{
open_profiles()
}
\value{
\verb{<list>} of Open Payments API profile information
}
\description{
Load Open Payments Profile Sets
}
\examples{
open_profiles()
}
