% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_open_payments.R
\name{open_research}
\alias{open_research}
\title{Load Open Payments Research Data}
\usage{
open_research(x)
}
\arguments{
\item{x}{\verb{<list>} of Open Payments API catalog information}
}
\value{
\verb{<list>} of Open Payments API research data
}
\description{
Load Open Payments Research Data
}
\examples{
open_research(open_catalog())
}
