% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_summary.R
\name{open_summary}
\alias{open_summary}
\alias{open_dashboard}
\alias{open_national}
\title{Open Payments Summaries}
\usage{
open_dashboard()

open_national()
}
\value{
A \verb{<tibble>}
}
\description{
Open Payments Summaries

Open Payments Dashboard

Open Payments National Overall Total and Averages
}
\examples{
open_dashboard()
open_national()
}
