% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_open.R
\name{openMain}
\alias{openMain}
\title{Open Payments Endpoint}
\usage{
openMain(alias)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint alias}
}
\value{
An S7 \verb{<openMain>} object.
}
\description{
Open Payments Endpoint
}
\examples{
openMain("PROF_covered")
openMain("PROF_physician")
openMain("PROF_information")
openMain("PROF_mapping")
openMain("PROF_entity")
openMain("PROF_teaching")
#openMain("SUMM_dashboard")
openMain("SUMM_state_all")
openMain("SUMM_state_group")
openMain("SUMM_nation_all")
openMain("SUMM_nation_group")
}
