% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-alias_open.R
\name{openpayments}
\alias{openpayments}
\alias{open_endpoint}
\alias{open_temporal}
\alias{open_group}
\title{Open Payments API Endpoints}
\usage{
open_endpoint(alias, call = caller_env())

open_temporal(alias, call = caller_env())

open_group(alias, call = caller_env(), ...)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint or group alias}

\item{call}{\verb{<env>} environment to use for error reporting}

\item{...}{Additional arguments passed to the group constructor}
}
\value{
An S7 \verb{<class_endpoint>}, \verb{<class_temporal>} or \verb{<class_group>} object
}
\description{
Open Payments API Endpoints
}
\examples{
open_endpoint("profile_covered")
open_temporal("grouped_state_nature")
open_group("payment_detailed")
}
