% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_open.R
\name{open_endpoint}
\alias{open_endpoint}
\alias{open_group}
\alias{open_temporal}
\alias{open_troup}
\title{Open Payments Endpoint}
\usage{
open_endpoint(alias)

open_group(alias)

open_temporal(alias)

open_troup(alias)
}
\arguments{
\item{alias}{\verb{<chr>} title alias}
}
\value{
An S7 \verb{<open_endpoint>} object.

An S7 \verb{<open_group>} object.

An S7 \verb{<open_temporal>} object.

An S7 \verb{<open_troup>} object.
}
\description{
Open Payments Endpoint

Open Payments Endpoint Group

Open Payments Temporal Endpoint

Group of Open Payments Temporal Endpoints
}
\examples{
open_endpoint("PROF_covered")
open_endpoint("PROF_physician")
open_endpoint("PROF_information")
open_endpoint("PROF_mapping")
open_endpoint("PROF_entity")
open_endpoint("PROF_teaching")
open_endpoint("SUMM_dashboard")
open_endpoint("SUMM_state_all")
open_endpoint("SUMM_state_group")
open_endpoint("SUMM_nation_all")
open_endpoint("SUMM_nation_group")
open_group("profile")
open_temporal("DATA_general")
open_temporal("DATA_ownership")
open_temporal("DATA_research")
open_temporal("GROUP_recip_nature")
open_temporal("GROUP_recip_entity")
open_temporal("GROUP_entity_nature")
open_temporal("GROUP_all")
open_troup("grouped_payment")
open_troup("detailed_payment")
}
