% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_open.R
\name{openMain}
\alias{openMain}
\alias{openGroup}
\alias{openTemp}
\alias{openTempGroup}
\title{Open Payments Endpoint}
\usage{
openMain(alias)

openGroup(alias)

openTemp(alias)

openTempGroup(alias)
}
\arguments{
\item{alias}{\verb{<chr>} title alias}
}
\value{
An S7 \verb{<openMain>} object.

An S7 \verb{<openGroup>} object.

An S7 \verb{<openTemp>} object.

An S7 \verb{<openTempGroup>} object.
}
\description{
Open Payments Endpoint

Open Payments Endpoint Group

Open Payments Temporal Endpoint

Group of Open Payments Temporal Endpoints
}
\examples{
openMain("PROF_covered")
openMain("PROF_physician")
openMain("PROF_information")
openMain("PROF_mapping")
openMain("PROF_entity")
openMain("PROF_teaching")
openMain("SUMM_dashboard")
openMain("SUMM_state_all")
openMain("SUMM_state_group")
openMain("SUMM_nation_all")
openMain("SUMM_nation_group")
openGroup("profile")
openTemp("DATA_general")
openTemp("DATA_ownership")
openTemp("DATA_research")
openTemp("GROUP_recip_nature")
openTemp("GROUP_recip_entity")
openTemp("GROUP_entity_nature")
openTemp("GROUP_all")
openTempGroup("grouped_payment")
openTempGroup("detailed_payment")
}
