% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_open.R
\name{openpayments}
\alias{openpayments}
\alias{open_endpoint}
\alias{open_group}
\alias{open_temporal}
\alias{open_troup}
\title{Open Payments API Endpoint Classes}
\usage{
open_endpoint(alias)

open_group(alias)

open_temporal(alias)

open_troup(alias)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint alias}
}
\value{
An S7 \verb{<open_endpoint>}, \verb{<open_group>}, \verb{<open_temporal>}, or \verb{<open_troup>} object
}
\description{
Open Payments API Endpoint Classes
}
\examples{
open_endpoint("PROF_covered")
open_group("profile")
open_temporal("GROUP_all")
open_troup("grouped_payment")
}
