% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr2_utils.R
\name{parse_json_response_public}
\alias{parse_json_response_public}
\title{Parse JSON response}
\usage{
parse_json_response_public(response)
}
\arguments{
\item{response}{\verb{<httr_response>} API response}
}
\value{
\verb{<tibble>} Parsed JSON response as tibble
}
\description{
Parse JSON response
}
\keyword{internal}
