% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_request.R
\name{perform_request_public}
\alias{perform_request_public}
\title{Perform API request}
\usage{
perform_request_public(url, query, limit)
}
\arguments{
\item{url}{\verb{<url>} API identifier url}

\item{query}{\verb{<chr>} vector of query parameters}

\item{limit}{\verb{<int>} API rate limit}
}
\value{
\verb{<int>} Number of results
}
\description{
Perform API request
}
\keyword{internal}
