% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_httr2.R
\name{perform_simple}
\alias{perform_simple}
\title{Perform and Parse Simple JSON Response}
\usage{
perform_simple(x)
}
\arguments{
\item{x}{\verb{<httr2_request>} object}
}
\value{
\verb{<list>} Parsed JSON response
}
\description{
Perform and Parse Simple JSON Response
}
\keyword{internal}
