% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_provider.R
\name{prov_nrows_fields}
\alias{prov_nrows_fields}
\title{Get Field Names and Number of Rows from Provider Endpoint}
\usage{
prov_nrows_fields(url)
}
\arguments{
\item{url}{\verb{<chr>} Endpoint UUID}
}
\value{
\verb{<list>} of number of rows and field names
}
\description{
Get Field Names and Number of Rows from Provider Endpoint
}
\keyword{internal}
