% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_backend.R
\name{proMain}
\alias{proMain}
\title{Current Provider API Endpoint}
\usage{
proMain(alias)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint alias}
}
\value{
An S7 \verb{<proMain>} object.
}
\description{
Current Provider API Endpoint
}
\examples{
proMain("affiliations")
proMain("clinicians")
proMain("utilization")
}
