% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-alias_pro.R, R/pro_affiliations.R,
%   R/pro_clinicians.R, R/pro_utilization.R
\name{provider}
\alias{provider}
\alias{prov_endpoint}
\alias{prov_group}
\alias{affiliations}
\alias{clinicians}
\alias{utilization}
\title{Provider API Endpoints}
\usage{
prov_endpoint(alias, call = caller_env())

prov_group(alias, call = caller_env(), ...)

affiliations(
  npi = NULL,
  pac = NULL,
  last_name = NULL,
  first_name = NULL,
  middle_name = NULL,
  suffix = NULL,
  facility_type = NULL,
  ccn_facility = NULL,
  ccn_primary = NULL
)

clinicians(
  npi = NULL,
  pac = NULL,
  enid = NULL,
  last_name = NULL,
  first_name = NULL,
  middle_name = NULL,
  suffix = NULL,
  gender = NULL,
  credential = NULL,
  med_school = NULL,
  grad_year = NULL,
  spec_prim = NULL,
  spec_sec = NULL,
  telehealth = NULL,
  pac_org = NULL,
  org_n_members = NULL,
  facility_name = NULL,
  add1 = NULL,
  add2 = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  phone = NULL,
  ind_assign = NULL,
  grp_assign = NULL,
  add_id = NULL
)

utilization(
  npi = NULL,
  pac = NULL,
  last_name = NULL,
  first_name = NULL,
  middle_name = NULL,
  suffix = NULL,
  procedure = NULL,
  count = NULL,
  percentile = NULL
)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint or group alias}

\item{call}{\verb{<env>} environment to use for error reporting}

\item{...}{Additional arguments passed to the group constructor}

\item{npi}{\verb{<chr>} 10-digit Individual NPI}

\item{pac}{\verb{<chr>} 10-digit PECOS Associate Control (PAC) ID}

\item{first_name, middle_name, last_name, suffix}{\verb{<chr>} Individual provider's name}

\item{facility_type}{\verb{<chr>} Type of facility, one of the following:
\itemize{
\item \code{"Hospital"}
\item \code{"Long-term care hospital"}
\item \code{"Nursing home"}
\item \code{"Inpatient rehabilitation facility"}
\item \code{"Home health agency"}
\item \code{"Skilled nursing facility"}
\item \code{"Hospice"}
\item \code{"Dialysis facility"}
}}

\item{ccn_facility}{\verb{<chr>} 6-digit CMS Certification Number (CCN)
of a facility or unit within hospital where an individual provider
provides service.}

\item{ccn_primary}{\verb{<int>} 6-digit CMS Certification Number (CCN)
of a sub-unit's primary hospital, should the provider provide
services in said unit.}

\item{enid}{\verb{<chr>} 15-digit Medicare Enrollment ID}

\item{gender}{Individual provider's gender}

\item{credential}{Individual provider's credentials}

\item{med_school}{Medical school attended}

\item{grad_year}{Year of graduation from medical school}

\item{spec_prim}{Primary specialty}

\item{spec_sec}{Secondary specialty}

\item{telehealth}{Telehealth services}

\item{pac_org}{\verb{<chr>} 10-digit Organization PECOS Associate Control (PAC) ID}

\item{org_n_members}{Number of members in the organization}

\item{facility_name}{Facility name}

\item{add1}{Address line 1}

\item{add2}{Address line 2}

\item{city}{City}

\item{state}{State}

\item{zip}{Zip code}

\item{phone}{Phone number}

\item{ind_assign}{Individual assignment}

\item{grp_assign}{Group assignment}

\item{add_id}{Address ID}

\item{procedure}{\verb{<chr>} Procedure category}

\item{count}{\verb{<int>} Number of procedures performed}

\item{percentile}{\verb{<int>} Percentile of procedures performed}
}
\value{
An S7 \verb{<class_endpoint>} or \verb{<class_group>} object

\verb{<tibble>} of search results

\verb{<tibble>} of search results

\verb{<tibble>} of search results
}
\description{
Facility affiliations data publicly reported in the Provider Data Catalog.

The Doctors and Clinicians national downloadable file is organized such
that each line is unique at the clinician/enrollment record/group/address
level. Clinicians with multiple Medicare enrollment records and/or single
enrollments linking to multiple practice locations are listed on
multiple lines.

The Doctors and Clinicians utilization data file reports volume
information for procedures of interest on clinician profile pages and
in the provider data catalog (PDC) to inform patients and caregivers
about clinicians' experience.
}
\section{Periodicity}{

\itemize{
\item Monthly (R/P1M)
}


\itemize{
\item Monthly (R/P1M)
}
}

\examples{
\dontshow{if (rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
prov_endpoint("asc_facility")
prov_group("pro_dialysis")
\dontshow{\}) # examplesIf}
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
affiliations(facility_type = "Home health agency")
affiliations(last_name = "CURRY")
affiliations(ccn_primary = "670055")
affiliations(ccn_facility = "370781")
affiliations(ccn_facility = "331302")
affiliations(ccn_facility = "33Z302")
affiliations(npi = "1043245657")
\dontshow{\}) # examplesIf}
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
clinicians(last_name = "CURRY")
clinicians(npi = "1043245657")
clinicians(state = "GA")
\dontshow{\}) # examplesIf}
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
utilization(last_name = "CURRY")
utilization(npi = "1043245657")
utilization(npi = "1003001785")
\dontshow{\}) # examplesIf}
}
