% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_backend.R
\name{proCurr}
\alias{proCurr}
\title{Current Provider API Endpoint}
\usage{
proCurr(alias)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint alias}
}
\value{
An S7 \verb{<proCurr>} object.
}
\description{
Current Provider API Endpoint
}
\examples{
proCurr("affiliations")
proCurr("clinicians")
proCurr("utilization")
}
