% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_backend.R
\name{ProviderCurrent}
\alias{ProviderCurrent}
\title{Current Provider API Endpoint}
\usage{
ProviderCurrent(alias)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint alias}
}
\value{
An S7 \verb{<ProviderCurrent>} object.
}
\description{
Current Provider API Endpoint
}
\examples{
ProviderCurrent("affiliations")
ProviderCurrent("clinicians")
ProviderCurrent("utilization")

}
