% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_pro.R
\name{provider}
\alias{provider}
\alias{pro_endpoint}
\alias{pro_group}
\title{Provider API Endpoint Classes}
\usage{
pro_endpoint(alias)

pro_group(alias)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint alias}
}
\value{
An S7 \verb{<pro_endpoint>} or\verb{<pro_group>} object
}
\description{
Provider API Endpoint Classes
}
\examples{
pro_endpoint("PSI90_6digit")
pro_endpoint("joint_replace")
pro_endpoint("footnote_xwalk")
pro_group("PDC")
pro_group("MIPS")
}
