% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_pro.R
\name{provider}
\alias{provider}
\alias{pro_endpoint}
\title{Provider API Endpoint Class}
\usage{
pro_endpoint(alias)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint alias}
}
\value{
An S7 \verb{<pro_endpoint>}
}
\description{
Provider API Endpoint Class
}
\examples{
pro_endpoint("asc_facility")
}
