% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_pro.R
\name{proMain}
\alias{proMain}
\title{Provider Endpoint}
\usage{
proMain(alias)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint alias}
}
\value{
An S7 \verb{<proMain>} object.
}
\description{
Provider Endpoint
}
\examples{
proMain("affiliations") |> str()
proMain("clinicians") |> str()
proMain("utilization") |> str()
}
