% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_provider.R
\name{provider_current}
\alias{provider_current}
\title{Load \code{CurrentProvider} API Endpoint}
\usage{
provider_current(alias)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint alias}
}
\value{
\verb{<CurrentProvider>} object
}
\description{
Load \code{CurrentProvider} API Endpoint
}
\examples{
provider_current("affiliations")
provider_current("clinicians")
provider_current("utilization")
provider_current("group_mips")
provider_current("group_patient")
provider_current("clin_mips")
provider_current("clin_overall")
provider_current("vgroup_mips")
}
