% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_Dataset_Distribution.R
\name{provider_Dataset}
\alias{provider_Dataset}
\title{Load Provider API \code{Dataset}}
\usage{
provider_Dataset(dataset, fname = TRUE)
}
\arguments{
\item{dataset}{\verb{<chr>} dataset title}

\item{fname}{\verb{<lgl>} Is \code{dataset} a function name?; default is \code{TRUE}}
}
\value{
\verb{<Dataset>} object
}
\description{
Load Provider API \code{Dataset}
}
\examples{
provider_Dataset("affiliations")

provider_Dataset("clinicians")

}
