% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_provider.R
\name{provider_dataset}
\alias{provider_dataset}
\title{Load \code{CurrentProvider} API Endpoint}
\usage{
provider_dataset(alias)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint alias}
}
\value{
\verb{<CurrentProvider>} object
}
\description{
Load \code{CurrentProvider} API Endpoint
}
\examples{
provider_dataset("affiliations")

provider_dataset("clinicians")

}
