% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{provider_dataset}
\alias{provider_dataset}
\title{Provider CMS Dataset Object}
\usage{
provider_dataset()
}
\value{
\verb{<list>} of \verb{<tibbles>}: \code{dataset}, \code{distribution}
}
\description{
Provider CMS Dataset Object
}
\examples{
provider_dataset()

}
