% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_provider.R
\name{provider_dataset}
\alias{provider_dataset}
\title{Load Provider API \code{Dataset}}
\usage{
provider_dataset(dataset)
}
\arguments{
\item{dataset}{\verb{<chr>} dataset title}
}
\value{
\verb{<Dataset>} object
}
\description{
Load Provider API \code{Dataset}
}
\examples{
provider_dataset("affiliations")

provider_dataset("clinicians")

}
