% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/providertwo-package.R
\docType{package}
\name{providertwo-package}
\alias{providertwo}
\alias{providertwo-package}
\title{providertwo: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andrewallenbruce/providertwo}
  \item \url{https://andrewallenbruce.github.io/providertwo/}
  \item Report bugs at \url{https://github.com/andrewallenbruce/providertwo/issues}
}

}
\author{
\strong{Maintainer}: Andrew Allen Bruce \email{andrewallenbruce@gmail.com} [copyright holder]

}
\keyword{internal}
