% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_class.R
\name{public_Dataset}
\alias{public_Dataset}
\title{Populate a public_Dataset class object}
\usage{
public_Dataset(dataset)
}
\arguments{
\item{dataset}{\verb{<chr>} dataset title}
}
\value{
\verb{<S7_class>} object
}
\description{
Populate a public_Dataset class object
}
\examples{
public_Dataset(dataset = "Public Provider Enrollment")
public_Dataset(dataset = "Hospital Enrollments")

}
