% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrollees.R
\name{public_Dataset}
\alias{public_Dataset}
\title{Public Provider Enrollment Dataset Object}
\usage{
public_Dataset(dataset)
}
\arguments{
\item{dataset}{\verb{<chr>} dataset title}
}
\value{
\verb{<S7_class>} object
}
\description{
Public Provider Enrollment Dataset Object
}
\examples{
public_Dataset(dataset = "Public Provider Enrollment")
public_Dataset(dataset = "Hospital Enrollments")

}
