% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_main.R
\name{public_dataset}
\alias{public_dataset}
\title{Load Public API \code{Dataset}}
\usage{
public_dataset(dataset = "enrollees", fname = TRUE)
}
\arguments{
\item{dataset}{\verb{<chr>} dataset title}

\item{fname}{\verb{<lgl>} Is \code{dataset} a function name?; default is \code{TRUE}}
}
\value{
\verb{<Dataset>} object
}
\description{
Load Public API \code{Dataset}
}
\examples{
public_dataset("enrollees")

public_dataset("hospitals")

public_dataset("reassignments")

public_dataset("opt_out")

public_dataset("laboratories")

}
