% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_class.R
\name{public_Dataset}
\alias{public_Dataset}
\title{Load Public API \code{Dataset}}
\usage{
public_Dataset(dataset)
}
\arguments{
\item{dataset}{\verb{<chr>} dataset title}
}
\value{
\verb{<Dataset>} object
}
\description{
Load Public API \code{Dataset}
}
\examples{
public_Dataset("Public Provider Enrollment")

public_Dataset("Hospital Enrollments")

}
