% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_class.R
\name{public_Distribution}
\alias{public_Distribution}
\title{Populate a public_Distribution class object}
\usage{
public_Distribution(dataset)
}
\arguments{
\item{dataset}{\verb{<chr>} dataset title}
}
\value{
\verb{<S7_class>} object
}
\description{
Populate a public_Distribution class object
}
\examples{
public_Distribution(dataset = "Public Provider Enrollment")

}
