% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_public.R
\name{public_Distribution}
\alias{public_Distribution}
\title{Load Public API \code{Distribution}}
\usage{
public_Distribution(dataset, fname = TRUE)
}
\arguments{
\item{dataset}{\verb{<chr>} dataset title}

\item{fname}{\verb{<lgl>} Is \code{dataset} a function name?; default is \code{TRUE}}
}
\value{
\verb{<Distribution>} object
}
\description{
Load Public API \code{Distribution}
}
\examples{
public_Distribution("Public Provider Enrollment", FALSE)

public_Distribution("Quality Payment Program Experience", FALSE)

}
