% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{Publisher}
\alias{Publisher}
\title{Publisher Class}
\usage{
Publisher(
  type = "org:Organization",
  name = "Centers for Medicare & Medicaid Services"
)
}
\arguments{
\item{type}{\verb{<chr>} Publisher type; default is \code{org:Organization}}

\item{name}{\verb{<chr>} Publisher name; default is \verb{Centers for Medicare & Medicaid Services}}
}
\value{
\verb{<S7_class>} Publisher object
}
\description{
Publisher Class
}
