% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{Publisher}
\alias{Publisher}
\title{Publisher Class}
\usage{
Publisher(
  type = "org:Organization",
  name = "Centers for Medicare & Medicaid Services (CMS)"
)
}
\arguments{
\item{type}{\verb{<chr>} Schema publisher type; default is \code{org:Organization}}

\item{name}{\verb{<chr>} publisher name; default is \verb{Centers for Medicare & Medicaid Services}}
}
\value{
\verb{<S7_class>} dataset_Publisher object
}
\description{
\code{Publisher} object
}
