% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpp.R
\name{quality_payment}
\alias{quality_payment}
\alias{quality_metrics}
\alias{quality_eligibility}
\title{Quality Payment Program}
\usage{
quality_metrics(year)

quality_eligibility(year, npi)
}
\arguments{
\item{year}{A vector of years from 2018 to 2025}

\item{npi}{A vector of NPIs}
}
\description{
Quality Payment Program
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
qpp_metrics <- quality_metrics(year = 2018:2025)

qpp_eligible <- quality_eligibility(
  year = 2018:2024,
  npi = c(1144544834, 1043477615, 1932365699, 1225701881))

qpp_experience <- build(
  endpoint("qppe"),
  query(npi = any_of(1144544834, 1043477615, 1932365699, 1225701881)))


qpp_exp <- qpp_experience@string |>
  providertwo:::gremove("/stats") |>
  providertwo:::greplace("size=1", "size=5000") |>
  providertwo:::map_perform_parallel() |>
  providertwo:::set_clean(qpp_experience@year) |>
  purrr::list_rbind(names_to = "prog_year") |>
  providertwo:::map_na_if() |>
  fastplyr::as_tbl()

list(
  experience  = providertwo:::set_clean(qpp_exp, names(qpp_exp)),
  eligibility = qpp_eligible,
  metrics     = qpp_metrics)
\dontshow{\}) # examplesIf}
}
