% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_qpp.R
\name{quality_payment}
\alias{quality_payment}
\alias{qpp_stats}
\alias{qpp_elig}
\title{Quality Payment Program (QPP)}
\usage{
qpp_stats(year)

qpp_elig(npi, year)
}
\arguments{
\item{year}{A vector of years from 2018 to 2025}

\item{npi}{A vector of NPIs}
}
\description{
Quality Payment Program (QPP)
}
\examples{
qpp_stats(year = 2018:2025)

qpp_elig(npi = 1043477615, year = 2018)
qpp_elig(npi = 1144544834, year = 2025)
qpp_elig(npi = 1932365699, year = 2024)

# CORRECT FORM: eligibility/npis/1043477615/1144544834?year=2018
# Multiple NPIs, One year at a time
qpp_elig(npi = c(1144544834, 1043477615, 1932365699), year = 2024)
}
