% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{new_query}
\alias{new_query}
\title{Format Queries}
\usage{
new_query(args)
}
\arguments{
\item{args}{named \verb{<list>} of \verb{<chr>} arguments}
}
\value{
\verb{<list>} of formatted query \verb{<exprs>}
}
\description{
Format Queries
}
\examples{
x = list(state = c("GA", "MD"),
            last_name = "SMITH",
            npi = 1234567890,
            PECOS = NULL)

new_query(x)

}
