% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{seq_along0}
\alias{seq_along0}
\alias{process_params}
\alias{eval_params}
\alias{format_query_public}
\alias{format_query_provider}
\alias{query-helpers}
\alias{greater_than}
\alias{less_than}
\alias{starts_with}
\alias{ends_with}
\alias{contains}
\alias{between}
\alias{is_in}
\title{\code{seq_along()} Starting at 0}
\usage{
seq_along0(x)

process_params(arg_names, field_names)

eval_params(args, fields)

format_query_public(args, operator = "=")

format_query_provider(args, operator = "=")

greater_than(x, equals = FALSE)

less_than(x, equals = FALSE)

starts_with(x)

ends_with(x)

contains(x)

between(x, negate = FALSE)

is_in(x, negate = FALSE)
}
\arguments{
\item{x}{\verb{<chr>}  input}

\item{arg_names}{\verb{<chr>} vector of argument names}

\item{field_names}{\verb{<chr>} vector of API field names}

\item{args}{named \verb{<list>} or vector of \verb{<chr>} arguments}

\item{fields}{\verb{<chr>} vector of API field names}

\item{operator}{\verb{<chr>} logical operator; acceptable options are:
\code{=}, \code{>=}, \code{<=}, \code{>}, \code{<}, \verb{<>}, \code{STARTS_WITH},
\code{ENDS_WITH}, \code{CONTAINS}, \code{IN}, \verb{NOT IN}, \code{BETWEEN},
\verb{NOT BETWEEN}, \verb{IS NULL}, \verb{IS NOT NULL}}

\item{equals}{\verb{<lgl>}  append "="; default is \code{FALSE}}

\item{negate}{\verb{<lgl>}  prepend "NOT"; default is \code{FALSE}}
}
\value{
\verb{<expr>} of processed query parameters

\verb{<list>} of evaluated query parameters

\verb{<list>} of formatted query \verb{<exprs>}

\verb{<list>} of formatted query \verb{<exprs>}

\if{html}{\out{<div class="sourceCode">}}\preformatted{ `<list>` of query parameters
}\if{html}{\out{</div>}}
}
\description{
\code{seq_along()} Starting at 0

Process Query Parameters

Evaluate Query Parameters

Format Public API Queries

Format Provider API Queries

Query Formatting Helpers
}
\examples{
format_query_public(list("NPI" = "1417918293", "PECOS" = NULL))

format_query_public(list(NPI = "1417918293", PECOS = "001132"))

format_query_provider(list("NPI" = "1417918293", "PECOS" = NULL))

format_query_provider(list(NPI = "1417918293", PECOS = "001132"))

\dontshow{if (rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
greater_than(1)
starts_with("foo")
is_in(state.abb[10:15])
\dontshow{\}) # examplesIf}
}
\keyword{internal}
