% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query-helpers}
\alias{query-helpers}
\alias{greater_than}
\alias{less_than}
\alias{starts_with}
\alias{ends_with}
\alias{contains}
\alias{between}
\alias{is_in}
\title{Query Formatting Helpers}
\usage{
greater_than(x, equals = FALSE)

less_than(x, equals = FALSE)

starts_with(x)

ends_with(x)

contains(x)

between(x, negate = FALSE)

is_in(x, negate = FALSE)
}
\arguments{
\item{x}{\verb{<chr>}  input}

\item{equals}{\verb{<lgl>}  append "="; default is \code{FALSE}}

\item{negate}{\verb{<lgl>}  prepend "NOT"; default is \code{FALSE}}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ `<list>` of query parameters
}\if{html}{\out{</div>}}
}
\description{
Query Formatting Helpers
}
\examples{
\dontshow{if (rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
greater_than(1)
starts_with("foo")
is_in(state.abb[10:15])
\dontshow{\}) # examplesIf}
}
\keyword{internal}
