% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query-helpers}
\alias{query-helpers}
\title{Query Formatting Helpers}
\arguments{
\item{x}{\verb{<chr>}  input}

\item{equals}{\verb{<lgl>}  append "="; default is \code{FALSE}}

\item{negate}{\verb{<lgl>}  prepend "NOT"; default is \code{FALSE}}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ `<list>` of query parameters
}\if{html}{\out{</div>}}
}
\description{
Query Formatting Helpers
}
\examples{
\dontshow{if (rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
greater_than(1)
starts_with("foo")
is_in(state.abb[10:15])
\dontshow{\}) # examplesIf}
}
