% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-helpers.R
\name{query-helpers}
\alias{query-helpers}
\alias{greaterthan}
\alias{greaterequal}
\alias{lessthan}
\alias{lessequal}
\alias{startswith}
\alias{endswith}
\alias{includes}
\alias{inside}
\alias{outside}
\alias{is_in}
\alias{not_in}
\title{Query Formatting Helpers}
\usage{
greaterthan(x)

greaterequal(x)

lessthan(x)

lessequal(x)

startswith(x)

endswith(x)

includes(x)

inside(x)

outside(x)

is_in(x)

not_in(x)
}
\arguments{
\item{x}{\verb{<chr>} function argument input}
}
\value{
\verb{<list>} of function argument and operator
}
\description{
Query Formatting Helpers
}
\examples{
greaterthan(1)

startswith("foo")

is_in(state.abb[10:15])
}
