% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\alias{query3}
\title{Create a Query Object}
\usage{
query(...)

query3(...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Named conditions}
}
\value{
S7 \verb{<class_query>} object.
}
\description{
Create a Query Object
}
\examples{
query(
  first_name  = starts_with("And"),
  middle_name = NULL,
  last_name   = contains("J"),
  state       = any_of("CA", "GA", "NY"),
  state_own   = c("GA", "MD"),
  ccn         = "01256",
  rate        = between(0.45, 0.67),
  year        = 2014:2025)
query3(
  state = any_of("CA", "GA", "NY"),
  state_owner = c("GA", "MD"),
  or("state", "state_owner"))

try(query3(
  first_name  = starts_with("And"),
  ccn         = "01256",
  and("ccn", "npii")))
}
