% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{Create a Query Object}
\usage{
query(..., .type = "default")
}
\arguments{
\item{...}{Query arguments. See details for valid query modifiers.}

\item{.type}{Query type, \code{"default"} or \code{"medicare"}.}
}
\value{
S7 \verb{<class_query>} object.
}
\description{
Create a Query Object
}
\examples{
query(
  first_name = starts_with_("Andr"),
  last_name = contains_("J"),
  state = in_(c("CA", "GA", "NY")),
  city = equals_(c("Atlanta", "Los Angeles"), negate = TRUE),
  state_owner = c("GA", "MD"),
  npi = npi_ex$k,
  ccn = "01256",
  pac = NULL,
  .type = "medicare")

query(
  first_name = starts_with_("Andr"),
  last_name = contains_("J"),
  state = in_(c("CA", "GA", "NY")),
  city = equals_(c("Atlanta", "Los Angeles"), negate = TRUE),
  state_own = c("GA", "MD"),
  npi = npi_ex$k,
  ccn = "01256",
  pac = NULL,
  .type = "default")
}
