% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{Create a Query Object}
\usage{
query(...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Named conditions where the names are API fields.}
}
\value{
S7 \verb{<class_query>} object.
}
\description{
Create a Query Object
}
\examples{
query(
  first_name  = starts_with("And"),
  middle_name = NULL,
  last_name   = contains("J"),
  state       = any_of(c("CA", "GA", "NY")),
  state_own   = c("GA", "MD"),
  npi         = npi_ex$k,
  ccn         = "01256",
  rate        = between(0.45, 0.67),
  year        = 2014:2025,
  group       = or(c("first_name", "last_name")))
}
