% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\alias{query2}
\title{Create a Query Object}
\usage{
query(...)

query2(..., call = rlang::caller_env())
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Key-value pairs of query parameters.
See Details.}

\item{call}{The environment from which the function is called.}
}
\value{
S7 \verb{<class_query>} object.
}
\description{
Create a Query Object
}
\examples{
query(
  first_name  = starts_with("And"),
  middle_name = NULL,
  last_name   = contains("J"),
  state       = any_of("CA", "GA", "NY"),
  state_own   = c("GA", "MD"),
  ccn         = "01256",
  rate        = between(0.45, 0.67),
  year        = 2014:2025)
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
try(query2())
try(query2(1))
try(query2(a = 1, a = 2))
try(query2(a = 1, or("a")))
try(query2(a = 1, or("a", "year")))
try(query2(year = none_of(2000:2020)))

query2(
  first_name  = starts_with("And"),
  middle_name = NULL,
  last_name   = contains("J"),
  state       = c("CA", "GA", "NY"),
  state_own   = c("GA", "MD"),
  ccn         = "01256",
  rate        = between(0.45, 0.67),
  year        = 2014:2025)

query2(
  year  = 2018:2025,
  state = c("CA", "GA", "NY"),
  state_owner = c("GA", "MD"))

query2(
  year = 2022:2024,
  state = c("GA", "NY"),
  enid_id = "I20040309000221",
  city = "Atlanta",
  provider_name = starts_with("C"),
  first_name = starts_with("An"),
  last_name = contains("JE"),
  practice_size = less_than(10, or_equal = TRUE))
\dontshow{\}) # examplesIf}
}
