% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_modifiers.R
\name{query_modifier}
\alias{query_modifier}
\alias{any_of}
\alias{none_of}
\alias{equals}
\alias{all_but}
\alias{between}
\alias{not_between}
\alias{greater_than}
\alias{less_than}
\alias{starts_with}
\alias{ends_with}
\alias{contains}
\alias{like}
\title{Query Modifiers}
\usage{
any_of(x, negate = FALSE)

none_of(x)

equals(x, negate = FALSE)

all_but(x)

between(x, y, negate = FALSE)

not_between(x, y)

greater_than(x, or_equal = FALSE)

less_than(x, or_equal = FALSE)

starts_with(x)

ends_with(x)

contains(x)

like(x)
}
\arguments{
\item{x, y}{input}

\item{negate}{\verb{<lgl>} prepend \code{NOT} to operator}

\item{or_equal}{\verb{<lgl>} append \code{=} to \code{>} or \code{<}}
}
\value{
An S7 \verb{<class_modifier>} object.
}
\description{
Helpers for use in constructing conditions in queries.
}
\details{
Query modifiers are a small DSL for use in constructing query conditions,
in the \href{https://www.drupal.org/docs/core-modules-and-themes/core-modules/jsonapi-module/filtering}{JSON-API} format.
}
\examples{
any_of(state.abb[10:15])
any_of(state.abb[1:5], negate = TRUE)
none_of(state.abb[10:15])
equals(1000)
equals(1000, negate = TRUE)
all_but(1000)
between(1000, 1100)
between(0.125, 2, negate = TRUE)
try(between(0.95, 0.67))
not_between(1000, 1100)
not_between(0.125, 2)
try(not_between(0.95, 0.67))
greater_than(1000)
greater_than(0.125, or_equal = TRUE)
less_than(1000)
less_than(0.125, or_equal = TRUE)
starts_with("foo")
ends_with("bar")
contains("baz")
like("baz")
}
