% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_modifiers.R
\name{query_modifier}
\alias{query_modifier}
\alias{greater_than_}
\alias{less_than_}
\alias{starts_with_}
\alias{ends_with_}
\alias{contains_}
\alias{in_}
\alias{equals_}
\alias{is_blank_}
\title{Query Modifiers}
\usage{
greater_than_(x, or_equals = FALSE)

less_than_(x, or_equals = FALSE)

starts_with_(x)

ends_with_(x)

contains_(x)

in_(x, negate = FALSE)

equals_(x, negate = FALSE)

is_blank_(negate = FALSE)
}
\arguments{
\item{x}{input}

\item{or_equals}{\verb{<lgl>} append "="; default is \code{FALSE}}

\item{negate}{\verb{<lgl>} prepend "NOT"; default is \code{FALSE}}
}
\value{
object of class \verb{<modifier>}
}
\description{
Helpers for use in constructing conditions in queries.
}
\details{
Query modifiers are a small DSL for use in constructing query conditions,
in the \strong{JSON:API} format.
}
\examples{
greater_than_(1000)
greater_than_(1000, or_equals = TRUE)
less_than_(1000)
less_than_(1000, or_equals = TRUE)
starts_with_("foo")
ends_with_("bar")
contains_("baz")
in_(state.abb[10:15])
in_(state.abb[10:15], negate = TRUE)
equals_(1000)
equals_(1000, negate = TRUE)
is_blank_()
is_blank_(negate = TRUE)
}
