% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_nresults.R
\name{query_nresults}
\alias{query_nresults}
\title{Return the number of results for a given query}
\usage{
query_nresults(obj, query = NULL, ...)
}
\arguments{
\item{obj}{An object of class \verb{<class_endpoint>} or \verb{<class_temporal>}.}

\item{query}{A \verb{<chr>} string representing the query to be passed to the API. If \code{NULL}, the default query for the endpoint is used.}

\item{...}{Additional arguments passed to the query.}
}
\value{
An \verb{<int>} representing the number of results found.
}
\description{
Return the number of results for a given query
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
query_nresults(
   endpoint("care_dial_end"),
   query(state = any_of(c("CA", "GA", "NY", "TX"))))

query_nresults(
   endpoint("quality_payment"),
   query(`practice state or us territory` = any_of(c("CA", "GA", "NY", "TX"))))
\dontshow{\}) # examplesIf}
}
