% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_request.R
\name{query_nrows_provider}
\alias{query_nrows_provider}
\title{Request number of results from Provider Query}
\usage{
query_nrows_provider(request)
}
\arguments{
\item{request}{\verb{<httr_request>} API request}
}
\value{
\verb{<int>} Number of results
}
\description{
Request number of results from Provider Query
}
\keyword{internal}
