% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_request.R
\name{query_nrows_public}
\alias{query_nrows_public}
\title{Request number of results for query}
\usage{
query_nrows_public(request)
}
\arguments{
\item{request}{\verb{<httr_request>} API request}
}
\value{
\verb{<int>} Number of results
}
\description{
Request number of results for query
}
\keyword{internal}
