% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_results.R
\name{query_results}
\alias{query_results}
\title{Return the number of results for a given query}
\usage{
query_results(obj, qry = NULL)
}
\arguments{
\item{obj}{A \verb{<class_endpoint>} or \verb{<class_temporal>} object.}

\item{qry}{A \verb{<class_query>} object. If \code{NULL} (the default), no query is used.}
}
\value{
A \verb{<list>} with the number of results found, total number of rows, and the URL used for the query.
}
\description{
Return the number of results for a given query
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
query_results(endpoint("dial_facility"), query(state = any_of(c("GA", "TX"))))
query_results(endpoint("lab_fee_sched"))

query_results(
   endpoint("quality_payment"),
   query(year = 2017:2023,
         `practice state or us territory` = any_of(c("GA", "TX"))))
\dontshow{\}) # examplesIf}
}
