% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_group.R
\name{quick}
\alias{quick}
\title{Quickly access CMS data}
\usage{
quick(alias, offset = 0, limit = 10000)
}
\arguments{
\item{alias}{\verb{<chr>} Alias representing the CMS data endpoint or category.}

\item{offset}{\verb{<int>} The offset for pagination. Default is \code{0}.}

\item{limit}{\verb{<int>} The maximum number of records to retrieve. Default is \code{10000}.}
}
\value{
A data frame containing the requested CMS data.
}
\description{
Convenience function to quickly access various CMS data endpoints.
Mostly for debugging purposes.
}
\examples{
quick("revalid_group")
quick("out_img_national")
}
