% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick.R
\name{quick}
\alias{quick}
\title{Quickly access CMS data}
\usage{
quick(x, offset = 0L, limit = 1L, call = caller_env())
}
\arguments{
\item{x}{A string representing the CMS data endpoint or category.}

\item{offset}{An integer representing the offset for pagination. Default is 0.}

\item{limit}{An integer representing the maximum number of records to retrieve. Default is 10,000.}

\item{call}{The calling environment. Default is the current environment.}
}
\value{
A data frame or object containing the requested CMS data.
}
\description{
\code{quick()} is a convenience function to quickly access various CMS data endpoints.
It provides a simple way to retrieve data from different CMS programs and categories.
}
\examples{
quick("MLR") |> str()
}
