% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick.R
\name{quick}
\alias{quick}
\title{Quickly access CMS data}
\usage{
quick(x, offset = 0L, limit = 1L, call = caller_env())
}
\arguments{
\item{x}{A string representing the CMS data endpoint or category.}

\item{offset}{An integer representing the offset for pagination. Default is 0.}

\item{limit}{An integer representing the maximum number of records to retrieve. Default is 10,000.}

\item{call}{The calling environment. Default is the current environment.}
}
\value{
A data frame or object containing the requested CMS data.
}
\description{
\code{quick()} is a convenience function to quickly access various CMS data endpoints.
It provides a simple way to retrieve data from different CMS programs and categories.
}
\examples{
# MEDICAID
quick("MLR") |> str()
quick("enterprise") |> str()
quick("demographics") |> str()


# PROVIDER
quick("suppliers") |> str()
quick("PDC") |> str()
quick("MIPS") |> str()
quick("LTCH") |> str()
quick("IRF") |> str()
quick("SPICE") |> str()
quick("CAHPS_hospice") |> str()
quick("HHVBP") |> str()
quick("HHCAHPS") |> str()
quick("HHCAHPS") |> str()
quick("SNF_VBP") |> str()
quick("SNF_quality") |> str()
quick("NH_pro") |> str()

# OPEN PAYMENTS
quick("summary") |> str()
quick("profile") |> str()

# MEDICARE
quick("HHA") |> str()
quick("hospice") |> str()
quick("hospital") |> str()
quick("RHC") |> str()
quick("FQHC") |> str()
quick("pending") |> str()
quick("reassignment") |> str()
quick("SNF") |> str()

quick("contact") |> str()
quick("crosswalk") |> str()
quick("CARE_dialysis") |> str()
quick("enrollees") |> str()
quick("facilities") |> str()
quick("IQIES") |> str()
quick("laboratories") |> str()
quick("long_term") |> str()
quick("opt_out") |> str()
quick("order_refer") |> str()
quick("RBCS") |> str()
quick("transparency") |> str()
}
