% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick.R
\name{quick}
\alias{quick}
\title{Quickly access CMS data}
\usage{
quick(x, offset = 0L, limit = 10L, call = caller_env())
}
\arguments{
\item{x}{\verb{<chr>} Alias representing the CMS data endpoint or category.}

\item{offset}{\verb{<int>} The offset for pagination. Default is \code{"0"}.}

\item{limit}{\verb{<int>} The maximum number of records to retrieve. Default is \code{"10"}.}

\item{call}{Call environment. Default is the current environment.}
}
\value{
A data frame or list containing the requested CMS data.
}
\description{
Convenience function to quickly access various CMS data endpoints.
Mostly for debugging purposes.
}
\examples{
quick("pro_dialysis") |> str()
quick("care_nhome")
}
