% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_quick_query.R
\name{quick_query_}
\alias{quick_query_}
\title{Quick Query}
\usage{
quick_query_(x, ...)
}
\arguments{
\item{x}{An S7 \verb{<endpoint>} object}

\item{...}{Additional arguments}
}
\value{
A list of results from the API query, or \code{NULL} if no results are found.
}
\description{
Quick Query
}
