% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_misc.R
\name{recode_iso8601}
\alias{recode_iso8601}
\title{ISO 8601 Recurring Time Intervals}
\source{
\href{https://resources.data.gov/resources/dcat-us/#accrualPeriodicity}{DCAT Schema: accrualPeriodicity}
}
\usage{
recode_iso8601(x)
}
\arguments{
\item{x}{\verb{<chr>} vector of ISO8601 recurrence rules}
}
\value{
\verb{<chr>} vector of human-readable recurrence rule descriptions
}
\description{
ISO 8601 Recurring Time Intervals
}
\section{References}{

\itemize{
\item \href{https://en.wikipedia.org/wiki/ISO_8601}{ISO 8601}
\item \href{https://en.wikipedia.org/wiki/ISO_8601#Repeating_intervals}{ISO 8601 Repeating_intervals}
\item \href{https://sentenz.github.io/convention/convention/iso-8601/#19-recurring-time-intervals}{Recurring Time Intervals}
}
}

\examples{
accrualPeriodicity = c(
  "R/PT1S",   "R/PT1H",  "R/P1D", "R/P3.5D",
  "R/P0.33W", "R/P0.5W", "R/P1W", "R/P2W",
  "R/P0.33M", "R/P0.5M", "R/P1M", "R/P2M",
  "R/P3M",    "R/P4M",   "R/P6M", "R/P1Y",
  "R/P2Y",    "R/P3Y",   "R/P4Y", "R/P10Y")

recode_iso8601(accrualPeriodicity)
}
\keyword{internal}
