% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{Resources}
\alias{Resources}
\title{Resources Class}
\usage{
Resources(url = NULL)
}
\arguments{
\item{url}{\verb{<chr>} \code{resourcesAPI} url}
}
\value{
\verb{<S7_class>} Resources object
}
\description{
\code{Resources} object
}
\examples{
# Resources(ex_resource_url)

}
