% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{Resources}
\alias{Resources}
\title{Resources Class}
\usage{
Resources(url = NULL)
}
\arguments{
\item{url}{\code{NULL} | \code{class_character} API identifier url}
}
\description{
The \code{Resources} class object contains validated API metadata
}
