% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{Resources}
\alias{Resources}
\title{Resources Class}
\usage{
Resources(url = NULL)
}
\arguments{
\item{url}{\verb{<chr>} resourcesAPI url}
}
\value{
\verb{<S7_class>} object
}
\description{
\code{Resources} object
}
\examples{
Resources(url = paste0(
   "https://data.cms.gov/",
   "data-api/v1/dataset-resources/",
   "7dcf9ea6-ee2f-4bf1-8b5d-39c18b0e8541"
   ))

}
