% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_httr2.R
\name{resp_simple_json}
\alias{resp_simple_json}
\title{Parse Simple JSON Response}
\usage{
resp_simple_json(resp, ...)
}
\arguments{
\item{resp}{\verb{<httr2_response>} object}

\item{...}{Additional arguments}
}
\value{
\verb{<list>} Parsed JSON response
}
\description{
Parse Simple JSON Response
}
\keyword{internal}
