% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_standardize.R
\name{standardize}
\alias{standardize}
\title{Standardize Query to Endpoint}
\usage{
standardize(obj, qry)
}
\arguments{
\item{obj}{An \verb{<endpoint>}, \verb{<collection>} or \verb{<group>} object.}

\item{qry}{A \verb{<query>} object.}
}
\value{
A list of query parameters matched to an endpoint's fields.
}
\description{
Standardize Query to Endpoint
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
standardize(
  endpoint("drug_state"),
  query2(
    year = 2022:2024,
    state = c("GA", "NY")
  )
)

standardize(
  endpoint("enroll_prov"),
  query2(
    enrlmt_id = "I20040309000221")
)

standardize(
  collection("dialysis"),
  query2(state = "AL")
)

standardize(
  endpoint("dial_facility"),
  query2(
    state = "GA",
    city = "Atlanta",
    provcity = "Atlanta",
    provider_name = starts_with("C"),
    provname = starts_with("C")
  )
)

standardize(
  endpoint("qppe"),
  query2(
    year = 2021:2023,
    practice_state_or_us_territory = c("GA", "FL"),
    practice_size = less_than(10, or_equal = TRUE),
    or("practice_state_or_us_territory",
       "practice_size")
  )
)

standardize(
  endpoint("hc_quality"),
  query2(
    state = c("Georgia", "Alabama"),
    year = 2020:2023
  )
)

standardize(
  endpoint("pdc_clinician"),
  query2(
    provider_first_name = starts_with("An"),
    provider_last_name = contains("JE"),
    state = c("CA", "GA", "NY")
  )
)
\dontshow{\}) # examplesIf}
}
