% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_backend.R
\name{Temporal}
\alias{Temporal}
\alias{TemporalMain}
\alias{TemporalMainGroup}
\alias{TemporalOpen}
\alias{TemporalOpenGroup}
\title{Temporal}
\usage{
Temporal(
  title = character(0),
  description = character(0),
  contact = character(0),
  modified = character(0),
  endpoints = list()
)

TemporalMain(
  title = character(0),
  description = character(0),
  contact = character(0),
  modified = character(0),
  endpoints = list()
)

TemporalMainGroup(
  title = character(0),
  description = character(0),
  contact = character(0),
  modified = character(0),
  endpoints = list()
)

TemporalOpen(
  title = character(0),
  description = character(0),
  contact = character(0),
  modified = character(0),
  endpoints = list()
)

TemporalOpenGroup(
  title = character(0),
  description = character(0),
  contact = character(0),
  modified = character(0),
  endpoints = list()
)
}
\arguments{
\item{title}{\verb{<chr>} dataset title}

\item{description}{\verb{<chr>} dataset description}

\item{contact}{\verb{<chr>} dataset contact}

\item{modified}{\verb{<chr>} dataset modified date}

\item{endpoints}{\verb{<list>} dataset endpoints}
}
\description{
Temporal

TemporalMain

TemporalMainGroup

TemporalOpen

TemporalOpenGroup
}
\concept{base-classes}
\concept{main-classes}
\concept{open-classes}
