% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{Temporal}
\alias{Temporal}
\alias{TemporalMain}
\alias{TemporalOpen}
\title{Temporal}
\usage{
Temporal(
  title = character(0),
  description = character(0),
  contact = character(0),
  modified = character(0),
  endpoints = list()
)

TemporalMain(
  title = character(0),
  description = character(0),
  contact = character(0),
  modified = character(0),
  endpoints = list()
)

TemporalOpen(
  title = character(0),
  description = character(0),
  contact = character(0),
  modified = character(0),
  endpoints = list()
)
}
\arguments{
\item{title}{\verb{<chr>} dataset title}

\item{description}{\verb{<chr>} dataset description}

\item{contact}{\verb{<chr>} dataset contact}

\item{modified}{\verb{<chr>} dataset modified date}

\item{endpoints}{\verb{<list>} dataset endpoints}
}
\description{
Temporal

TemporalMain

TemporalOpen
}
\seealso{
Other base-classes: 
\code{\link{Current}()}

Other main-classes: 
\code{\link{Current}()}

Other open-classes: 
\code{\link{Current}()}
}
\concept{base-classes}
\concept{main-classes}
\concept{open-classes}
