% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_backend.R
\name{TemporalGroup}
\alias{TemporalGroup}
\title{Main API Endpoint Group (Temporal)}
\usage{
TemporalGroup(title = character(0), contact = character(0), endpoints = list())
}
\arguments{
\item{title}{\verb{<chr>} title}

\item{contact}{\verb{<chr>} contact name and email address}

\item{endpoints}{\verb{<data.frame>} endpoints}
}
\value{
An S7 \verb{<TemporalGroup>} object.
}
\description{
Main API Endpoint Group (Temporal)
}
\keyword{internal}
