% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7_backend.R
\name{TemporalOpen}
\alias{TemporalOpen}
\title{Temporal Open Payments API Endpoint}
\usage{
TemporalOpen(alias)
}
\arguments{
\item{alias}{\verb{<chr>} endpoint alias}
}
\value{
An S7 \verb{<TemporalOpen>} object.
}
\description{
Temporal Open Payments API Endpoint
}
\examples{
TemporalOpen("general")
TemporalOpen("ownership")
TemporalOpen("research")
TemporalOpen("recipient_nature")
TemporalOpen("recipient_entity")
TemporalOpen("entity_nature")
TemporalOpen("entity_recipient_nature")
TemporalOpen("state_nature")
}
